update layout mechanisms update documentation
