+++ release/entry.js
@@ -1,1 +1,1 @@
-!function(e){var t={};function n(i){if(t[i])return t[i].exports;var s=t[i]={i:i,l:!1,exports:{}};return e[i].call(s.exports,s,s.exports,n),s.l=!0,s.exports}n.m=e,n.c=t,n.d=function(e,t,i){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:i})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var i=Object.create(null);if(n.r(i),Object.defineProperty(i,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var s in e)n.d(i,s,function(t){return e[t]}.bind(null,s));return i},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=0)}([function(e,t,n){"use strict";n.r(t),n.d(t,"MDWDialog",(function(){return k})),n.d(t,"MDWSnackbar",(function(){return C})),n.d(t,"MDWTemplate",(function(){return S})),n.d(t,"MDWSurface",(function(){return H})),n.d(t,"MDWDrag",(function(){})),n.d(t,"MDWSwipe",(function(){}));const i=function(){if("undefined"!=typeof FORCE_MOBILE&&"true"===FORCE_MOBILE)return!0;var e=!1;return function(t){(/(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows ce|xda|xiino/i.test(t)||/1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i.test(t.substr(0,4)))&&(e=!0)}(navigator.userAgent||navigator.vendor||window.opera),e}(),o=function(){if("undefined"!=typeof FORCE_MOBILE&&"true"===FORCE_MOBILE)return!0;var e=!1;return function(t){(/(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows ce|xda|xiino|android|ipad|playbook|silk/i.test(t)||/1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i.test(t.substr(0,4)))&&(e=!0)}(navigator.userAgent||navigator.vendor||window.opera),e}();const r=new class{constructor(){this._uid=1,this._setupTransitionEvent(),this._setTransformPropertyName(),this.isPhone=i,this.isPhoneAndTablet=o,this.routeChangeCallbacks=new Set,this.onReady(()=>{this.isMobile?document.body.classList.add("mdw-is-mobile"):document.body.classList.remove("mdw-is-mobile"),window.addEventListener("hashchange",this.onRouteChange)})}onReady(e){document.body?e():setTimeout(()=>{this.onReady(e)},0)}uid(e="id"){return`${e}_${this._uid++}`}get isMobile(){return this.isPhoneAndTablet}lockPageScroll(){(document.body.classList.contains("prevent-over-scroll")?document.querySelector("mdw-page > mdw-content"):document.body).style.overflow="hidden"}unlockPageScroll(){(document.body.classList.contains("prevent-over-scroll")?document.querySelector("mdw-page > mdw-content"):document.body).style.overflow=""}disableUserSelect(){document.body.classList.add("mdw-no-select")}enableUserSelect(){document.body.classList.remove("mdw-no-select")}debounce(e,t){let n;return function(){const i=arguments,s=this;clearTimeout(n),n=setTimeout(()=>{n=void 0,e.apply(s,i)},t||10)}}throttle(e,t){let n;return function(){const i=arguments,s=this;n||(n=!0,e.apply(s,i),setTimeout(()=>{n=!1},t))}}rafThrottle(e){let t;return function(){const n=arguments,i=this;t||(t=!0,e.apply(i,n),requestAnimationFrame(()=>{t=!1}))}}querySlotted(e,t){if(!e)throw Error("requires either component");if(!t)throw Error("requires selector");return e.shadowRoot.querySelector("slot")?e.shadowRoot.querySelector("slot").assignedNodes().find(e=>!!e.matches&&e.matches(t)):null}querySlottedAll(e,t){if(!e)throw Error("requires either component");if(!t)throw Error("requires selector");return e.shadowRoot.querySelector("slot").assignedNodes({flatten:!0}).reduce((e,n)=>n.querySelectorAll?e.concat([...n.querySelectorAll(t)]):e,[])}slottedChildren(e){if(!e)throw Error("requires either component");return e.shadowRoot.querySelector("slot").assignedNodes()}get transitionEventName(){return this.transitionEventName_}get transformPropertyName(){return this.transformPropertyName_}addBackdrop(e,t,n={sheet:!1}){const i=this.uid();e.insertAdjacentHTML("afterend",`<div id="${i}" class="mdw-backdrop"></div>`);const s=document.querySelector(`#${i}`);return!0===n.sheet&&s.classList.add("mdw-sheet-backdrop"),t&&s.addEventListener("click",t),{remove(){t&&s.removeEventListener("click",t),s.remove()}}}get searchParamters(){return this._extractSearchParameters(this._clean(window.location.href)).split(",").filter(e=>!!e).reduce((e,t)=>{const n=t.split("=");return e[n[0]]=n[1],e},{})}setSearchParamter(e,t){const n=this.searchParamters;null==t?delete n[e]:n[e]=t;let i=window.location.href.split("?")[0];Object.keys(n).length>0&&(i+="?"+Object.keys(n).map(e=>`${e}=${n[e]}`).join(",")),window.history.pushState({path:i},"",i)}removeSearchParamter(e){this.setSearchParamter(e,void 0)}addOnRouteChange(e){return this.routeChangeCallbacks.add(e),()=>{this.routeChangeCallbacks.delete(e)}}removeOnRouteChange(e){this.routeChangeCallbacks.has(e)&&this.routeChangeCallbacks.delete(e)}_routeChange({oldUrl:e,newURL:t}={oldUrl:void 0,newURL:void 0}){this.routeChangeCallbacks.forEach(n=>n({oldUrl:e,newURL:t}))}_clean(e){return e instanceof RegExp?s:e.replace(/\/+$/,"").replace(/^\/+/,"/")}_extractSearchParameters(e){return e.split(/\?(.*)?$/).slice(1).join("")}_setupTransitionEvent(){const e=document.createElement("fakeelement"),t={transition:"transitionend",OTransition:"oTransitionEnd",MozTransition:"transitionend",WebkitTransition:"webkitTransitionEnd"};for(let n in t)void 0!==e.style[n]&&(this.transitionEventName_=t[n])}_setTransformPropertyName(e=!1){if(void 0===this.transformPropertyName_||e){const e=document.createElement("div");this.transformPropertyName_="transform"in e.style?"transform":"webkitTransform"}}};window.MDWUtils=r;var a=r;const d=new Map,l={start:"start",move:"move",end:"end"};function c(e,t){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");if("function"!=typeof t)throw Error("callback must be a function");const n=new m(e,t);n.addEvents(),d.get(e)||d.set(e,new Map),d.get(e).set(t,n)}function h(e,t){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");const n=d.get(e);if(n)if(t){const e=n.get(t);e&&e.removeEvents(),n.delete(t)}else n.forEach(e=>e.removeEvents()),d.delete(e)}class m{constructor(e,t){this.element=e,this.callback=t,this.hasPointerEvent=!!window.PointerEvent,this.bound_handleGestureStart=this.handleGestureStart.bind(this),this.bound_handleGestureMove=this.handleGestureMove.bind(this),this.bound_handleGestureEnd=this.handleGestureEnd.bind(this),this.callbackThrottle=a.rafThrottle(e=>{e.distance=this.getDistance(e),e.direction=this.getDirection(this.lastDistance,e.distance),this.lastDistance=e.distance,this.callback(e)})}addEvents(){this.disableTouchEvents(),a.isMobile?this.element.addEventListener("touchstart",this.bound_handleGestureStart,!1):this.element.addEventListener("mousedown",this.bound_handleGestureStart)}disable(){this.removeEvents()}enable(){this.addEvents()}enableTouchEvents(){this.element.style["touch-action"]=""}disableTouchEvents(){this.element.style["touch-action"]="none"}removeEvents(){this.enableTouchEvents(),a.isMobile?(this.element.removeEventListener("touchstart",this.bound_handleGestureStart),this.element.removeEventListener("touchmove",this.bound_handleGestureMove),this.element.removeEventListener("touchend",this.bound_handleGestureEnd),this.element.removeEventListener("touchcancel",this.bound_handleGestureEnd)):(this.element.removeEventListener("mousedown",this.bound_handleGestureStart),window.removeEventListener("mousemove",this.bound_handleGestureMove),window.removeEventListener("mouseup",this.bound_handleGestureEnd))}handleGestureStart(e){"mousedown"===e.type&&0!==e.button||(e.state="start",a.isMobile?(this.element.addEventListener("touchmove",this.bound_handleGestureMove,!1),this.element.addEventListener("touchend",this.bound_handleGestureEnd,!1),this.element.addEventListener("touchcancel",this.bound_handleGestureEnd,!1)):(window.addEventListener("mousemove",this.bound_handleGestureMove),window.addEventListener("mouseup",this.bound_handleGestureEnd)),this.startTime=Date.now(),this.initialTouchPos=this.getClientXY(e),this.lastDistance=this.getDistance(e),this.moved=!1)}handleGestureMove(e){e.state=this.moved?"move":"start",this.moved=!0,this.initialTouchPos&&this.callbackThrottle(e)}handleGestureEnd(e){if(e.state="end",a.isMobile?(this.element.removeEventListener("touchmove",this.bound_handleGestureMove),this.element.removeEventListener("touchend",this.bound_handleGestureEnd),this.element.removeEventListener("touchcancel",this.bound_handleGestureEnd)):(window.removeEventListener("mousemove",this.bound_handleGestureMove),window.removeEventListener("mouseup",this.bound_handleGestureEnd)),!1!==this.moved){if(this.endTime=Date.now(),e.runTime=this.endTime-this.startTime,e.distance=this.getDistance(e),e.direction=this.getDirection({x:0,y:0},e.distance),e.velocity=this.getVelocity(e.distance,e.runTime),void 0===e.clientX){const t=this.getClientXY(e);e.clientX=t.x,e.clientY=t.y}this.callback(e)}}getDistance(e){const t=this.getClientXY(e);return{x:t.x-this.initialTouchPos.x,y:t.y-this.initialTouchPos.y}}getDirection(e,t){const n=t.x>e.x?1:t.x===e.x?0:-1,i=t.y>e.y?1:t.y===e.y?0:-1;return{x:n,y:i,xDescription:0===n?"none":1===n?"right":"left",yDescription:0===i?"none":1===i?"down":"up"}}getClientXY(e){return{x:e.targetTouches&&e.targetTouches.length?e.targetTouches[0].clientX:e.changedTouches&&e.changedTouches.length?e.changedTouches[0].clientX:e.clientX,y:e.targetTouches&&e.targetTouches.length?e.targetTouches[0].clientY:e.changedTouches&&e.changedTouches.length?e.changedTouches[0].clientY:e.clientY}}getVelocity(e,t){return{x:e.x/t,y:e.y/t}}}window.MDWDrag={states:l,addDragListener:c,removeDragListener:h,enableDragListenerForElement:function(e){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");const t=d.get(e);t&&t.forEach(e=>e.enable())},disableDragListenerForElement:function(e){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");const t=d.get(e);t&&t.forEach(e=>e.disable())}};const u=new Map;class p{constructor(e,t){this.bound_dragEvent=this.dragEvent.bind(this),this.element=e,this.callback=t}addEvents(){c(this.element,this.bound_dragEvent)}removeEvents(){h(this.element,this.bound_dragEvent)}dragEvent(e){"end"===e.state&&(Math.abs(e.velocity.x)>1.4||Math.abs(e.velocity.y)>1.4)&&this.callback(e)}}function w(e,t){const n=new p(e,t);n.addEvents(),u.get(e)||u.set(e,new Map),u.get(e).set(t,n)}function b(e,t){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");const n=u.get(e);if(n)if(t){const e=n.get(t);e&&e.removeEvents(),n.delete(t)}else n.forEach(e=>e.removeEvents()),u.delete(e)}window.MDWSwipe={addSwipeListener:w,removeSwipeListener:b};new class{constructor(){this.routes={},this.classReference={},this.intercepter=void 0,this.pageClassnameRegex=/class\s(.*)\sextends/,this.bound_resolve=this._resolve.bind(this),this.PARAMETER_REGEXP=/([:*])(\w+)/g,this.WILDCARD_REGEXP=/\*/g,this.REPLACE_VARIABLE_REGEXP="([^/]+)",this.REPLACE_WILDCARD="(?:.*)",this.FOLLOWED_BY_SLASH_REGEXP="(?:/$|$)",this.MATCH_REGEXP_FLAGS="",this._transitionPages=!1,this.bound_onTransitionComplete=this._onTransitionComplete.bind(this),this.__mutationObserver=new MutationObserver(()=>{window.activePage.connectedCallback&&window.activePage.connectedCallback(),this._stopWatchingForConnect()})}init(){window.addEventListener("hashchange",this.bound_resolve),window.addEventListener("DOMContentLoaded",()=>{this._resolve(void 0,!0)})}interceptRouteChange(e){"function"==typeof e&&(this.intercepter=e)}get transitionPages(){return this._transitionPages}set transitionPages(e){this._transitionPages=!!e}get path(){let e=window.location.hash.replace(/.*#/,"");return e.indexOf("?")>-1&&(e=e.split("?")[0]),"/"!==e.charAt(0)&&(e="/"+e),e}get urlParameters(){const e=this._match(this.path);return e?e.params:{}}get searchParamters(){return this._extractSearchParameters(this._clean(window.location.href)).split(",").filter(e=>!!e).reduce((e,t)=>{const n=t.split("=");return e[n[0]]=n[1],e},{})}set hash(e){window.location=`#/${e}`}setSearchParamter(e,t){const n=this.searchParamters;null==t?delete n[e]:n[e]=t;let i=window.location.href.split("?")[0];Object.keys(n).length>0&&(i+="?"+Object.keys(n).map(e=>`${e}=${n[e]}`).join(",")),window.history.pushState({path:i},"",i)}removeSearchParamter(e){this.setSearchParamter(e,void 0)}addTransitionCSS(){this._addTransitionCSSAdded||(document.body.insertAdjacentHTML("beforebegin","<style>\n      page-container {\n        display: flex;\n      }\n      page-container.in-transition {\n        overflow-x: hidden;\n      }\n      page-render-block {\n        width: 100%;\n        flex-shrink: 0;\n        opacity: 1;\n      }\n      page-render-block.before-transition-page-out {\n        pointer-events: none;\n        user-select: none;\n      }\n      page-render-block.before-transition-page-in {\n        transform: scale(0.9) translateX(-100%);\n        opacity: 0;\n      }\n      page-render-block.transition-page-out {\n        transform: scale(0.9);\n        opacity: 0;\n        transition: opacity .16s linear,\n                    transform .26s cubic-bezier(0,0,.2,1);\n      }\n      page-render-block.transition-page-in {\n        transform: scale(1) translateX(-100%);\n        transform-origin: -50% 0;\n        opacity: 1;\n        transition: opacity .18s linear,\n                    transform .26s cubic-bezier(0,0,.2,1);\n      }\n    </style>"),this._addTransitionCSSAdded=!0)}addPageHideCSS(){this._addPageHideCSSAdded||(document.body.insertAdjacentHTML("beforebegin","<style>\n      .mdw-hide-non-page-container {\n        display: none;\n      }\n    </style>"),this._addPageHideCSSAdded=!0)}addPageClass(e,t){const n=this.getClassName(e,t);t&&this.addPageClassPath(e,t),(e.routes||[]).forEach(i=>{if(t!==i){if(this.routes[i])throw Error(`Path already exists: ${i}`);this.classReference[n]=e,this.routes[i]=n}})}addPageClassPath(e,t){const n=this.getClassName(e,t);if(this.routes[t])throw Error(`Path already exists: ${optionalPath}`);this.classReference[n]=e,this.routes[t]=n}getClassName(e,t){const n=this.pageClassnameRegex.exec(e);return n?n[1]:t.split("/").pop().replace(".js","")}setRoot(e){if(this.routes[e]){const t=this.classReference[this.routes[e]];this.addPageClassPath(t,"/")}else this.addPageClassPath(e,"/")}set404({Class:e}){e&&(this._notFoundRouteClass=e)}showPageOnly(){const e=document.querySelector("page-container"),t=document.documentElement;let n,i=e,s=e;for(;i.parentNode&&i.parentNode!==t;){for(i=i.parentNode,n=i.firstChild;n;)1===n.nodeType&&n!==s&&n.classList.add("mdw-hide-non-page-container"),n=n.nextSibling;s=i}this.addPageHideCSS(),this._isShowingPageOnly=!0}undoShowPageOnly(){!0===this._isShowingPageOnly&&([...document.querySelectorAll(".mdw-hide-non-page-container")||[]].forEach(e=>e.classList.remove("mdw-hide-non-page-container")),this._isShowingPageOnly=!1)}_resolve(e,t=!1){const{oldURL:n,newURL:i}=e||{};if(!1===t&&void 0!==n&&n===i)return;const s=this.intercepter?this.intercepter(i,n):void 0;if(s&&s.then&&"function"==typeof s.then&&console.error("you cannot return a Promise to the router.intercepter callback. Expecting either true or false"),!1===s)return void window.history.go();const o=this.path,r=this._match(o);if(!r)return this._notFoundRouteClass?this._changePage(this._notFoundRouteClass):console.warn("no page found and no default not found page setup");let a=o;if(t&&this._pageIsPreRendered())return;let d=this._extractSearchParameters(this._clean(window.location.href));if(d&&(a+=`?${d}`),void 0!==n){const e=n.length>i.length?n.replace(i,""):i.replace(n,"");if(""===e||"#/"===e)return}return this._changePage(r)}_watchForConnect(){const e=document.querySelector("page-render-block:not(.previous)");this.__mutationObserver.observe(e,{childList:!0})}_stopWatchingForConnect(){this.__mutationObserver.disconnect()}resetPageScroll(){const e=document.querySelector("mdw-page mdw-page-content");if(e&&e.scrollTop>0)return e.scrollTop=0;const t=document.querySelector("mdw-page mdw-content");if(t&&t.scrollTop>0)return t.scrollTop=0;const n=document.querySelector("mdw-page");if(n&&n.scrollTop>0)return n.scrollTop=0;const i=document.querySelector("body");if(i.scrollTop>0)return i.scrollTop=0;const s=document.documentElement;return s.scrollTop>0?s.scrollTop=0:void 0}_changePage({Class:e}){if(!e)throw Error("no class found");const t=document.querySelector("page-container");if(!t)throw Error("<page-container> required for router to work");this._stopWatchingForConnect(),this.undoShowPageOnly();const n=document.querySelector("page-render-block");if(!n)return t.appendChild(document.createElement("page-render-block")),window.activePage=new e,this._watchForConnect(),window.activePage.render(),void this.resetPageScroll();if(!this._transitionPages)return window.activePage.disconnectedCallback(),window.activePage=new e,this._watchForConnect(),window.activePage.render(),void this.resetPageScroll();n.classList.add("previous"),n.classList.add("before-transition-page-out"),window.activePage._disableRender=!0,window.activePage.disconnectedCallback();const i=document.createElement("page-render-block");i.classList.add("before-transition-page-in"),n.insertAdjacentElement("afterend",i);const s=new e;window.activePage=s,this._watchForConnect(),s.render();const o=document.querySelector("title");o&&(o.innerText=s.title),t.classList.add("in-transition"),n.classList.add("transition-page-out"),i.classList.add("transition-page-in"),n.addEventListener("transitionend",this.bound_onTransitionComplete),i.addEventListener("transitionend",this.bound_onTransitionComplete)}_onTransitionComplete({target:e}){e.removeEventListener("transitionend",this.bound_onTransitionComplete),e.classList.contains("transition-page-out")?e.remove():(e.classList.remove("before-transition-page-in"),e.classList.remove("transition-page-in")),document.querySelector("page-render-block.previous")||document.querySelector("page-render-block.next")||document.querySelector("page-container").classList.remove("in-transition")}_pageIsPreRendered(){const e=document.querySelector("page-render-block");return!!(e&&e.children.length>0)}_clean(e){return e instanceof RegExp?s:e.replace(/\/+$/,"").replace(/^\/+/,"/")}_extractSearchParameters(e){return e.split(/\?(.*)?$/).slice(1).join("")}_match(e){let t=this._findMatchedRoutes(e);return!!t.length&&(1===t.length?t[0]:t.sort((e,t)=>t.params?1:-1)[0])}_findMatchedRoutes(e){return Object.keys(this.routes).map(t=>{const n=this.routes[t],{regexp:i,paramNames:s}=this._replaceDynamicURLParts(this._clean(t)),o=e.replace(/^\/+/,"/").match(i),r=this._regExpResultToParams(o,s),a=this.classReference[n];return!!o&&{match:o,route:t,params:r,className:n,Class:a}}).filter(e=>e&&""!==e.match[0])}_replaceDynamicURLParts(e){let t=[],n="";return n=e instanceof RegExp?e:new RegExp(this._clean(e).replace(this.PARAMETER_REGEXP,(e,n,i)=>(t.push(i),this.REPLACE_VARIABLE_REGEXP)).replace(this.WILDCARD_REGEXP,this.REPLACE_WILDCARD)+this.FOLLOWED_BY_SLASH_REGEXP,this.MATCH_REGEXP_FLAGS),{regexp:n,paramNames:t}}_regExpResultToParams(e,t){return 0===t.length?null:e?e.slice(1,e.length).reduce((e,n,i)=>(null===e&&(e={}),e[t[i]]=decodeURIComponent(n),e),null):null}};class g extends HTMLElement{constructor(){super()}get _templateId(){return`${this.nodeName.toLowerCase()}--template`}cloneTemplate(e){let t=document.getElementById(this._templateId);t||(t=this._createTemplate());const n=t.content,i=this.shadowRoot?this.shadowRoot:this.attachShadow({mode:"open"}),s=n.cloneNode(!0);e&&(!this.__isBuildProcess&&this.beforeRender&&this.beforeRender(),s.querySelector("render-block").innerHTML=this.template()),i.appendChild(s),!this.__isBuildProcess&&this.afterRender&&this.afterRender()}_createTemplate(){const e=document.createElement("template");return e.setAttribute("id",this._templateId),e.innerHTML=`\n        <style>\n          ${this.styles()}\n        </style>\n        <render-block>\n          ${this.template()}\n        </render-block>\n    `,document.body.insertAdjacentElement("beforeend",e),e}connectedCallback(){!this.__isBuildProcess&&this.addEvents&&this.addEvents()}disconnectedCallback(){!this.__isBuildProcess&&this.removeEvents&&this.removeEvents()}render(){if(this.__isBuildProcess)return;const e=this.shadowRoot.querySelector("render-block");if(!e)throw Error("Could not find <render-block>");this.removeEvents&&this.removeEvents(),this.beforeRender&&this.beforeRender(),e.innerHTML=this.template(),this.afterRender&&this.afterRender(),this.addEvents&&this.addEvents()}beforeRender(){}afterRender(){}addEvents(){}removeEvents(){}styles(){}externalStyles(){}template(){}}const _=new class{constructor(){this.queue=[]}add(e,t){this.queue.push({el:e,resolver:t}),this.handleQueue()}remove(e){this.current&&this.current.el===e?(this.current.resolver(!1),e._dissmiss()):this.queue=this.queue.filter(t=>t.el!==e)}accept(e){this.current&&this.current.el===e?(this.current.resolver(!0),e._dissmiss()):this.queue=this.queue.filter(t=>t.el!==e)}handleQueue(){0!==this.queue.length&&(this.current||(this.current=this.queue.shift(),this.current.el._show(),this.current.el.addEventListener("close",()=>{this.current=void 0,setTimeout(()=>{this.handleQueue()},300)})))}create({message:e,dismissLabel:t="dismiss",acceptLabel:n=null,template:i,parent:s}){if(!e&&!i)throw Error("Either `message` or `template` is required");if(!i&&!t&&!n)throw Error("When not using a `template` you are required to provide either a `dismissLabel` or an `acceptLabel`");const o=a.uid();i||(i=this.template(e,t,n,o));let r=s||document.querySelector("mdw-page > mdw-top-app-bar");r||(r=document.querySelector("mdw-page")),r||(r=document.querySelector("body"));let d,l=void 0;"MDW-TOP-APP-BAR"===r.nodeName?(r.insertAdjacentHTML("afterend",i),l=document.querySelector(`mdw-banner#${o}`)):(r.insertAdjacentHTML("afterbegin",i),l=document.querySelector(`mdw-banner#${o}`));const c=new Promise(e=>{d=e});return this.add(l,d),c}template(e,t,n,i){return`\n      <mdw-banner id="${i}" class="mdw-elevation-1">\n        <div>${e}</div>\n        <div>\n          ${t?`<mdw-button onclick="${i}.dismiss()" class="mdw-secondary">${t}</mdw-button>`:""}\n          ${n?`<mdw-button onclick="${i}.accept()" class="mdw-secondary">${n}</mdw-button>`:""}\n        </div>\n      </mdw-banner>\n    `}};window.MDWBanner=_;var v=_;customElements.define("mdw-banner",class extends g{constructor(){super()}connectedCallback(){this.style.marginBottom=`-${this.clientHeight+1}px`}show(){v.add(this)}dismiss(){v.remove(this)}accept(){v.accept(this)}_show(){this.classList.add("mdw-show")}_dissmiss(){const e=this;e.addEventListener(a.transitionEventName,(function t(){e.removeEventListener(a.transitionEventName,t),e.remove()})),this.classList.add("mdw-dismiss"),this.dispatchClose()}dispatchClose(){this.dispatchEvent(new CustomEvent("close"))}}),customElements.define("mdw-bottom-navigation",class extends g{constructor(){super(),!o&&this.hasAttribute("mdw-mobile-only")?(this.style.display="none",this.style.pointerEvents="none"):(this.bound_routeChange=this.routeChange.bind(this),document.body.classList.add("mdw-has-bottom-navigation"))}connectedCallback(){window.addEventListener("hashchange",this.bound_routeChange),window.addEventListener("DOMContentLoaded",this.bound_routeChange)}disconnectedCallback(){window.removeEventListener("hashchange",this.bound_routeChange),window.removeEventListener("DOMContentLoaded",this.bound_routeChange)}get path(){let e=window.location.hash.replace(/.*#/,"");return e.indexOf("?")>-1&&(e=e.split("?")[0]),"/"!==e.charAt(0)&&(e="/"+e),e}routeChange(){this.querySelectorAll(".mdw-current-link").forEach(e=>e.classList.remove("mdw-current-link"));let e=this.querySelectorAll(`[href="#${this.path}"]`);e&&0!==e.length||(e=this.querySelectorAll(`[alt-href="#${this.path}"]`)),e.forEach(e=>e.classList.add("mdw-current-link"))}});class f{constructor(e={}){if(this.RIPPLE_FADE_IN_DURATION=200,this.RIPPLE_FADE_OUT_DURATION=150,this.RIPPLE_STATE={FADING_IN:"FADING_IN",VISIBLE:"VISIBLE",FADING_OUT:"FADING_OUT",HIDDEN:"HIDDEN"},!e.element)throw Error("requires config.element");if(!e.triggerElement)throw Error("requires config.triggerElement");this.element=e.element,this.triggerElement=[].concat(e.triggerElement).filter(e=>!!e),this.centered=!!e.centered,this.speedFactor=e.speedFactor||1,this.radius=e.radius,this.color=e.color||null,this.persistent=!!e.persistent,this.activeRipples=new Set,this.isMousedown=!1,this.bound_mousesdown_=this.mousesdown_.bind(this),this.bound_mouseup_=this.mouseup_.bind(this),this.bound_mouseleave_=this.mouseleave_.bind(this),this.triggerElement.forEach(e=>{e.addEventListener("mousedown",this.bound_mousesdown_)})}destroy(){this.triggerElement.forEach(e=>{e.removeEventListener("mousedown",this.bound_mousesdown_)})}mousesdown_(e){this.isMousedown=!0,this.triggerElement.forEach(e=>{e.addEventListener("mouseup",this.bound_mouseup_),e.addEventListener("mouseleave",this.bound_mouseleave_)}),this.fadeInRipple(e.pageX,e.pageY)}mouseup_(e){this.isMousedown=!1,this.activeRipples.forEach(e=>{e.config.persistent||e.state!==this.RIPPLE_STATE.VISIBLE||e.fadeOut()}),this.triggerElement.forEach(e=>{e.removeEventListener("mouseup",this.bound_mouseup_),e.removeEventListener("mouseleave",this.bound_mouseleave_)})}mouseleave_(){this.isMousedown&&this.mouseup_()}fadeInRipple(e,t){const n=this.element.getBoundingClientRect();if(this.centered)e=n.left+n.width/2,t=n.top+n.height/2;else{const n=this.getViewportScrollPosition();e-=n.left,t-=n.top}const i=this.radius||this.distanceToFurthestCorner(e,t,n),s=this.RIPPLE_FADE_IN_DURATION*(1/this.speedFactor),o=e-n.left,r=t-n.top,a=document.createElement("div");a.classList.add("mdw-ripple-element"),a.style.left=`${o-i}px`,a.style.top=`${r-i}px`,a.style.height=`${2*i}px`,a.style.width=`${2*i}px`,a.style.backgroundColor=this.color,a.style.transitionDuration=`${s}ms`,this.element.appendChild(a),this.enforceStyleRecalculation(a),a.style.transform="scale(1)";let d={config:{centered:this.centered,tiggerElement:this.triggerElement,speedFactor:this.speedFactor,radius:i,color:this.color,persistent:this.persistent,duration:s},element:a,fadeOut:()=>l(),state:this.RIPPLE_STATE.FADING_IN};const l=()=>{this.fadeOutRipple(d)};this.activeRipples.add(d),setTimeout(()=>{d.state=this.RIPPLE_STATE.VISIBLE,this.persistent||this.isMousedown||d.fadeOut()},s)}fadeOutRipple(e){if(!this.activeRipples.delete(e))return;const t=e.element;t.style.transitionDuration=`${this.RIPPLE_FADE_OUT_DURATION}ms`,t.style.opacity="0",e.state=this.RIPPLE_STATE.FADING_OUT,setTimeout(()=>{e.state=this.RIPPLE_STATE.HIDDEN,t.parentNode.removeChild(t)},this.RIPPLE_FADE_OUT_DURATION)}distanceToFurthestCorner(e,t,n){const i=Math.max(Math.abs(e-n.left),Math.abs(e-n.right)),s=Math.max(Math.abs(t-n.top),Math.abs(t-n.bottom));return Math.sqrt(i*i+s*s)}getViewportScrollPosition(){const e=document.documentElement.getBoundingClientRect();return{top:-e.top||document.body.scrollTop||window.scrollY||document.documentElement.scrollTop||0,left:-e.left||document.body.scrollLeft||window.scrollX||document.documentElement.scrollLeft||0}}enforceStyleRecalculation(e){window.getComputedStyle(e).getPropertyValue("opacity")}}customElements.define("mdw-button",class extends g{constructor(){super(),this.bound_asyncClick=this.asyncClick.bind(this),this.bound_hrefClick=this.hrefClick.bind(this),this.bound_checkHREFActive=this.checkHREFActive.bind(this),this.cloneTemplate(),this.setupAsync(),this.connectHREF()}connectedCallback(){this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:this})}disconnectedCallback(){this.ripple.destroy(),this.removeEventListener("click",this.bound_asyncClick),this.removeEventListener("click",this.bound_hrefClick),window.removeEventListener("hashchange",this.bound_checkHREFActive)}get spinnerContainer(){return this._spinnerContainer||(this._spinnerContainer=this.shadowRoot.querySelector(".mdw-spinner-container")),this._spinnerContainer}get pending(){return this._pending}setupAsync(){this.hasAttribute("mdw-async")&&this.addEventListener("click",this.bound_asyncClick)}resolve(){!1!==this._pending&&(this._pending=!1,this.hideSpinner())}asyncClick(e){!0!==this._pending&&(this._pending=!0,this.showSpinner())}showSpinner(){this._showSpinner=!0,this.classList.add("mdw-show-spinner");const e=this.classList.contains("mdw-primary")||this.classList.contains("mdw-secondary")||this.classList.contains("mdw-error");this.spinnerContainer.innerHTML=`<mdw-circular-progress mdw-mode="indeterminate" mdw-diameter="24" class="${e?"mdw-white":"mdw-grey"}" style="position: absolute; left: calc(50% - 12px); top: 6px;"></mdw-circular-progress>`}hideSpinner(){this._showSpinner=!1,this.classList.remove("mdw-show-spinner"),this.spinnerContainer.innerHTML=""}connectHREF(){this.hasAttribute("href")&&(this.checkHREFActive(),window.addEventListener("hashchange",this.bound_checkHREFActive),this.addEventListener("click",this.bound_hrefClick))}checkHREFActive(){if(!this.hasAttribute("href"))return;const e=document.location.href,t=document.location.hash;e===this.getAttribute("href")||e===this.getAttribute("href-alt")||t===this.getAttribute("href")||t===this.getAttribute("href-alt")?this.setAttribute("active","active"):this.removeAttribute("active")}hrefClick(){"_blank"!==this.getAttribute("target")?document.location.href=this.getAttribute("href"):window.open(this.getAttribute("href"),"_blank")}template(){return'\n      <span class="text"><slot></slot></span>\n      <span class="mdw-spinner-container"></span>\n      <div class="mdw-ripple mdw-button-ripple"></div>\n    '}styles(){return'\n      :host {\n        user-select: none;\n        -webkit-user-select: none;\n        align-items: center;\n        border: none;\n        box-sizing: border-box;\n        display: inline-flex;\n        font-size: 0.875rem;\n        font-weight: 500;\n        justify-content: center;\n        letter-spacing: 0.08929em;\n        outline: none;\n        overflow: hidden;\n        position: relative;\n        text-decoration: none;\n        text-transform: uppercase;\n        vertical-align: middle;\n        will-change: transform, opacity;\n        margin: 0;\n        background-color: transparent;\n        white-space: nowrap;\n\n        border-radius: 4px;\n        line-height: 2.25rem;\n        padding: 0 8px 0 8px;\n        height: 36px;\n        min-width: 64px;\n\n        color: var(--mdw-theme-on-primary);\n      }\n\n      :host-context(.mdw-density-comfortable),\n      :host(.mdw-density-comfortable) {\n        height: 28px;\n        margin-top: 0;\n        margin-bottom: 0;\n        padding: 0 8px 0 8px;\n      }\n\n      :host-context(.mdw-density-compact),\n      :host(.mdw-density-compact) {\n        height: 24px;\n        margin-top: 0;\n        margin-bottom: 0;\n        padding: 0 4px 0 4px;\n      }\n\n      :host(.mdw-icon):host-context(.mdw-density-comfortable),\n      :host(.mdw-density-comfortable.mdw-icon) {\n        height: 28px;\n        width: 28px;\n        margin-top: 0;\n        margin-bottom: 0;\n      }\n\n      :host(.mdw-icon):host-context(.mdw-density-compact),\n      :host(.mdw-density-compact.mdw-icon) {\n        height: 24px;\n        width: 24px;\n        margin-top: 0;\n        margin-bottom: 0;\n      }\n\n      :host-context(.mdw-density-comfortable) .mdw-spinner-container mdw-circular-progress,\n      :host(.mdw-density-comfortable) .mdw-spinner-container mdw-circular-progress {\n        top: 2px !important;\n      }\n\n      :host-context(.mdw-density-compact) .mdw-spinner-container mdw-circular-progress,\n      :host(.mdw-density-compact) .mdw-spinner-container mdw-circular-progress {\n        top: 0 !important;\n      }\n\n\n      :host::before,\n      :host::after {\n        position: absolute;\n        border-radius: 50%;\n        opacity: 0;\n        pointer-events: none;\n        content: "";\n        top: calc(50% - 100%);\n        left: calc(50% - 100%);\n        width: 200%;\n        height: 200%;\n        background-color: var(--mdw-theme-foreground);\n      }\n\n      :host::before {\n        z-index: 1;\n        transition: opacity 15ms linear,\n                    background-color 15ms linear;\n      }\n\n      :host(:hover) {\n        cursor: pointer;\n      }\n\n      :host([disabled]) {\n        background-color: transparent !important;\n        color: var(--mdw-theme-text-disabled-on-background);\n        cursor: default;\n        pointer-events: none;\n      }\n      :host::-moz-focus-inner {\n        padding: 0;\n        border: 0;\n      }\n\n      :host(:active) {\n        outline: none;\n      }\n\n      :host(.mdw-contained),\n      :host(.mdw-raised),\n      :host(.mdw-unelevated) {\n        background-color: var(--mdw-theme-background);\n        padding: 0 16px 0 16px;\n      }\n\n      :host(.mdw-contained)::before,\n      :host(.mdw-raised)::before,\n      :host(.mdw-unelevated)::before {\n        opacity: 0.08;\n      }\n\n      :host(.mdw-contained),\n      :host(.mdw-raised) {\n        box-shadow: 0px 3px 1px -2px rgba(0, 0, 0, 0.2), 0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 1px 5px 0px rgba(0, 0, 0, 0.12);\n        transition: box-shadow 280ms cubic-bezier(0.4, 0, 0.2, 1);\n      }\n\n      :host(.mdw-contained:hover),\n      :host(.mdw-contained:focus),\n      :host(.mdw-raised:hover),\n      :host(.mdw-raised:focus) {\n        box-shadow: 0px 2px 4px -1px rgba(0, 0, 0, 0.2), 0px 4px 5px 0px rgba(0, 0, 0, 0.14), 0px 1px 10px 0px rgba(0, 0, 0, 0.12);\n      }\n\n      :host(.mdw-contained:active),\n      :host(.mdw-raised:active) {\n        box-shadow: 0px 5px 5px -3px rgba(0, 0, 0, 0.2), 0px 8px 10px 1px rgba(0, 0, 0, 0.14), 0px 3px 14px 2px rgba(0, 0, 0, 0.12);\n      }\n\n      :host(.mdw-contained[disabled]),\n      :host(.mdw-raised[disabled]) {\n        box-shadow: 0px 0px 0px 0px rgba(0, 0, 0, 0.2), 0px 0px 0px 0px rgba(0, 0, 0, 0.14), 0px 0px 0px 0px rgba(0, 0, 0, 0.12);\n      }\n\n      :host(.mdw-outlined) {\n        border-color: rgba(0, 0, 0, 0.37);\n        line-height: inherit;\n        border-style: solid;\n        padding: 0 14px 0 14px;\n        border-width: 2px;\n      }\n\n      :host(.mdw-shaped) {\n        border-radius: var(--mdw-theme--button-shape-radius, 18px);\n      }\n\n      :host(.mdw-icon) {\n        border-radius: 50%;\n        min-width: 0;\n        width: 48px;\n        height: 48px;\n        padding: 12px;\n        line-height: 19px;\n      }\n      \n      :host(.mdw-icon) span.text {\n        display: flex;\n        flex-direction: column;\n        align-items: center;\n        font-size: 0.75rem;\n        text-transform: none;\n      }\n\n      :host(.mdw-bottom-navigation) {\n        border-radius: 50%;\n        min-width: 0;\n        max-width: 100px;\n        width: 56px;\n        height: 56px;\n        padding: 28px;\n      }\n\n      :host(:not(.mdw-icon)) .text ::slotted(mdw-icon) {\n        display: inline-block;\n        height: 28px;\n        vertical-align: middle;\n      }\n\n      /*\n      :host(.mdw-icon) ::slotted(mdw-icon) {\n        line-height: 19px;\n        font-weight: normal;\n        font-style: normal;\n        font-size: 24px;\n        letter-spacing: normal;\n        text-transform: none;\n        display: inline-block;\n        white-space: nowrap;\n        word-wrap: normal;\n      }\n\n      ::slotted(mdw-icon) {\n        width: 18px;\n        height: 18px;\n        font-size: 18px;\n        margin-left: -4px;\n        margin-right: 2px;\n        vertical-align: top;\n        line-height: 36px;\n      }\n      */\n\n      :host ::slotted(svg.mdw-icon) {\n        fill: currentColor;\n      }\n\n\n      /* primary */\n      :host(.mdw-primary) {\n        color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-primary.mdw-contained),\n      :host(.mdw-primary.mdw-raised),\n      :host(.mdw-primary.mdw-unelevated)  {\n        background-color: var(--mdw-theme-primary);\n        color: var(--mdw-theme-text-primary-on-background);\n      }\n\n      :host(.mdw-primary.mdw-outlined) {\n        border-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-primary)::before,\n      :host(.mdw-primary)::after {\n        background-color: var(--mdw-theme-primary);\n      }\n\n\n      /* secondary */\n\n      :host(.mdw-secondary) {\n        color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-secondary.mdw-contained),\n      :host(.mdw-secondary.mdw-raised),\n      :host(.mdw-secondary.mdw-unelevated) {\n        background-color: var(--mdw-theme-secondary);\n        color: var(--mdw-theme-text-primary-on-background);\n      }\n\n      :host(.mdw-secondary.mdw-outlined) {\n        border-color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-secondary)::before,\n      :host(.mdw-secondary)::after {\n        background-color: var(--mdw-theme-secondary);\n      }\n\n      /* error */\n\n      :host(.mdw-error) {\n        color: var(--mdw-theme-error);\n      }\n\n      :host(.mdw-error.mdw-contained),\n      :host(.mdw-error.mdw-raised),\n      :host(.mdw-error.mdw-unelevated) {\n        background-color: var(--mdw-theme-error);\n        color: var(--mdw-theme-text-primary-on-background);\n      }\n\n      :host(.mdw-error.mdw-outlined) {\n        border-color: var(--mdw-theme-error);\n      }\n\n      :host(.mdw-error)::before,\n      :host(.mdw-error)::after {\n        background-color: var(--mdw-theme-error);\n      }\n\n      :host(:not(.mdw-bottom-navigation):hover)::before {\n        opacity: 0.04;\n      }\n\n      :host(.mdw-show-spinner) span.text {\n        opacity: 0;\n      }\n\n      /* --- Ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-on-primary--rgb), 0.16);\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n      }\n\n      .mdw-button-ripple,\n      .mdw-button-focus-overlay {\n        border-radius: inherit;\n        position: absolute;\n        top: 0;\n        left: 0;\n        bottom: 0;\n        right: 0;\n        pointer-events: none;\n      }\n\n\n      :host(.mdw-primary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n\n      :host(.mdw-primary.mdw-contained) .mdw-ripple-element,\n      :host(.mdw-primary.mdw-raised) .mdw-ripple-element,\n      :host(.mdw-primary.mdw-unelevated) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n      :host(.mdw-secondary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.16);\n      }\n\n      :host(.mdw-secondary.mdw-contained) .mdw-ripple-element,\n      :host(.mdw-secondary.mdw-raised) .mdw-ripple-element,\n      :host(.mdw-secondary.mdw-unelevated) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n      :host(.error) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.16);\n      }\n\n      :host(.error.mdw-contained) .mdw-ripple-element,\n      :host(.error.mdw-raised) .mdw-ripple-element,\n      :host(.error.mdw-unelevated) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n\n      /* bottom navigation */\n      :host(.mdw-bottom-navigation) span.text {\n        display: flex;\n        flex-direction: column;\n        align-items: center;\n        font-size: 12px;\n        text-transform: none;\n        line-height: 12px;\n      }\n    '}}),customElements.define("mdw-card",class extends g{constructor(){super(),this.classList.add("mdw-elevation-1")}}),customElements.define("mdw-checkbox",class extends g{constructor(){super(),this.bound_handleClick=this.handleClick.bind(this),this.cloneTemplate(),this._defaultIcons(),this.state="unchecked"}connectedCallback(){this.hasAttribute("indeterminate")&&(this.indeterminate=!0),this.hasAttribute("checked")&&(this.checked=!0),this.hasAttribute("mdw-no-ripple")||(this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:[this],centered:!0})),this.addEventListener("click",this.bound_handleClick)}disconnectedCallback(){this.removeEventListener("click",this.bound_handleClick),this.ripple.destroy()}static get observedAttributes(){return["checked","indeterminate"]}attributeChangedCallback(e,t,n){this[e]=n}get state(){return this._state}set state(e){["checked","unchecked","indeterminate"].includes(e)||console.error(`ivalid state fro checkbox "${e}". Only excepts "checked", "unchecked", "indeterminate"`),this.setAttribute("mdw-state",e);const t=this._state!==e;this._state=e,!0===t&&this.dispatchEvent(new Event("change"))}get checked(){return"checked"===this.state}set checked(e){""===e&&(e=!0),this.state=!0===e?"checked":"unchecked"}get indeterminate(){return this.input.indeterminate}set indeterminate(e){""===e&&(e=!0),!0===e&&(this.state="indeterminate")}toggle(){this.checked=!this.checked}handleClick(){let e=!1;switch(this.state){case"indeterminate":this.state="checked",e=!0;break;case"checked":this.state="unchecked",e=!0;break;case"unchecked":this.state="checked",e=!0}}_defaultIcons(){0===this.children.length&&this.insertAdjacentHTML("afterbegin","\n        <mdw-icon mdw-checked>check_box</mdw-icon>\n        <mdw-icon mdw-unchecked>check_box_outline_blank</mdw-icon>\n        <mdw-icon mdw-indeterminate>indeterminate_check_box</mdw-icon>\n      ")}template(){return'\n      <slot></slot>\n      <div class="mdw-ripple mdw-checkbox-ripple"></div>\n    '}styles(){return'\n      :host {\n        --mdw-checkbox-size: 24px;\n        -webkit-tap-highlight-color: rgba(0, 0, 0, 0);\n        -webkit-box-sizing: content-box;\n        box-sizing: content-box;\n        cursor: pointer;\n        display: inline-block;\n        position: relative;\n        vertical-align: bottom;\n        white-space: nowrap;\n        cursor: pointer;\n\n        width: var(--mdw-checkbox-size);\n        height: var(--mdw-checkbox-size);\n        line-height: 0;\n        padding: 8px 0;\n      }\n\n      :host(.mdw-no-padding) {\n        padding: 0;\n      }\n\n      :host([disabled]) {\n        pointer-events: none;\n        cursor: not-allowed;\n        opacity: 0.5;\n      }\n\n      :host-context(.mdw-density-comfortable),\n      :host(.mdw-density-comfortable) {\n        padding: 9px;\n        margin: 0;\n      }\n\n      :host-context(.mdw-density-compact),\n      :host(.mdw-density-compact) {\n        padding: 5px;\n        margin: 0;\n      }\n\n      ::slotted(mdw-icon) {\n        position: absolute;\n        user-select: none;\n        outline: none;\n        font-size: var(--mdw-checkbox-size);\n      }\n\n      :host(.mdw-large) {\n        --mdw-checkbox-size: 48px;\n      }\n\n      :host(.mdw-primary) ::slotted(mdw-icon) {\n        color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-secondary) ::slotted(mdw-icon) {\n        color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-error) ::slotted(mdw-icon) {\n        color: var(--mdw-theme-error);\n      }\n      \n\n      /* --- state: checked --- */\n      :host([mdw-state="checked"]) ::slotted(mdw-icon[mdw-unchecked]),\n      :host([mdw-state="checked"]) ::slotted(mdw-icon[mdw-indeterminate]) {\n        display: none;\n      }\n\n      /* --- state: unchecked --- */\n      :host([mdw-state="unchecked"]) ::slotted(mdw-icon[mdw-checked]),\n      :host([mdw-state="unchecked"]) ::slotted(mdw-icon[mdw-indeterminate]) {\n        display: none;\n      }\n\n      /* --- state: indeterminate --- */\n      :host([mdw-state="indeterminate"]) ::slotted(mdw-icon[mdw-checked]),\n      :host([mdw-state="indeterminate"]) ::slotted(mdw-icon[mdw-unchecked]) {\n        display: none;\n      }\n\n\n      /* --- Ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n        background-color: rgba(var(--mdw-theme-on-background--rgb), 0.16);\n      }\n\n      .mdw-checkbox-ripple {\n        position: absolute;\n        left: 0;\n        top: 0;\n        right: 0;\n        bottom: 0;\n        border-radius: 50%;\n        z-index: 1;\n        pointer-events: none;\n        overflow: visible;\n      }\n\n      :host(.mdw-primary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n\n      :host(.mdw-secondary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.16);\n      }\n\n      :host(.mdw-error) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.16);\n      }\n    '}}),customElements.define("mdw-circular-progress",class extends g{constructor(){super(),this.insertedDiameters=[],this.cloneTemplate()}connectedCallback(){this.diameter=this.getAttribute("mdw-diameter")||100,this.render(),this.style.width=this.style.height=this.diameter+"px",this.value&&(this.value=this.value)}static get observedAttributes(){return["value"]}attributeChangedCallback(e,t,n){this[e]=n}get diameter(){return this._diameter}set diameter(e){this._diameter=parseInt((""+e).replace("px","")),this.insertedDiameters[this._diameter]||(this.insertedDiameters.push(this._diameter),this.shadowRoot.querySelector("style").sheet.insertRule(this._getAnimationText(),0))}get svg(){return this._svg||(this._svg=this.shadowRoot.querySelector("svg")),this._svg}get strokeWidth(){return this._strokeWidth||this.diameter/10}set strikeWidth(e){this._strokeWidth=parseInt((""+e).replace("px",""))}get value(){return this.getAttribute("value")}set value(e){this._value=Math.max(0,Math.min(100,parseInt((""+e).replace("px","")))),void 0!==this.diameter&&(this.circle.style.strokeDashoffset=this._strokeCircumference*(100-this._value)/100+"px")}get mode(){return"determinate"===this.getAttribute("mdw-mode")?"determinate":"indeterminate"}get circle(){return this._circle||(this._circle=this.shadowRoot.querySelector("circle")),this._circle}get _circleRadius(){return(this.diameter-10)/2}get _circleStrokeWidth(){return this.strokeWidth/this.diameter*100}get _strokeCircumference(){return 2*Math.PI*this._circleRadius}get INDETERMINATE_ANIMATION_TEMPLATE(){return"\n     @keyframes mat-progress-spinner-stroke-rotate-DIAMETER {\n        0%      { stroke-dashoffset: START_VALUE;  transform: rotate(0); }\n        12.5%   { stroke-dashoffset: END_VALUE;    transform: rotate(0); }\n        12.5001%  { stroke-dashoffset: END_VALUE;    transform: rotateX(180deg) rotate(72.5deg); }\n        25%     { stroke-dashoffset: START_VALUE;  transform: rotateX(180deg) rotate(72.5deg); }\n        25.0001%   { stroke-dashoffset: START_VALUE;  transform: rotate(270deg); }\n        37.5%   { stroke-dashoffset: END_VALUE;    transform: rotate(270deg); }\n        37.5001%  { stroke-dashoffset: END_VALUE;    transform: rotateX(180deg) rotate(161.5deg); }\n        50%     { stroke-dashoffset: START_VALUE;  transform: rotateX(180deg) rotate(161.5deg); }\n        50.0001%  { stroke-dashoffset: START_VALUE;  transform: rotate(180deg); }\n        62.5%   { stroke-dashoffset: END_VALUE;    transform: rotate(180deg); }\n        62.5001%  { stroke-dashoffset: END_VALUE;    transform: rotateX(180deg) rotate(251.5deg); }\n        75%     { stroke-dashoffset: START_VALUE;  transform: rotateX(180deg) rotate(251.5deg); }\n        75.0001%  { stroke-dashoffset: START_VALUE;  transform: rotate(90deg); }\n        87.5%   { stroke-dashoffset: END_VALUE;    transform: rotate(90deg); }\n        87.5001%  { stroke-dashoffset: END_VALUE;    transform: rotateX(180deg) rotate(341.5deg); }\n        100%    { stroke-dashoffset: START_VALUE;  transform: rotateX(180deg) rotate(341.5deg); }\n      }\n    "}_getAnimationText(){return this.INDETERMINATE_ANIMATION_TEMPLATE.replace(/START_VALUE/g,`${.95*this._strokeCircumference}`).replace(/END_VALUE/g,`${.2*this._strokeCircumference}`).replace(/DIAMETER/g,`${this.diameter}`)}template(){return`\n      <svg style="width: ${this.diameter}px; height: ${this.diameter}px;">\n        <circle\n          cx="50%"\n          cy="50%"\n          r="${this._circleRadius}"\n          style="\n            animation-name: mat-progress-spinner-stroke-rotate-${this.diameter};\n            stroke-dasharray: ${this._strokeCircumference}px;\n            stroke-width: ${this._circleStrokeWidth}%;\n          "\n          ></circle>\n      </svg>\n    `}styles(){return"\n      :host {\n        display: block;\n        position: relative;\n      }\n\n      svg {\n        position: absolute;\n        transform: rotate(-90deg);\n        top: 0;\n        left: 0;\n        transform-origin: center;\n        overflow: visible;\n      }\n\n      circle {\n        fill: transparent;\n        transform-origin: center;\n        transition: stroke-dashoffset 225ms linear;\n        stroke: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-white) circle {\n        stroke: white;\n      }\n\n      :host(.mdw-grey) circle {\n        stroke: grey;\n      }\n\n      :host(.mdw-secondary) circle {\n        stroke: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-error) circle {\n        stroke: var(--mdw-theme-error);\n      }\n\n      :host([mdw-mode='indeterminate']) {\n        animation: mat-progress-spinner-linear-rotate 2000ms linear infinite;\n      }\n\n      :host([mdw-mode='indeterminate']) circle {\n        transition-property: stroke;\n        animation-duration: 4000ms;\n        animation-timing-function: cubic-bezier(0.35, 0, 0.25, 1);\n        animation-iteration-count: infinite;\n      }\n\n      @keyframes mat-progress-spinner-linear-rotate {\n        0%       { transform: rotate(0deg); }\n        100%     { transform: rotate(360deg); }\n      }\n\n      @keyframes mat-progress-spinner-stroke-rotate-100 {\n        0% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotate(0);\n        }\n        12.5% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotate(0);\n        }\n        12.5001% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotateX(180deg) rotate(72.5deg);\n        }\n        25% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotateX(180deg) rotate(72.5deg);\n        }\n        25.0001% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotate(270deg);\n        }\n        37.5% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotate(270deg);\n        }\n        37.5001% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotateX(180deg) rotate(161.5deg);\n        }\n        50% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotateX(180deg) rotate(161.5deg);\n        }\n        50.0001% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotate(180deg);\n        }\n        62.5% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotate(180deg);\n        }\n        62.5001% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotateX(180deg) rotate(251.5deg);\n        }\n        75% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotateX(180deg) rotate(251.5deg);\n        }\n        75.0001% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotate(90deg);\n        }\n        87.5% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotate(90deg);\n        }\n        87.5001% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotateX(180deg) rotate(341.5deg);\n        }\n        100% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotateX(180deg) rotate(341.5deg);\n        }\n      }\n    "}});const y=new class{constructor(){this.currentDialog=null}open({title:e,message:t,okLabel:n,cancelLabel:i,position:s="center center",clickOutsideClose:o=!1}){return new Promise(r=>{const d=a.uid("dialog"),l=this.template({id:d,title:e,message:t,okLabel:n,cancelLabel:i,position:s});document.body.insertAdjacentHTML("beforeend",l);const c=document.querySelector(`#${d}`),h=e=>{r(e.detail.ok),c.removeEventListener("close",h),c.remove(),this.currentDialog=null};c.addEventListener("close",h),c.clickOutsideClose=o,this.currentDialog=c,requestAnimationFrame(()=>{c.open()})})}removeCurrent(){this.currentDialog.close()}template({id:e,title:t,message:n,okLabel:i,cancelLabel:s,position:o}){return`\n      <mdw-dialog id="${e}">\n        <mdw-panel mdw-position="${o}">\n          <mdw-dialog-container>\n            ${t?`<mdw-dialog-title>${t}</mdw-dialog-title>`:""}\n            <mdw-dialog-content>${n}</mdw-dialog-content>\n            <mdw-dialog-actions>\n              ${s?`<mdw-button class="mdw-error" onclick="${e}.close(false)">${s}</mdw-button>`:""}\n              ${i?`<mdw-button onclick="${e}.close(true)">${i}</mdw-button>`:""}\n            </mdw-dialog-actions>\n          </mdw-dialog-container>\n        </mdw-panel>\n      </mdw-dialog>\n    `}};window.MDWDialog=y;var k=y;customElements.define("mdw-dialog",class extends g{constructor(){super(),this.bound_onPanelClose=this.onPanelClose.bind(this),this.clickOutsideClose_=!1}disconnectedCallback(){this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose),this.panel.remove(),this.backdrop&&(this.backdrop.remove(),this.backdrop=void 0)}get panel(){return this._panel||(this._panel=this.querySelector("mdw-panel")),this._panel}get position(){return this._position||"center center"}set position(e){this._position=e}get clickOutsideClose(){return this.clickOutsideClose_}set clickOutsideClose(e){this.clickOutsideClose_=e}open(e=!1){this._fromService=e,this.panel.hoistToBody(),this.panel.setPosition(this.position),this.panel.addEventListener("MDWPanel:closed",this.bound_onPanelClose),this.backdrop=a.addBackdrop(this.panel,()=>{!0===this.clickOutsideClose&&this.close()}),requestAnimationFrame(()=>{this.panel.open()})}close(e){this.panel.close(),this.dispatchClose(e)}onPanelClose(){if(!this._fromService)return this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose),void(this.backdrop&&(this.backdrop.remove(),this.backdrop=void 0));this.remove()}dispatchClose(e=!1){this.dispatchEvent(new CustomEvent("close",{detail:{ok:e}}))}}),customElements.define("mdw-fab",class extends g{constructor(){super(),this.bound_asyncClick=this.asyncClick.bind(this),this.cloneTemplate(),this.setupAsync()}connectedCallback(){this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:this})}disconnectedCallback(){this.ripple.destroy(),this.removeEventListener("click",this.bound_asyncClick)}template(){return'\n      <span class="text"><slot></slot></span>\n      <span class="mdw-spinner-container"></span>\n      <div class="mdw-ripple mdw-fab-ripple"></div>\n    '}get dense(){return this.classList.contains("mdw-dense")}get spinnerContainer(){return this._spinnerContainer||(this._spinnerContainer=this.shadowRoot.querySelector(".mdw-spinner-container")),this._spinnerContainer}set disabled(e){e||""===e?this.setAttribute("disabled","disabled"):this.removeAttribute("disabled")}get pending(){return this.pending_}setupAsync(){this.hasAttribute("mdw-async")&&this.addEventListener("click",this.bound_asyncClick)}resolve(){!1!==this.pending_&&(this.pending_=!1,this.hideSpinner())}asyncClick(e){!0!==this.pending_&&(this.pending_=!0,this.showSpinner())}get spinnerStyle(){return this.dense?"position: absolute; left: calc(50% - 12px); top: 8px;":"position: absolute; left: calc(50% - 16px); top: 12px;"}get spinnerDiameter(){return this.dense?24:32}showSpinner(){this._showSpinner=!0,this.classList.add("mdw-show-spinner");const e=this.classList.contains("mdw-primary")||this.classList.contains("mdw-secondary")||this.classList.contains("mdw-error");this.spinnerContainer.innerHTML=`<mdw-circular-progress mdw-mode="indeterminate" mdw-diameter="${this.spinnerDiameter}" class="${e?"mdw-white":"mdw-grey"}" style="${this.spinnerStyle}"></mdw-circular-progress>`}hideSpinner(){this._showSpinner=!1,this.classList.remove("mdw-show-spinner"),this.spinnerContainer.innerHTML=""}styles(){return"\n      :host(.mdw-show-spinner) span.text {\n        opacity: 0;\n      }\n\n      :host-context(.mdw-density-comfortable) .mdw-spinner-container mdw-circular-progress {\n        top: 3px !important;\n      }\n\n      :host-context(.mdw-density-compact) .mdw-spinner-container mdw-circular-progress {\n        top: 0 !important;\n      }\n\n      /* --- Ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n      }\n\n      .mdw-fab-ripple,\n      .mdw-fab-focus-overlay {\n        border-radius: inherit;\n        position: absolute;\n        top: 0;\n        left: 0;\n        bottom: 0;\n        right: 0;\n        pointer-events: none;\n      }\n\n\n      :host(.mdw-primary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n      :host(.mdw-secondary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n      :host(.mdw-error) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n    "}}),customElements.define("mdw-icon",class extends g{constructor(){super(),this.cloneTemplate(),this.hasAttribute("mdw-src")&&this.render()}get src(){return this.getAttribute("mdw-src")}styles(){return"\n      :host {\n        font-family: 'Material Icons';\n        font-weight: normal;\n        font-style: normal;\n        font-size: 24px;\n        line-height: 1;\n        letter-spacing: normal;\n        text-transform: none;\n        display: inline-block;\n        white-space: nowrap;\n        word-wrap: normal;\n        direction: ltr;\n        font-feature-settings: 'liga';\n        -webkit-font-feature-settings: 'liga';\n        -webkit-font-smoothing: antialiased;\n      }\n\n      :host img {\n        width: 24px;\n        height: 24px;\n      }\n\n\n      :host(.mdw-primary) {\n        color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-secondary) {\n        color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-error) {\n        color: var(--mdw-theme-error);\n      }\n    "}template(){const e=this.src;return e?`<img src="${e}"></img>`:"<slot></slot>"}}),customElements.define("mdw-linear-progress",class extends g{constructor(){super(),this.cloneTemplate()}connectedCallback(){null===this.percent&&this.classList.add("mdw-query")}static get observedAttributes(){return["mdw-percent"]}attributeChangedCallback(e,t,n){switch(e){case"mdw-percent":this.percent=n}}get bar(){return this._bar||(this._bar=this.shadowRoot.querySelector(".mdw-bar")),this._bar}get percent(){return this.getAttribute("mdw-percent")}set percent(e){e<0&&(e=0),e>100&&(e=100),this.bar.style.width=`${e}%`}template(){return'\n      <div class="mdw-bar"></div>\n    '}styles(){return"\n      :host {\n        display: block;\n        position: relative;\n        width: 100%;\n        height: 4px;\n        padding-top: 0;\n        margin-bottom: 0;\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.18);\n      }\n\n      .mdw-bar {\n        position: absolute;\n        left: 0;\n        top: 0;\n        bottom: 0;\n        width: 100%;\n        height: 4px;\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-white) {\n        background-color: rgba(255, 255, 255, 0.18);\n      }\n\n      :host(.mdw-white) .mdw-bar {\n        background-color: white;\n      }\n\n      :host(.mdw-grey) {\n        background-color: rgba(50, 50, 50, 0.18);\n      }\n\n      :host(.mdw-grey) .mdw-bar {\n        background-color: grey;\n      }\n\n      :host(.mdw-secondary) {\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.18);\n      }\n\n      :host(.mdw-secondary) .mdw-bar {\n        background-color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-error) {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.18);\n      }\n\n      :host(.mdw-error) .mdw-bar {\n        background-color: var(--mdw-theme-error);\n      }\n\n\n      :host(.mdw-query) .mdw-bar {\n        transition: all 0.2s linear;\n        animation: query .8s infinite cubic-bezier(0.390, 0.575, 0.565, 1.000);\n      }\n\n      @keyframes query {\n        0% {\n          opacity: 1;\n          transform: translateX(35%) scale(.3, 1);\n        }\n        100% {\n          opacity: 0;\n          transform: translateX(-50%) scale(0, 1);\n        }\n      }\n    "}}),customElements.define("mdw-list",class extends g{constructor(){super(),this.selectedIndexes_=[]}static get observedAttributes(){return["mdw-select","mdw-select-onclick"]}attributeChangedCallback(e,t,n){switch(e){case"mdw-select":this.selectType=n;break;case"mdw-select-onclick":this.selectOnclick=null!==n}}set selectOnclick(e){this.selectOnclick_=e}get selectOnclick(){return this.selectOnclick_}set selectType(e){["single","multiple"].includes(e)||console.warn('mdw-list[mdw-select] attribute - only accepts "single" or "multiple"'),this.selectType_=e}get selectType(){return this.selectType_}get selected(){return[].concat(this.selectedIndexes_)}deselectAll(){[...this.querySelectorAll("mdw-list-item")].forEach(e=>e.deselect()),this.selectedIndexes_=[]}itemSelected(e){const t=Array.prototype.indexOf.call(this.children,e)-1;if("single"===this.selectType_){const e=[...this.children];this.selectedIndexes_.forEach(t=>e[t].deselect()),this.selectedIndexes_=[]}this.selectedIndexes_.push(t),this.handleChange()}itemDeselected(e){const t=Array.prototype.indexOf.call(this.children,e)-1;this.selectedIndexes_.splice(this.selectedIndexes_.indexOf(t),1),this.handleChange()}handleChange(){this.dispatchEvent(new CustomEvent("change",this))}closeExpanded(){this.dispatchEvent(new CustomEvent("MDWList:closeExpanded",this))}}),customElements.define("mdw-list-item",class extends g{constructor(){super(),this.bound_hrefClick=this.hrefClick.bind(this),this.bound_onSelect=this.onSelect.bind(this),this.bound_onclickSelect=this.onclickSelect.bind(this),this.bound_checkHREFCurrent=this.checkHREFCurrent.bind(this)}get list(){return this.parentNode}get expanded(){return this.querySelector("mdw-list-item-expanded")}get key(){return this.getAttribute("mdw-key")}isSelect(){return["single","multiple"].includes(this.list.selectType)}selectOnclick(){return!!this.list.selectOnclick}connectedCallback(){this.connectRipple(),this.connectHREF(),this.connectSelect();const e=this.expanded;e&&requestAnimationFrame(()=>{e.listItem=this})}disconnectedCallback(){this.ripple&&this.ripple.destroy(),this.removeEventListener("click",this.bound_hrefClick),this._selectEl&&this._selectEl.removeEventListener("change",this.bound_onSelect),this.removeEventListener("click",this.bound_onclickSelect),window.removeEventListener("hashchange",this.bound_checkHREFCurrent)}expand(){const e=this.expanded;e&&e.open()}connectRipple(){const e=this.querySelector(".mdw-ripple");e&&(this.ripple=new f({element:e,triggerElement:this}),this.classList.add("mdw-has-ripple"))}connectHREF(){this.hasAttribute("href")&&(this.checkHREFCurrent(),window.addEventListener("hashchange",this.bound_checkHREFCurrent),this.addEventListener("click",this.bound_hrefClick))}checkHREFCurrent(){if(!this.hasAttribute("href"))return;const e=document.location.href,t=document.location.hash;e===this.getAttribute("href")||e===this.getAttribute("href-alt")||t===this.getAttribute("href")||t===this.getAttribute("href-alt")?this.classList.add("mdw-current-link"):this.classList.remove("mdw-current-link")}hrefClick(){"_blank"!==this.getAttribute("target")?document.location.href=this.getAttribute("href"):window.open(this.getAttribute("href"),"_blank")}onSelect(e){e.target.checked?this.list.itemSelected(this):this.list.itemDeselected(this)}onclickSelect(e){this.selectOnclick()&&e.target!==this._selectEl&&(this._selectEl.checked=!this._selectEl.checked)}connectSelect(){this.isSelect()&&(this._selectEl=this.querySelector("mdw-checkbox"),this._selectEl&&this._selectEl.addEventListener("change",this.bound_onSelect),this.selectOnclick()&&this.addEventListener("click",this.bound_onclickSelect))}deselect(){this._selectEl.checked=!1}}),customElements.define("mdw-menu",class extends g{constructor(){super(),this.bound_onClick=this.onClick.bind(this),this.bound_onPanelClick=this.onPanelClick.bind(this),!0===a.isMobile?this.createSheet():this.createPanel()}connectedCallback(){this.button.addEventListener("click",this.bound_onClick),!0!==a.isMobile&&(this.classList.add("mdw-panel--container"),this.panel.classList.add("mdw-menu"))}onClick(){!0!==a.isMobile?(this.panel.setPosition(this.panelPosition),this.panel.autoPosition(),this.panel.clickBodyToClose(),this.panel.open(!0),this.panel.addEventListener("click",this.bound_onPanelClick)):this.sheet.open()}onPanelClick(){this.panel.close()}set panelPosition(e){this.panelPosition_=e}get panelPosition(){return this.panelPosition_||"inner-top inner-left"}get button(){return this.children[0]}get contentElement(){return this.querySelector("mdw-menu-content")}get panel(){return this.querySelector("mdw-panel")}get sheet(){return this.querySelector("mdw-sheet")}createSheet(){this.insertAdjacentHTML("beforeend",`\n      <mdw-sheet mdw-modal>\n        <mdw-sheet-content>\n          ${this.contentElement.innerHTML}\n        </mdw-sheet-content>\n      </mdw-sheet>\n    `),this.contentElement.remove()}createPanel(){this.contentElement&&(this.insertAdjacentHTML("beforeend",`\n      <mdw-panel>\n        ${this.contentElement.innerHTML}\n      </mdw-panel>\n    `),this.contentElement.remove())}}),customElements.define("mdw-navigation-rail",class extends g{constructor(){super(),o&&this.hasAttribute("mdw-desktop-only")?(this.style.display="none",this.style.pointerEvents="none"):(this.bound_routeChange=this.routeChange.bind(this),document.body.classList.add("mdw-has-navigation-rail"))}connectedCallback(){window.addEventListener("hashchange",this.bound_routeChange),window.addEventListener("DOMContentLoaded",this.bound_routeChange)}disconnectedCallback(){window.removeEventListener("hashchange",this.bound_routeChange),window.removeEventListener("DOMContentLoaded",this.bound_routeChange)}get path(){let e=window.location.hash.replace(/.*#/,"");return e.indexOf("?")>-1&&(e=e.split("?")[0]),"/"!==e.charAt(0)&&(e="/"+e),e}routeChange(){this.querySelectorAll(".mdw-current-link").forEach(e=>e.classList.remove("mdw-current-link"));let e=this.querySelectorAll(`[href="#${this.path}"]`);e&&0!==e.length||(e=this.querySelectorAll(`[alt-href="#${this.path}"]`)),e.forEach(e=>e.classList.add("mdw-current-link"))}}),customElements.define("mdw-panel-draggable-header",class extends g{constructor(){super(),this.bound_onDrag=this.onDrag.bind(this),this.cloneTemplate()}get draggableElement(){return this.shadowRoot.querySelector(".mdw-panel-draggable-header")}get panel(){return this.parentNode}connectedCallback(){c(this.draggableElement,this.bound_onDrag)}disconnectedCallback(){h(this.draggableElement,this.bound_onDrag)}onDrag(e){switch(e.state){case l.start:this._initialLeft=parseInt((this.panel.style.left||"0").replace("px","")),this._initialTop=parseInt((this.panel.style.top||"0").replace("px",""));break;case l.move:const t=this.panel;t.style.left=`${this._initialLeft+e.distance.x}px`,t.style.top=`${this._initialTop+e.distance.y}px`}}template(){return'\n      <div class="mdw-panel-draggable-header">\n        <slot></slot>\n        <mdw-icon>close</mdw-icon>\n      </div>\n    '}styles(){return"\n      .mdw-panel-draggable-header {\n        display: flex;\n        align-items: center;\n        justify-content: space-between;\n        cursor: move;\n        padding: 12px;\n        user-select: none;\n        -webkit-user-select: none;\n      }\n      mdw-icon {\n        cursor: pointer;\n      }\n    "}}),customElements.define("mdw-panel",class extends g{constructor(){super(),this.FOCUSABLE_ELEMENTS=["button:not(:disabled)",'[href]:not([aria-disabled="true"])',"input:not(:disabled)","select:not(:disabled)","textarea:not(:disabled)",'[tabindex]:not([tabindex="-1"]):not([aria-disabled="true"])'].join(", "),this._clickOutsideClose=!1,this._boundHandleBodyClick=this._handleBodyClick.bind(this),this._boundHandleKeydown=this._handleKeydown.bind(this),this.bound_close=this.close.bind(this),this._clickOutsideCloseIgnorElement=[],this._autoPosition=!1,this._animationConfig={type:"scale",opacity:!0},this.bound_onOpenTransitionEnd=this.onOpenAnimationEnd.bind(this)}connectedCallback(){this.classList.add("mdw-upgraded"),this.transformPropertyName=a.transformPropertyName}disconnectedCallback(){this.removeBodyClickEvent_(),this.removeKeydownEvent_(),clearTimeout(this._openAnimationEndTimerId),clearTimeout(this._closeAnimationEndTimerId),cancelAnimationFrame(this._animationRequestId)}static get observedAttributes(){return["mdw-position"]}attributeChangedCallback(e,t,n){switch(e){case"mdw-position":this._position=n}}set clickOutsideClose(e){this._clickOutsideClose=e}set setQuickOpen(e){this._isQuickOpen=e}get position(){return this._position}fullscreen(){this.classList.add("mdw-fullscreen")}setPosition(e){const t=e.split(" ");this._position=`${t[0]||"top"} ${t[1]||"left"}`,this.setAttribute("mdw-position",this._position),this._positionSet=!0}autoPosition(){this._autoPosition=!0}setAnimation(e){this._animationConfig=e}clickBodyToClose(){this._clickOutsideClose=!0}isOpen(){return this._isOpen}onOpenAnimationEnd(){this.style.transition="",this.style.transformOrigin="",this.style.overflow="",this.style.maxHeight="",this.classList.remove("mdw-panel--animating-open"),this.removeEventListener("transitionend",this.bound_onOpenTransitionEnd),this.notifyOpen()}open(e){void 0!==e&&(this._clickOutsideClose=e);const t=this.querySelectorAll(this.FOCUSABLE_ELEMENTS);this._firstFocusableElement=t[0],this._lastFocusableElement=t[t.length-1],this.saveFocus(),this._isQuickOpen?(this.classList.add("mdw-open"),this._isHoisted?this.setHoistedPosition():this.setPositionStyle()):(this.prepareAnimation(),this._isHoisted?this.setHoistedPosition():this.setPositionStyle(),this._animationRequestId=this._runNextAnimationFrame(()=>{switch(this._animationConfig.fullscreen&&this.classList.add("mdw-fullscreen"),this._animationConfig.type){case"height":this.style.transition="max-height .22s cubic-bezier(0,0,.2,1), transform .22s cubic-bezier(0,0,.2,1), opacity .22s linear",this.style.maxHeight=this.classList.contains("mdw-fullscreen")?"100%":`${this.scrollHeight}px`,this.style.transform="";break;case"scale":default:this.style.transition="transform .1s cubic-bezier(0,0,.2,1), opacity 0.1s linear",this.style.transform=""}this.style.opacity=1,this.addEventListener("transitionend",this.bound_onOpenTransitionEnd)})),this._addBodyClickEvent(),this._addKeydownEvent(),this.addEventListener("MDWPanel:close",this.bound_close),this._isOpen=!0}prepareAnimation(){switch(this.classList.add("mdw-open"),this.classList.add("mdw-panel--animating-open"),this._animationConfig.target&&this._animationConfig.fullscreen&&(this.style.width="100%"),this._animationConfig.type){case"height":switch(this.style.overflow="hidden",this.style.maxHeight=this._animationConfig.target?`${this._animationConfig.target.offsetHeight}px`:"0",this._animationConfig.origin){case"center":let e=this.classList.contains("mdw-fullscreen")?window.innerHeight/2:this.scrollHeight/2;this._animationConfig.target&&(e=this._animationConfig.target.getBoundingClientRect().y),this.style.transform=`translateY(${e}px)`;break;case"top":default:this._animationConfig.target&&(e=this._animationConfig.target.offsetTop,this.style.transform=`translateY(${e}px)`)}break;case"scale":default:this.style.transform="scale(0.9)",this.style.transformOrigin=this._animationConfig.origin||"center"}this._animationConfig.opacity&&(this.style.opacity=0)}async close(e){return new Promise(t=>{e&&e.stopPropagation(),this.removeEventListener("MDWPanel:close",this.bound_close),this._isQuickOpen?(this.classList.remove("mdw-open"),this.resetPosition(),t()):(this.classList.add("mdw-panel--animating-closed"),this.removeBodyClickEvent_(),this._animationRequestId=this._runNextAnimationFrame(()=>{this.classList.remove("mdw-open"),this._closeAnimationEndTimerId=setTimeout(()=>{this._closeAnimationEndTimerId=0,this.classList.remove("mdw-panel--animating-closed"),this.resetPosition(),this.notifyClose(),t()},75)})),this.removeKeydownEvent_(),this._isOpen=!1;const n=this.isFocused(),i=document.activeElement&&this.contains(document.activeElement);(n||i)&&this.restoreFocus()})}_runNextAnimationFrame(e){cancelAnimationFrame(this._animationFrame),this._animationFrame=requestAnimationFrame(()=>{this._animationFrame=0,clearTimeout(this._animationTimer),this._animationTimer=setTimeout(e,0)})}isFocused(){return document.activeElement===this}saveFocus(){this._previousFocus=document.activeElement}restoreFocus(){this.contains(document.activeElement)&&this._previousFocus&&this._previousFocus.focus&&this._previousFocus.focus()}focusFirstElement(){this._firstFocusableElement&&this._firstFocusableElement.focus&&this._firstFocusableElement.focus()}focusLastElement(){this._lastFocusableElement&&this._lastFocusableElement.focus&&this._lastFocusableElement.focus()}isFirstElementFocused(){this._firstFocusableElement&&(this._firstFocusableElement,document.activeElement)}isLastElementFocused(){this._lastFocusableElement&&(this._lastFocusableElement,document.activeElement)}_addBodyClickEvent(){this._clickOutsideClose&&setTimeout(()=>{this.hasBodyEvent=!0,document.body.addEventListener("click",this._boundHandleBodyClick)},0)}removeBodyClickEvent_(){this.hasBodyEvent&&document.body.removeEventListener("click",this._boundHandleBodyClick),this.hasBodyEvent=!1}_addKeydownEvent(){this.hasKeydownEvent=!0,document.body.addEventListener("keydown",this._boundHandleKeydown)}removeKeydownEvent_(){this.hasKeydownEvent&&document.body.removeEventListener("keydown",this._boundHandleKeydown),this.hasKeydownEvent=!1}ignoreElementOnClickToClose(e){this._clickOutsideCloseIgnorElement.push(e)}_handleBodyClick(e){const t=e.target;this._clickOutsideCloseIgnorElement.includes(t)||this.contains(t)||(this.removeBodyClickEvent_(),this.close())}_handleKeydown(e){const{key:t,keyCode:n,shiftKey:i}=e,s="Tab"===t||9===n;"Escape"===t||27===n?this.close():s&&(this.isLastElementFocused()&&!i?(this.focusFirstElement(),e.preventDefault()):this.isFirstElementFocused()&&i&&(this.focusLastElement(),e.preventDefault()))}notifyClose(){this.dispatchEvent(new Event("MDWPanel:closed",this))}notifyOpen(){this.dispatchEvent(new Event("MDWPanel:opened"),this)}hoistToBody(e){this._isHoisted||(this._container=e||this.parentNode,document.body.appendChild(this),this.classList.add("mdw-panel-hoisted"),this._isHoisted=!0)}_autoPositionHoisted(){if(!this._autoPosition)return;const e=window.innerHeight,t=(this.getBoundingClientRect(),this.offsetHeight);let n=this.offsetTop;if(n+t>e&&t<=e){let i=n-(e-t);i>20?i+=10:i/=2,n-=i}this.style.top=`${n}px`}setHoistedPosition(){const e=this._container.getBoundingClientRect();if(this.style.top=`${e.top}px`,this.style.left=`${e.left}px`,this._positionSet){let t=0,n=0;this.style.top=`${t}px`,this.style.left=`${n}px`,setTimeout(()=>{const{clientWidth:i,clientHeight:s}=document.documentElement,o=this.offsetHeight,r=this.offsetWidth,a=(this.position||" ").split(" "),d=a[0],l=a[1];switch(d){case"top":t=0;break;case"inner-top":t=e.y+12;break;case"bottom":t=s;break;case"center":t=s/2-o/2;break;case"inner-bottom":t=s-o-12}switch(l){case"left":n=0;break;case"inner-left":n=e.x+12;break;case"right":n=i;break;case"inner-right":n=i-r-12;break;case"center":n=i/2-r/2}this.style.width=`${this.width}px`,this.style.top=`${t}px`,this.style.left=`${n}px`},0)}else this._autoPositionHoisted()}setPositionStyle(e){e?this._parentOverride=e:this._parentOverride&&(e=this._parentOverride);this.position;let t=0,n=0;if(e)t=e.offsetWidth,n=e.offsetHeight;else{let e=this.parentNode;"MDW-SNACKBAR"===e.nodeName&&(e=e.parentNode);const i=e.getBoundingClientRect();t=i.width,n=i.height}const i=this.offsetWidth,s=this.offsetHeight,o=(this.position||" ").split(" "),r=o[0],a=o[1];let d=0,l=0;switch(r){case"top":d=-s;break;case"bottom":d=n;break;case"center":d=n/2-s/2;break;case"inner-bottom":d=n-s}switch(a){case"left":l=-i;break;case"right":l=t;break;case"inner-right":l=t-i;break;case"center":l=t/2-i/2}if(this._autoPosition){const{clientWidth:e,clientHeight:o}=document.documentElement,{x:r,y:a}=this.getBoundingClientRect();a+s>o&&(d=n-s),r+i>e&&(l=t-i)}this.style.top=`${parseInt(d)}px`,this.style.left=`${parseInt(l)}px`}resetPosition(){this.style.top="",this.style.left=""}}),customElements.define("mdw-radio",class extends g{constructor(){super(),this.insertAdjacentHTML("beforeend",this.rippleTemplate())}connectedCallback(){this.parentNode.initialValue===this.value&&(this.input.checked=!0),this.input.hasAttribute("type")||this.input.setAttribute("type","radio"),this.input.hasAttribute("name")||this.input.setAttribute("name",this.name),this.ripple=new f({element:this.querySelector(".mdw-ripple"),triggerElement:[this.input],radius:20,centered:!0})}disconnectedCallback(){this.ripple.destroy()}get value(){return this.input.value}get input(){return this.querySelector("input")}get name(){return this.parentNode&&this.parentNode.hasAttribute("name")?this.name_=this.parentNode.getAttribute("name"):this.hasAttribute("name")&&(this.name_=this.getAttribute("name")),this.name_||(this.name_=a.uid(),this.parentNode?this.parentNode.setAttribute("name",this.name_):this.setAttribute("name",this.name_)),this.name_}rippleTemplate(){return'\n      <div class="mdw-radio-background">\n        <div class="mdw-radio__outer-circle"></div>\n        <div class="mdw-radio__inner-circle"></div>\n      </div>\n      <div class="mdw-ripple mdw-radio-ripple"></div>\n    '}}),customElements.define("mdw-radio-group",class extends g{constructor(){super(),this.initialValue=this.getAttribute("mdw-value")}}),customElements.define("mdw-select",class extends g{constructor(){super(),this.setupLabel_(),this._isEnhanced&&this.prepareEnhance_(),this.classList.add("mdw-no-animation"),this.cloneTemplate(!0),this.bound_onFocus=this.onFocus.bind(this),this.bound_onBlur=this.onBlur.bind(this),this.bound_onClick=this.onClick.bind(this),this.bound_onChange=this.onChange.bind(this),this.bound_onPanelClick=this.onPanelClick.bind(this),this.bound_onKeyDown=this.onKeyDown.bind(this)}connectedCallback(){this._isEnhanced?(this._selected&&(this.value=this._selected.value),this.shadowRoot.querySelector("render-block").addEventListener("click",this.bound_onClick),document.body.addEventListener("keydown",this.bound_onKeyDown)):(this.selectElement.addEventListener("focus",this.bound_onFocus),this.selectElement.addEventListener("blur",this.bound_onBlur),this.selectElement.addEventListener("change",this.bound_onChange)),this.onChange(),setTimeout(()=>{this.classList.remove("mdw-no-animation"),this._isEnhanced&&(this.panel.style.minWidth=`${this.offsetWidth}px`)},0)}disconnectedCallback(){this._isEnhanced?(this.shadowRoot.querySelector("render-block").removeEventListener("click",this.bound_onClick),document.body.removeEventListener("keydown",this.bound_onKeyDown)):(this.selectElement.removeEventListener("focus",this.bound_onFocus),this.selectElement.removeEventListener("blur",this.bound_onBlur),this.selectElement.removeEventListener("change",this.bound_onChange))}get value(){return this._isEnhanced?this._value:this.selectElement.value||this._value}set value(e){this._value=e,this.onChange(),this.dispatchEvent(new Event("change"))}get selectElement(){return a.querySlotted(this,"select")}get label(){return this.shadowRoot.querySelector("label")}get labelWidth(){return.9*this.label.offsetWidth}get enhacedElementId(){return this._enhacedElementId||(this._enhacedElementId=`select-enhanced-${a.uid()}`),this._enhacedElementId}get panel(){return document.querySelector(`#${this.enhacedElementId}`)}get sheet(){return document.querySelector(`#${this.enhacedElementId}`)}get _isEnhanced(){return null!==this.getAttribute("mdw-enhanced")}get outlined(){return[].slice.apply(this.classList||[]).includes("mdw-outlined")}get notch(){return this._notch||(this._notch=this.shadowRoot.querySelector(".mdw-outlined-notch")),this._notch}setupLabel_(){const e=this.querySelector("label");e&&(this._labelText=e.innerText,e.remove())}prepareEnhance_(){this._optionsMap=[...this.querySelectorAll("option")].map(e=>({text:e.innerText,value:e.value,selected:e.hasAttribute("selected")})),this._selected=this._optionsMap.filter(({selected:e})=>!0===e)[0]||{text:"",value:""};const e=this.querySelector("select");if(e){const t=e.getAttribute("onchange");t&&this.setAttribute("onchange",t),e.remove()}a.isMobile?this.prepareSheet_():this.preparePanel_()}preparePanel_(){const e=`\n      <mdw-panel id="${this.enhacedElementId}" mdw-position="bottom inner-left" class="mdw-panel-hoisted">\n        <mdw-list>\n          ${this._optionsMap.map(({text:e,value:t,selected:n})=>`\n            <mdw-list-item value="${t}"${n?" selected":""}>${e}</mdw-list-item>\n          `).join("\n")}\n        </mdw-list>\n      </mdw-panel>\n    `;document.body.insertAdjacentHTML("beforeend",e);const t=this.panel;t.setAnimation({target:this.querySelector("select"),type:"scale",origin:"top",opacity:!0}),t.hoistToBody(this)}prepareSheet_(){const e=`\n      <mdw-sheet mdw-modal id=${this.enhacedElementId}>\n        <mdw-sheet-content>\n          <mdw-list>\n            ${this._optionsMap.map(({text:e,value:t,selected:n})=>`\n              <mdw-list-item value="${t}"${n?" selected":""}>${e}</mdw-list-item>\n            `).join("\n")}\n          </mdw-list>\n        </mdw-sheet-content>\n      </mdw-sheet>\n    `;document.body.insertAdjacentHTML("beforeend",e)}onFocus(){this.classList.add("mdw-focused"),this.outlined&&(this.notch.style.width=this.labelWidth+"px")}onBlur(){this.classList.remove("mdw-focused"),this.classList.toggle("mdw-not-empty",this.value),this._isEnhanced&&(a.isMobile?(this.sheet.removeEventListener("MDWSheet:closed",this.bound_onBlur),this.sheet.removeEventListener("click",this.bound_onPanelClick)):(this.panel.removeEventListener("MDWPanel:closed",this.bound_onBlur),this.panel.removeEventListener("click",this.bound_onPanelClick))),a.unlockPageScroll()}onChange(){this.value&&this.label?(this.label.classList.add("mdw-select--float-above"),this.label.classList.remove("mdw-empty-no-float"),this.outlined&&(this.notch.style.width=this.labelWidth+"px")):(this.label.classList.remove("mdw-select--float-above"),this.label.classList.add("mdw-empty-no-float"),this.outlined&&(this.notch.style.width="0"))}onClick(e){if(this._focusIndex,this.onFocus(),a.isMobile){const e=this.sheet;e.open(),e.addEventListener("MDWSheet:closed",this.bound_onBlur),e.addEventListener("click",this.bound_onPanelClick);const t=e.querySelector(".mdw-focused");t&&t.classList.remove("mdw-focused");const n=e.querySelector("[selected]");n&&n.classList.add("mdw-focused")}else{const e=this.panel;e.autoPosition(),e.open(!0),e.addEventListener("MDWPanel:closed",this.bound_onBlur),e.addEventListener("click",this.bound_onPanelClick);const t=e.querySelector(".mdw-focused");t&&t.classList.remove("mdw-focused");const n=e.querySelector("[selected]");n&&n.classList.add("mdw-focused")}a.lockPageScroll()}onPanelClick(e){if(!e.target.hasAttribute("value"))return;this.value=e.target.getAttribute("value"),this.setSelectedText(e.target.innerText);const t=this.panel.querySelector("[selected]");t&&t.removeAttribute("selected"),e.target.setAttribute("selected",""),this.panel.close()}setSelectedText(e){this.shadowRoot.querySelector(".mdw-select__selected-text").innerText=e}get internalStylesFile(){return"./internal.css"}template(){return`\n      <i class="mdw-select__icon"></i>\n      ${this._isEnhanced?`\n        <div class="mdw-select__selected-text">${this._selected.text}</div>\n      `:"<slot></slot>"}\n      <label>${this._labelText}</label>\n      ${this.outlined?"":'<div class="mdw-line-ripple"></div>'}\n      ${this.outlined?'\n        <div class="mdw-outlined-border-container">\n          <div class="mdw-outlined-leading"></div>\n          <div class="mdw-outlined-notch"></div>\n          <div class="mdw-outlined-trailing"></div>\n        </div>\n      ':""}\n    `}onKeyDown(e){if(this.panel.isOpen())switch(e.keyCode){case 40:case 39:this.focusNext(),e.preventDefault();break;case 38:case 37:this.focusPrevious(),e.preventDefault();break;case 13:this.selectFocused(),e.preventDefault();break;default:if(e.keyCode>=31||e.keyCode<=90){const t=this.keyboardSearchNodes(e.keyCode);void 0!==t&&this.selectNode(t),e.stopPropagation(),e.preventDefault()}}}keyboardSearchNodes(e){void 0!==this._clearSearchTimeout&&clearTimeout(this._clearSearchTimeout),this._clearSearchTimeout=setTimeout(()=>{this._clearSearchTimeout=void 0,this._keyboardSearchStr="",this._keyboardOptionNames=void 0},300),void 0===this._keyboardSearchStr&&(this._keyboardSearchStr=""),this._keyboardSearchStr+=String.fromCharCode(e);const t=new RegExp("^"+this._keyboardSearchStr,"i");this._keyboardOptionNames||(this._keyboardOptionNames=[...this.panel.querySelectorAll("mdw-list-item")].map(e=>e.innerText));const n=this._keyboardOptionNames.length;let i=0;for(;i<n;){if(t.test(this._keyboardOptionNames[i]))return i;i+=1}}selectNode(e){const t=[...this.panel.querySelectorAll("mdw-list-item")];this._focusIndex=e,this._focusedOption&&this._focusedOption.classList.remove("mdw-focused"),this._focusedOption=t[this._focusIndex],this._focusedOption.classList.add("mdw-focused")}focusNext(){if(!this.panel.isOpen())return;const e=[...this.panel.querySelectorAll("mdw-list-item")];if(void 0===this._focusIndex){const t=e.findIndex(e=>e.classList.contains("mdw-focused"));t>=0&&(this._focusedOption=e[t]),this._focusIndex=t<=0?1:t+1}else this._focusIndex+=1;this._focusIndex>e.length-1&&(this._focusIndex=e.length-1),this._focusedOption&&this._focusedOption.classList.remove("mdw-focused"),this._focusedOption=e[this._focusIndex],this._focusedOption.classList.add("mdw-focused")}focusPrevious(){if(!this.panel.isOpen())return;const e=[...this.panel.querySelectorAll("mdw-list-item")];void 0===this._focusIndex?this._focusIndex=0:this._focusIndex-=1,this._focusIndex<=0&&(this._focusIndex=0),this._focusedOption&&this._focusedOption.classList.remove("mdw-focused"),this._focusedOption=e[this._focusIndex],this._focusedOption.classList.add("mdw-focused")}selectFocused(){if(!this.panel.isOpen())return;const e=[...this.panel.querySelectorAll("mdw-list-item")];(null==this._focusIndex||this._focusIndex>e.length-1)&&(this._focusIndex=0),this.onPanelClick({target:e[this._focusIndex]})}styles(){return"\n      :host-context(.mdw-density-comfortable) .mdw-select__icon {\n        bottom: 15px;\n      }\n\n      :host-context(.mdw-density-comfortable) .mdw-select__selected-text {\n        height: 48px;\n        padding-top: 16px;\n      }\n\n      :host(.mdw-outlined):host-context(.mdw-density-comfortable) .mdw-select__icon {\n        bottom: 20px;\n      }\n      \n      :host(.mdw-outlined):host-context(.mdw-density-comfortable) label {\n        top: 18px;\n      }\n\n      :host-context(.mdw-density-compact) .mdw-select__icon {\n        bottom: 12px;\n      }\n\n      :host(.mdw-outlined):host-context(.mdw-density-compact) .mdw-select__icon {\n        top: 18px;\n      }\n\n      :host(.mdw-outlined):host-context(.mdw-density-compact) label {\n        top: 12px;\n      }\n\n      :host-context(.mdw-density-compact) .mdw-select__selected-text {\n        height: 40px;\n        line-height: 1.1rem;\n      }\n\n      :host-context(.mdw-density-compact) ::slotted(select) {\n        height: 40px;\n        padding-top: 12px;\n      }\n\n      :host-context(.mdw-density-compact) label {\n        top: 16px;\n      }\n\n      :host(.mdw-outlined.mdw-focused):host-context(.mdw-density-compact) label,\n      :host(.mdw-outlined):host-context(.mdw-density-compact) label.mdw-select--float-above {\n        transform: translateY(-100%) scale(0.75);\n      }\n\n      ::slotted(label.mdw-empty-no-float) {\n        transform: none;\n      }\n\n      :host(.mdw-focused) .mdw-select__icon {\n        transform: rotate(180deg) translateY(-5px);\n        transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1);\n      }\n\n      .mdw-select__icon {\n        transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1);\n        pointer-events: none;\n        position: absolute;\n        bottom: 23px;\n        left: auto;\n        right: 8px;\n        width: 0;\n        height: 0;\n        border-left: 5px solid transparent;\n        border-right: 5px solid transparent;\n        border-top: 5px solid var(--mdw-theme-on-secondary);\n      }\n\n      ::slotted(select:focus) .mdw-select__icon,\n      :host(.mdw-focused:focus) .mdw-select__icon {\n        transform: rotate(180deg) translateY(-5px);\n        transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1);\n      }\n\n      :host(:not(.mdw-select--disabled)) ::slotted(select),\n      :host(:not(.mdw-select--disabled)) .mdw-select__selected-text {\n        border-bottom-color: rgba(var(--mdw-theme-on-background--rgb), 0.54);\n        color: var(--mdw-theme-on-primary);\n      }\n\n      :host(.mdw-focused:not(.mdw-select--disabled)) ::slotted(select),\n      :host(.mdw-focused:not(.mdw-select--disabled)) .mdw-select__selected-text,\n      :host(:not(.mdw-select--disabled)) ::slotted(select:focus),\n      :host(.mdw-focused:focus:not(.mdw-select--disabled)) .mdw-select__selected-text {\n        border-bottom: 2px solid;\n        border-bottom-color: var(--mdw-theme-primary);\n        height: calc(100% + 1px); /* add 1px to height so the text does not get pushed up by border size change */\n      }\n\n      :host(.mdw-outlined) ::slotted(select),\n      :host(.mdw-outlined) .mdw-select__selected-text {\n        border: none;\n      }\n\n      ::slotted(select),\n      .mdw-select__selected-text {\n        position: absolute;\n        padding: 20px 52px 4px 16px;\n        font-family: Roboto,sans-serif;\n        font-size: 1rem;\n        line-height: 1.75rem;\n        font-weight: 400;\n        letter-spacing: .009375em;\n        text-decoration: inherit;\n        text-transform: inherit;\n        box-sizing: border-box;\n        width: 100%;\n        height: 56px;\n        border: none;\n        border-bottom: 1px solid;\n        outline: none;\n        background-color: transparent;\n        color: inherit;\n        white-space: nowrap;\n        cursor: pointer;\n        appearance: none;\n        -webkit-appearance: none;\n        -moz-appearance: none;\n      }\n\n      /* outlined */\n      :host(.mdw-outlined) ::slotted(select),\n      :host(.mdw-outlined) .mdw-select__selected-text {\n        padding: 12px 52px 12px 16px;\n        display: flex;\n        border: none;\n        background-color: transparent;\n        z-index: 1;\n      }\n      :host(.mdw-outlined) ::slotted(select) {\n        border-radius: 4px;\n      }\n\n      ::slotted(select) {\n        border-radius: 4px 4px 0 0;\n      }\n\n      :host([dir=rtl]) ::slotted(select),\n      ::slotted(select[dir=rtl]),\n      :host([dir=rtl]) .mdw-select__selected-text,\n      .mdw-select__selected-text[dir=rtl] {\n        padding-left: 52px;\n        padding-right: 16px;\n      }\n\n\n      label {\n        font-size: 1rem;\n        line-height: 1.75rem;\n        font-weight: 400;\n        letter-spacing: 0.009375em;\n        text-decoration: inherit;\n        text-transform: inherit;\n        position: absolute;\n        left: 0;\n        transform-origin: left top;\n        transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1), color 150ms cubic-bezier(0.4, 0, 0.2, 1);\n        line-height: 1.15rem;\n        text-align: left;\n        text-overflow: ellipsis;\n        white-space: nowrap;\n        cursor: text;\n        overflow: hidden;\n        will-change: transform;\n        transform: none;\n        pointer-events: none;\n        color: rgba(var(--mdw-theme-on-background--rgb), .6);\n        z-index: 1;\n\n        left: 16px;\n        right: initial;\n        top: 21px;\n      }\n\n      :host(.mdw-focused) label {\n        color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-no-animation) label {\n        transition: none;\n      }\n\n      label:not(.mdw-empty-no-float) {\n        transform: translateY(-70%) scale(0.75);\n      }\n\n      ::slotted(select:focus) + label,\n      label.mdw-select--float-above {\n        transform: translateY(-70%) scale(0.75);\n      }\n\n      :host(.mdw-outlined.mdw-focused) label,\n      :host(.mdw-outlined) label.mdw-select--float-above {\n        transform: translateY(-132%) scale(0.75);\n      }\n\n      :host(.mdw-select--with-leading-icon) label {\n        left: 48px;\n        right: initial;\n      }\n\n      :host(.mdw-outlined) label {\n        left: 15px;\n        right: initial;\n        top: 18px;\n      }\n\n      :host(.mdw-outlined.mdw-select--with-leading-icon) label {\n        left: 36px;\n        right: initial;\n      }\n\n      :host(.mdw-outlined.mdw-select--with-leading-icon) label.mdw-select--float-above {\n        left: 36px;\n        right: initial;\n      }\n\n      .mdw-outlined-border-container {\n        display: -ms-flexbox;\n        display: flex;\n        position: absolute;\n        top: 0;\n        right: 0;\n        left: 0;\n        -webkit-box-sizing: border-box;\n        box-sizing: border-box;\n        width: 100%;\n        max-width: 100%;\n        height: 100%;\n        text-align: left;\n        pointer-events: none;\n      }\n\n      .mdw-outlined-leading {\n        border-radius: 4px 0 0 4px;\n        border-left: 1px solid;\n        border-right: none;\n        width: 12px;\n      }\n\n      .mdw-outlined-notch {\n        -ms-flex: 0 0 auto;\n        flex: 0 0 auto;\n        width: auto;\n        max-width: calc(100% - 12px * 2);\n      }\n\n      .mdw-outlined-trailing {\n        border-left: none;\n        border-right: 1px solid;\n        border-radius: 0 4px 4px 0;\n        -ms-flex-positive: 1;\n        flex-grow: 1;\n      }\n\n      .mdw-outlined-leading,\n      .mdw-outlined-notch,\n      .mdw-outlined-trailing {\n        box-sizing: border-box;\n        height: 100%;\n        border-top: 1px solid;\n        border-bottom: 1px solid;\n        pointer-events: none;\n\n        border-color: rgba(var(--mdw-theme-on-background--rgb), 0.54);\n      }\n\n      .mdw-outlined-notch {\n        border-top: none;\n      }\n\n      :host(.mdw-focused) .mdw-outlined-leading,\n      :host(.mdw-focused) .mdw-outlined-notch,\n      :host(.mdw-focused) .mdw-outlined-trailing,\n      ::slotted(select:focus) .mdw-outlined-leading,\n      ::slotted(select:focus) .mdw-outlined-notch,\n      ::slotted(select:focus) .mdw-outlined-trailing {\n        border-width: 2px;\n        border-color: var(--mdw-theme-primary);\n      }\n\n      :host(.invalid) .mdw-outlined-leading,\n      :host(.invalid) .mdw-outlined-notch,\n      :host(.invalid) .mdw-outlined-trailing {\n        border-color: var(--mdw-theme-error);\n      }\n    "}}),customElements.define("mdw-sheet-bottom-header",class extends g{constructor(){super(),this.parentNode._registerHeader&&this.parentNode._registerHeader(this),this.sheetInstance=this.parentNode,this.bound_close=this.close.bind(this),this.bound_toTop=this.toTop.bind(this),this.innerHTML=this.template(this.isModal,this.title,this.innerHTML);const e=this.querySelector("mdw-header");e&&(this.classList.add("has-standard-header"),e.setAttribute("class",this.classList.toString()))}connectedCallback(){this.closeButton&&this.closeButton.addEventListener("click",this.bound_close),this._hasMdwHeader&&this.mdwHeader.addEventListener("click",this.bound_toTop)}disconnectedCallback(){this.closeButton&&this.closeButton.removeEventListener("click",this.bound_close),this._hasMdwHeader&&this.mdwHeader.removeEventListener("click",this.bound_toTop)}get closeButton(){return this.querySelector("#mdw-sheet-close-action")}get mdwHeader(){return this.querySelector("mdw-header")}get isModal(){return"modal"===this.parentNode.type}get title(){return this._title?this._title:this.hasAttribute("mdw-title")?this.getAttribute("mdw-title"):""}set title(e){this._title=e}close(){if(this.isModal)return this.parentNode.close();this.parentNode.exitFullscreen()}disable(){this.classList.add("mdw-disabled")}show(){this.classList.add("mdw-show"),this.isShowing=!0}hide(){this.classList.remove("mdw-show"),this.isShowing=!1}toTop(){console.log("to top"),this.parentNode._transitionToNearestPosition()}template(e,t,n){const i=(new DOMParser).parseFromString(n,"text/html").querySelector("mdw-sheet-top-bar");return i&&(n=n.replace(i.outerHTML,"")),n&&(this._hasMdwHeader=!0),`\n      ${i?i.outerHTML:`\n        <mdw-sheet-top-bar>\n          <mdw-button id="mdw-sheet-close-action" class="mdw-icon">\n            ${e?"<mdw-icon>close</mdw-icon>":"<mdw-icon>keyboard_arrow_down</mdw-icon>"}\n          </mdw-button>\n          ${t}\n        </mdw-sheet-top-bar>\n      `}\n\n      ${n&&`<mdw-header>\n\n          ${n}\n        </mdw-header>`}\n    `}});class E{constructor(e){this.componentElement=e,this.isAnchored=e.hasAttribute("mdw-anchored")}get initialPosition(){return this.componentElement._headerHeight+window.innerHeight/4}get minimizedPosition(){return this.componentElement._headerHeight}get clientPosition(){return window.innerHeight/4}get isDraggable(){return this.componentElement.contentHeight>this.clientPosition}registerHeader(e){this.headerElement=e,this.headerElement.title=this.title}setupHeader(){const e=this.componentElement.querySelector("mdw-header");this.componentElement.insertAdjacentHTML("afterbegin",`<mdw-sheet-bottom-header mdw-title="${this.componentElement.title}" class="${e.classList.toString()}">${e&&e.innerHTML}</mdw-sheet-bottom-header>`),e&&e.remove()}addBackdrop(){}removeBackdrop(){}handleOnMove({position:e,isAtTop:t,isAboveTop:n,targetingTop:i,targetingInitial:s}){i||t||n?this.headerElement.show():this.headerElement.hide(),t||n?this.headerElement.classList.add("mdw-fullscreen"):this.headerElement.classList.remove("mdw-fullscreen"),n?this.headerElement.classList.add("mdw-is-above-top"):this.headerElement.classList.remove("mdw-is-above-top")}}class x{constructor(e){this.componentElement=e}get initialPosition(){return this.minimizedPosition}get minimizedPosition(){const e=this.componentElement.contentHeight;return Math.min(e,this.clientPosition)}get clientPosition(){return window.innerHeight/2}get isDraggable(){return this.componentElement.contentHeight>this.clientPosition}registerHeader(e){this.headerElement=e,this.headerElement.title=this.title}setupHeader(){this.componentElement.insertAdjacentHTML("afterbegin",`<mdw-sheet-bottom-header mdw-title="${this.componentElement.title}"></mdw-sheet-bottom-header>`)}addBackdrop(){this.backdrop=a.addBackdrop(this.componentElement,()=>{this.componentElement.close()})}removeBackdrop(){this.backdrop&&this.backdrop.remove(),this.backdrop=void 0}handleOnMove({position:e,isAtTop:t,isAboveTop:n,targetingTop:i,targetingInitial:s}){i||t||n?this.headerElement.show():this.headerElement.hide(),t||n?this.headerElement.classList.add("mdw-fullscreen"):this.headerElement.classList.remove("mdw-fullscreen"),n?this.headerElement.classList.add("mdw-is-above-top"):this.headerElement.classList.remove("mdw-is-above-top")}}customElements.define("mdw-sheet-bottom",class extends g{constructor(){switch(super(),this.type){case"modal":this._helpers=new x(this);break;default:this._helpers=new E(this)}this.bound_onTransitionEnd=this._onTransitionEnd.bind(this),this.bound_onTransitionEndClose=this._onTransitionEndClose.bind(this),this.bound_onDrag=this._onDrag.bind(this),this._helpers.setupHeader(),this._setupOverScroll()}get title(){return this.getAttribute("mdw-title")}get type(){return this.hasAttribute("mdw-modal")?"modal":"standard"}get contentElement(){return this.querySelector("mdw-content")}get contentHeight(){return this.contentElement.offsetHeight}get _isDraggable(){return this._helpers.isDraggable}get _headerHeight(){return this._helpers.headerElement.offsetHeight}get _maxScroll(){return this.contentHeight+this._headerHeight}get _topPosition(){const e=window.innerHeight,t=this.contentHeight,n=this._headerHeight;return"standard"===this.type?e-(55-n):e-(t>=e-n?55-n:e-n-t)}get _scrollDistanceRemaining(){return this._maxScroll-this._currentPosition}get _initialPosition(){return this._helpers.initialPosition}get _minimizedPosition(){return this._helpers.minimizedPosition}get _isAnchored(){return this.hasAttribute("mdw-anchored")}disconnectedCallback(){this.classList.contains("mdw-dragging")&&(a.unlockPageScroll(),a.disableUserSelect()),this._cancelTransitions()}open(){this._cancelTransitions(),this.classList.add("mdw-show"),this.style.height=`${this.contentHeight+this._headerHeight}px`,this._positionBottom(),this._isAnchored?this._transitionToPosition(this._minimizedPosition):this._transitionToPosition(this._initialPosition),this._helpers.addBackdrop(),this._isDraggable&&(c(this.contentElement,this.bound_onDrag),this._helpers.headerElement&&c(this._helpers.headerElement,this.bound_onDrag))}async close(){return new Promise(e=>{h(this.contentElement,this.bound_onDrag),this._helpers.headerElement&&h(this._helpers.headerElement,this.bound_onDrag),this._cancelTransitions(),this.classList.add("mdw-animating-close"),this._positionBottom(),this.addEventListener("transitionend",()=>{this._cancelTransitions(),this.classList.remove("mdw-show"),this._helpers.removeBackdrop(),e()},{once:!0})})}minimize(){this._cancelTransitions(),this._transitionToPosition(this._minimizedPosition)}exitFullscreen(){this._cancelTransitions(),this._isAnchored?this._transitionToPosition(this._minimizedPosition):this._transitionToPosition(this._initialPosition)}toggle(){this.classList.contains("mdw-show")?this.close():this.open()}_registerHeader(e){this._helpers.registerHeader(e)}_positionTop(){this._setPosition(this._topPosition)}_positionBottom(){this._currentPosition=0,this.style.top="100%"}_positionInitial(){this._setPosition(this._initialPosition)}_positionMinimized(){this._setPosition(this._minimizedPosition)}_setPosition(e){const t=this._maxScroll;let n=0;if(e>t){const i=100;n=i*Math.log(e-t+i)-i*Math.log(i),e=t}const i=window.innerHeight+(this._headerHeight-55);this._isAtTop=e===i,this._isAboveTop=e>i,this._isAtOrAboveTop=this._isAtTop||this._isAboveTop,this._currentPosition=e,this.style.top=`calc(100% - ${e+n}px)`;const s=window.innerHeight-this._initialPosition,o=!this._isAtOrAboveTop&&e-this._initialPosition>=s/2,r=!this._isAtOrAboveTop&&e-this._initialPosition<s/2;this._helpers.handleOnMove({position:e,isAtTop:this._isAtTop,isAboveTop:this._isAboveTop,targetingTop:o,targetingInitial:r})}_transitionToNearestPosition(){const e=this._currentPosition;if(this._initialOffsetTop===this.offsetTop)return e===this._minimizedPosition?this._transitionToPosition(this._initialPosition):this._transitionToPosition(this._topPosition);const t=(this._topPosition-this._initialPosition)/2;e-this._initialPosition>=t?this._transitionToPosition(this._topPosition):this._transitionToPosition(this._initialPosition)}_transitionToPosition(e){this.classList.add("mdw-animating-open"),requestAnimationFrame(()=>{this._setPosition(e),this.addEventListener("transitionend",this.bound_onTransitionEnd)})}_cancelTransitions(){this.classList.remove("mdw-animating-open"),this.classList.remove("mdw-animating-close"),this.classList.remove("mdw-animating-scroll"),this.style.transitionDuration="",this.removeEventListener("transitionend",this.bound_onTransitionEnd),this.removeEventListener("transitionend",this.bound_onTransitionEndClose)}_onTransitionEnd(){this._initialOffsetTop=this.offsetTop,this._cancelTransitions()}_onTransitionEndClose(){this._cancelTransitions(),this.classList.remove("mdw-show"),this._helpers.removeBackdrop()}_onDrag(e){switch(e.state){case"start":this._cancelTransitions(),this._startPosition=this._currentPosition,this.classList.add("mdw-dragging"),a.lockPageScroll(),a.disableUserSelect();break;case"move":this._setPosition(this._startPosition-e.distance.y);break;case"end":this.classList.add("mdw-animating-scroll"),this._handleScrollEnd(e.velocity.y,e.direction.y),this.addEventListener("transitionend",this.bound_onTransitionEnd),this.classList.remove("mdw-dragging"),a.unlockPageScroll(),a.enableUserSelect()}}_handleScrollEnd(e,t){if(!this._isAtOrAboveTop&&e<-1.1)this._positionTop();else{if(!this._isAtOrAboveTop&&e>.7)return"modal"===this.type?this.close():this.minimize();{const n=Math.abs(e)/3;let i=this._scrollDistanceRemaining*n*-t;i>this._scrollDistanceRemaining&&(i=this._scrollDistanceRemaining);const s=i+this._currentPosition;if(this._isAtOrAboveTop)s<this._topPosition+80?this._positionTop():(this.style.transitionDuration=`${.5*n+.3}s`,this._setPosition(s));else{const e=(this._topPosition-this._initialPosition)/2;s-this._initialPosition>=e?this._positionTop():this._positionInitial()}}}}_setupOverScroll(){this.insertAdjacentHTML("beforeend",'<div class="mdw-sheet-bottom-over-scroll"></div>')}}),customElements.define("mdw-sheet-side",class extends g{constructor(){super(),this._useBackdrop=!this.hasAttribute("mdw-no-backdrop"),this.bound_onSwipe=this.onSwipe.bind(this),this.bound_routeChange=this.routeChange.bind(this)}connectedCallback(){this._isNavigationDrawer=this.classList.contains("mdw-navigation-drawer"),this._isNavigationDrawer&&document.body.classList.add("mdw-has-navigation-drawer"),a.isMobile&&this.setAttribute("mdw-modal",""),this._isNavigationDrawer&&this.isModal&&document.body.classList.add("mdw-navigation-drawer-modal"),this.isModal&&this._isNavigationDrawer&&!this.isHidden&&!this.classList.contains("mdw-show")?this.classList.add("mdw-hide"):this._isNavigationDrawer?document.body.classList.add("mdw-navigation-drawer-open"):this.isModal&&!this.isHidden&&this._useBackdrop&&(this._backdrop=a.addBackdrop(this,()=>this.close(),{sheet:!0})),this._isNavigationDrawer&&(window.addEventListener("hashchange",this.bound_routeChange),window.addEventListener("DOMContentLoaded",this.bound_routeChange))}disconnectedCallback(){this._backdrop&&this._backdrop.remove(),b(document.body,this.bound_onSwipe),this._isNavigationDrawer&&(window.removeEventListener("hashchange",this.bound_routeChange),window.removeEventListener("DOMContentLoaded",this.bound_routeChange))}get isModal(){return this.hasAttribute("mdw-modal")}get isHidden(){return this.classList.contains("mdw-hide")}get isLeft(){return this.classList.contains("mdw-left")||this.classList.contains("mdw-navigation-drawer")}set useBackdrop(e){this._useBackdrop=!!e}get path(){let e=window.location.hash.replace(/.*#/,"");return e.indexOf("?")>-1&&(e=e.split("?")[0]),"/"!==e.charAt(0)&&(e="/"+e),e}routeChange(){this.querySelectorAll(".mdw-current-link").forEach(e=>e.classList.remove("mdw-current-link"));let e=this.querySelectorAll(`[href="#${this.path}"]`);e&&0!==e.length||(e=this.querySelectorAll(`[alt-href="#${this.path}"]`)),e.forEach(e=>e.classList.add("mdw-current-link"))}open(){console.log(this.classList.contains("mdw-hide")),setTimeout(()=>{this.classList.remove("mdw-hide"),this.classList.add("mdw-show"),this.isModal&&(w(document.body,this.bound_onSwipe),this._useBackdrop&&(this._backdrop=a.addBackdrop(this,()=>this.close(),{sheet:!0}))),this._isNavigationDrawer&&document.body.classList.add("mdw-navigation-drawer-open")},10)}async close(){return new Promise(e=>{this.classList.remove("mdw-show"),this.classList.add("mdw-hide"),this._backdrop&&this._backdrop.remove(),b(document.body,this.bound_onSwipe),this._isNavigationDrawer&&document.body.classList.remove("mdw-navigation-drawer-open"),setTimeout(()=>{e()},200)})}toggle(){this.isHidden?this.open():this.close()}onSwipe(e){this.isLeft?-1===e.direction.x&&e.velocity.x<-.8&&this.close():1===e.direction.x&&e.velocity.x>.8&&this.close()}}),customElements.define("mdw-slider",class extends g{constructor(){super(),this.cloneTemplate(),this.bound_onMouseDown=this.onMouseDown.bind(this),this.bound_onMouseUp=this.onMouseUp.bind(this),this.bound_onMouseMove=this.onMouseMove.bind(this),this.bound_onMouseEnter=this.onMouseEnter.bind(this),this.bound_onMouseLeave=this.onMouseLeave.bind(this),this.bound_trackClick=this.trackClick.bind(this),this.bound_onDrag=this.onDrag.bind(this)}connectedCallback(){this.value=this.attrValue,this.thumbContainer.style.left=`${(this.attrValue-this.min)/this.range*this.offsetWidth}px`,this.notchContainer.style.marginLeft=`-${this.offsetWidth-(this.attrValue-this.min)/this.range*this.offsetWidth}px`,this.throttled_dispatchChange=a.rafThrottle(this.dispatchChange),c(this.thumb,this.bound_onDrag)}disconnectedCallback(){this.track.removeEventListener("click",this.bound_trackClick),h(this.thumb,this.bound_onDrag)}static get observedAttributes(){return["value","min","max","step"]}attributeChangedCallback(e,t,n){this[e]=n,["min","max","step"].includes(e)&&this.render()}get min(){return this.min_||0}set min(e){this.min_=parseFloat(e)}get max(){return this.max_||100}set max(e){this.max_=parseFloat(e)}get range(){return this.max-this.min}get step(){return this.step_}set step(e){this.step_=parseFloat(e)}get stepCount(){return this.step?Math.floor(this.range/this.step):0}get attrValue(){let e=parseFloat(this.getAttribute("value")||0);return e<this.min&&(e=this.min),e}get value(){const{width:e}=this.getBoundingClientRect(),t=(this.thumbContainer.style.left||"0px").replace("px","")/e,n=this.range;return this.value_=this.min+t*n,(""+this.step).includes(".")||(this.value_=parseInt(this.value_)),this.value_||0}set value(e){this.value_=parseFloat(e)}get thumb(){return this.shadowRoot.querySelector(".mdw-slider__thumb-hover")}get thumbContainer(){return this.thumbContainer_||(this.thumbContainer_=this.shadowRoot.querySelector(".mdw-slider__thumb-container")),this.thumbContainer_}get notchContainer(){return this.notchContainer_||(this.notchContainer_=this.shadowRoot.querySelector(".mdw-slider__notch-container")),this.notchContainer_}get track(){return this.shadowRoot.querySelector(".mdw-slider__track-container")}trackClick(e){const{left:t,width:n}=this.getBoundingClientRect();let i=e.layerX;e.clientX<t&&(i=0),i>n&&(i=n),this.thumbContainer.style.left=`${this.snap(i,n)}px`,this.notchContainer.style.marginLeft=`-${this.offsetWidth-this.snap(i,n)}px`,this.dispatchChange()}onDrag(e){switch(e.state){case"start":this.classList.add("mdw-pressed"),this.initialX_=parseInt((this.thumbContainer.style.left||"0px").replace("px",""));break;case"move":const{left:t,width:n}=this.getBoundingClientRect();let i=e.distance.x+this.initialX_;i<0&&(i=0),i>n&&(i=n),this.thumbContainer.style.left=`${this.snap(i,n)}px`,this.notchContainer.style.marginLeft=`-${this.offsetWidth-this.snap(i,n)}px`,this.throttled_dispatchChange();break;case"end":this.classList.remove("mdw-pressed")}}onMouseDown(e){this.classList.add("mdw-pressed"),document.addEventListener("mouseup",this.bound_onMouseUp),document.addEventListener("mousemove",this.bound_onMouseMove)}onMouseUp(e){this.classList.remove("mdw-pressed"),document.removeEventListener("mouseup",this.bound_onMouseUp),document.removeEventListener("mousemove",this.bound_onMouseMove)}onMouseMove(e){const{left:t,width:n}=this.getBoundingClientRect();let i=e.layerX;e.clientX<t&&(i=0),i>n&&(i=n),this.thumbContainer.style.left=`${this.snap(i,n)}px`,this.notchContainer.style.marginLeft=`-${this.offsetWidth-this.snap(i,n)}px`,this.throttled_dispatchChange()}onMouseEnter(e){this.classList.add("mdw-hover"),this.thumb.addEventListener("mouseleave",this.bound_onMouseLeave)}onMouseLeave(e){this.classList.remove("mdw-hover"),this.thumb.removeEventListener("mouseleave",this.bound_onMouseLeave)}snap(e,t){if(!this.step)return e;const n=e/t,i=this.range,s=n*i;return(s-s%this.step)/i*t}dispatchChange(){this.dispatchEvent(new Event("change",this))}template(){return`\n      <div class="mdw-slider__track-container">\n        <div class="mdw-slider__track"></div>\n\n        <div class="mdw-slider__notch-container">\n          <div class="mdw-slider__notch-pre-container">\n            ${[...new Array(this.stepCount)].map(e=>'<div class="mdw-slider__notch"></div>').join("\n")}\n          </div>\n\n          <div class="mdw-slider__notch-post-container">\n            ${[...new Array(this.stepCount)].map(e=>'<div class="mdw-slider__notch"></div>').join("\n")}\n          </div>\n        </div>\n      </div>\n      <div class="mdw-slider__thumb-container">\n        <div class="mdw-slider__thumb"></div>\n        <div class="mdw-slider__thumb-hover"></div>\n      </div>\n    `}styles(){return"\n      .mdw-slider__track-container {\n        position: absolute;\n        top: 50%;\n        width: 100%;\n        height: 10px;\n        margin-top: -6px;\n        overflow: hidden;\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      .mdw-slider__track {\n        position: absolute;\n        width: 100%;\n        height: 2px;\n        top: 50%;\n        user-select: none;\n        -webkit-user-select: none;\n        /* background-color: var(--mdw-theme-secondary); */\n      }\n\n      /* :host(.mdw-primary) .mdw-slider__track {\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-error) .mdw-slider__track {\n        background-color: var(--mdw-theme-error);\n      } */\n\n\n      .mdw-slider__thumb-container {\n        position: absolute;\n        top: 50%;\n        left: 0;\n        user-select: none;\n        -webkit-user-select: none;\n        z-index: 2;\n      }\n\n      .mdw-slider__thumb {\n        box-sizing: border-box;\n        width: 12px;\n        height: 12px;\n        border-radius: 50%;\n        margin-top: -50%;\n        z-index: 2;\n        background-color: var(--mdw-theme-secondary);\n        cursor: pointer;\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      :host(.mdw-primary) .mdw-slider__thumb {\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-error) .mdw-slider__thumb {\n        background-color: var(--mdw-theme-error);\n      }\n\n      .mdw-slider__thumb-hover {\n        position: absolute;\n        box-sizing: border-box;\n        top: -12px;\n        left: -6px;\n        width: 24px;\n        height: 24px;\n        border-radius: 50%;\n        transform-origin: center center;\n        transition: opacity .1s ease-out,fill .1s ease-out,\n                    transform .1s ease-out,fill .1s ease-out;\n        opacity: 0;\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.16);\n        cursor: pointer;\n        user-select: none;\n      }\n\n      :host(.mdw-primary) .mdw-slider__thumb-hover {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n\n      :host(.mdw-error) .mdw-slider__thumb-hover {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.16);\n      }\n\n      :host(.mdw-hover) .mdw-slider__thumb-hover {\n        opacity: 1;\n      }\n\n      :host(.mdw-pressed) .mdw-slider__thumb-hover {\n        transform: scale(1.8);\n      }\n\n\n\n\n      /* --- notches --- */\n\n      .mdw-slider__notch-container {\n        display: flex;\n        width: 200%;\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      .mdw-slider__notch-pre-container {\n        width: 100%;\n        height: 2px;\n        display: flex;\n        flex-direction: row;\n        margin-top: 5px;\n        overflow: hidden;\n        z-index: 1;\n        background-color: var(--mdw-theme-secondary);\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      :host(.mdw-primary) .mdw-slider__notch-pre-container {\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-error) .mdw-slider__notch-pre-container {\n        background-color: var(--mdw-theme-error);\n      }\n\n      .mdw-slider__notch-pre-container .mdw-slider__notch {\n        height: 2px;\n        flex: 1;\n        border-left: 3px solid rgba(255, 255, 255, 0.6);\n      }\n\n      .mdw-slider__notch-post-container {\n        width: 100%;\n        height: 2px;\n        display: flex;\n        flex-direction: row;\n        margin-top: 5px;\n        overflow: hidden;\n        z-index: 1;\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.5);\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      :host(.mdw-primary) .mdw-slider__notch-post-container {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.5);\n      }\n\n      :host(.mdw-error) .mdw-slider__notch-post-container {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.5);\n      }\n\n      .mdw-slider__notch-post-container .mdw-slider__notch {\n        height: 2px;\n        flex: 1;\n        border-left: 3px solid rgba(0, 0, 0, 0.6);\n      }\n    "}});const L=new class{constructor(){this.queue=[]}add(e){this.queue.push({el:e}),this.handleQueue()}remove(e,t){this.current&&this.current.el===e?e._close(t):this.queue=this.queue.filter(t=>t.el!==e)}handleQueue(){0!==this.queue.length&&(this.current||(this.current=this.queue.shift(),this.current.el._open(),this.current.el.addEventListener("close",()=>{this.current=void 0,setTimeout(()=>{this.handleQueue()},300)})))}show({message:e,actionLabel:t,position:n}){return new Promise(i=>{const s=this.uid(),o=this.template({id:s,message:e,actionLabel:t});this.topLevelElement.insertAdjacentHTML("beforeend",o);const r=document.querySelector(`#${s}`),a=e=>{i(e.detail.ok),r.removeEventListener("close",a),r.remove()};r.addEventListener("close",a),n&&r.setPosition(n),r.show()})}get topLevelElement(){let e=document.body.querySelector("mdw-content");return e||(e=document.body.querySelector("mdw-body"),e||document.body)}uid(){return`snackbar_${parseInt(99999*Math.random())}`}template({id:e,message:t,actionLabel:n}){return`\n      <mdw-snackbar id="${e}">\n        <mdw-panel>\n          <mdw-snackbar-container>\n            <mdw-snackbar-content>${t}</mdw-snackbar-content>\n            <mdw-snackbar-actions>\n              ${n?`<mdw-button class="mdw-action-button">${n}</mdw-button>`:""}\n              <mdw-button onclick="${e}.close(true)" class="mdw-close-button mdw-icon">\n                <mdw-icon>close</mdw-icon>\n              </mdw-button>\n            </mdw-snackbar-actions>\n          </mdw-snackbar-container>\n        </mdw-panel>\n      </mdw-snackbar>\n    `}};window.MDWSnackbar=L;var C=L;customElements.define("mdw-snackbar",class extends g{constructor(){super(),this.bound_onPanelClose=this.onPanelClose.bind(this),this.panelId=`${this.getAttribute("id")}_panel`}connectedCallback(){this.querySelector("mdw-panel").setAttribute("id",`${this.panelId}`),this.hasBckdrop=!0,this.panel.clickOutsideClose=!1}disconnectedCallback(){this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose)}get panel(){return document.querySelector(`#${this.panelId}`)}get position(){return this.position_||"inner-bottom inner-left"}setPosition(e){const t=e.split(" ");this.position_=`${t[0]||"top"} ${t[1]||"left"}`,this.panel.setPosition(this.position)}show(){C.add(this)}close(e){C.remove(this,e)}_open(){this.panel.hoistToBody(this.parentNode),this.panel.setPosition(this.position),this.panel.autoPosition(),this.panel.open(),this.panel.addEventListener("MDWPanel:closed",this.bound_onPanelClose),this.autoCancelTimeout=setTimeout(()=>{this.close()},3e3)}_close(e){this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose),this.panel.close(),this.dispatchClose(e),clearTimeout(this.autoCancelTimeout),setTimeout(()=>{this.panel.remove()},200)}onPanelClose(){this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose)}dispatchClose(e=!1){this.dispatchEvent(new CustomEvent("close",{detail:{ok:e}}))}}),customElements.define("mdw-switch",class extends g{constructor(){super(),this.bound_onInputChange=this.onInputChange.bind(this),this.cloneTemplate()}connectedCallback(){this.input.addEventListener("change",this.bound_onInputChange),this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:[this.input],radius:20,centered:!0})}disconnectedCallback(){this.input.addEventListener("click",this.bound_click),this.ripple.destroy()}static get observedAttributes(){return["checked","disabled"]}attributeChangedCallback(e,t,n){this[e]=n}get input(){return this.input_||(this.input_=this.shadowRoot.querySelector("input")),this.input_}get checked(){return this.input.checked}set checked(e){""===e&&(e=!0),this.input.checked=e,this.updateCheckedClass()}set disabled(e){(e=!!e||""===e)?this.input.setAttribute("disabled","disabled"):this.input.removeAttribute("disabled")}updateCheckedClass(){this.checked?this.classList.add("checked"):this.classList.remove("checked")}dispatchChange(){this.dispatchEvent(new Event("change"))}onInputChange(e){this.updateCheckedClass(),this.dispatchChange()}template(){return'\n      <div class="mdw-track"></div>\n      <div class="mdw-thumb-underlay">\n        <div class="mdw-thumb">\n          <input type="checkbox" role="switch">\n          <div class="mdw-ripple mdw-switch-ripple"></div>\n        </div>\n      </div>\n    '}styles(){return'\n      .mdw-track {\n        box-sizing: border-box;\n        width: 32px;\n        height: 14px;\n        border: 1px solid;\n        border-radius: 7px;\n        opacity: .38;\n        transition: opacity 90ms cubic-bezier(.4,0,.2,1),\n                    background-color 90ms cubic-bezier(.4,0,.2,1),\n                    border-color 90ms cubic-bezier(.4,0,.2,1);\n      }\n\n      :host(:not(.checked)) .mdw-track {\n        background-color: rgba(var(--mdw-theme-on-background--rgb), 0.7);\n        border-color: rgba(var(--mdw-theme-on-background--rgb), 0.7);\n      }\n\n      :host(.checked) .mdw-track {\n        background-color: var(--mdw-theme-secondary);\n        border-color: var(--mdw-theme-secondary);\n        opacity: .54;\n      }\n\n      :host(.checked.mdw-primary) .mdw-track {\n        background-color: var(--mdw-theme-primary);\n        border-color: var(--mdw-theme-primary);\n      }\n\n      :host(.checked.mdw-error) .mdw-track {\n        background-color: var(--mdw-theme-error);\n        border-color: var(--mdw-theme-error);\n      }\n\n\n\n      /* --- thumb underlay --- */\n\n      .mdw-thumb-underlay {\n        -webkit-tap-highlight-color: rgba(0,0,0,0);\n        display: flex;\n        position: absolute;\n        will-change: transform,opacity;\n        left: -18px;\n        right: auto;\n        top: -17px;\n        align-items: center;\n        justify-content: center;\n        width: 48px;\n        height: 48px;\n        transform: translateX(0);\n        transition: transform 90ms cubic-bezier(.4,0,.2,1),\n                    background-color 90ms cubic-bezier(.4,0,.2,1),\n                    border-color 90ms cubic-bezier(.4,0,.2,1);\n      }\n\n      :host(.checked) .mdw-thumb-underlay {\n        transform: translateX(20px);\n      }\n\n      .mdw-thumb-underlay:after,\n      .mdw-thumb-underlay:before {\n        position: absolute;\n        border-radius: 50%;\n        opacity: 0;\n        pointer-events: none;\n        content: "";\n        background-color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-primary) .mdw-thumb-underlay:after,\n      :host(.mdw-primary) .mdw-thumb-underlay:before {\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-error) .mdw-thumb-underlay:after,\n      :host(.mdw-error) .mdw-thumb-underlay:before {\n        background-color: var(--mdw-theme-error);\n      }\n\n      .mdw-thumb-underlay:before {\n        transition: opacity 15ms linear,background-color 15ms linear;\n        z-index: 1;\n      }\n\n\n\n      /* --- thumb --- */\n\n      .mdw-thumb {\n        box-shadow: 0 3px 1px -2px rgba(0,0,0,.2),\n                    0 2px 2px 0 rgba(0,0,0,.14),\n                    0 1px 5px 0 rgba(0,0,0,.12);\n        box-sizing: border-box;\n        width: 20px;\n        height: 20px;\n        border: 10px solid;\n        border-radius: 50%;\n        pointer-events: none;\n        z-index: 1;\n      }\n\n      :host(.checked) .mdw-thumb {\n        background-color: var(--mdw-theme-secondary);\n        border-color: var(--mdw-theme-secondary);\n      }\n\n      :host(.checked.mdw-primary) .mdw-thumb {\n        background-color: var(--mdw-theme-primary);\n        border-color: var(--mdw-theme-primary);\n      }\n\n      :host(.checked.mdw-error) .mdw-thumb {\n        background-color: var(--mdw-theme-error);\n        border-color: var(--mdw-theme-error);\n      }\n\n      :host(:not(.checked)) .mdw-thumb {\n        background-color: #fff;\n        border-color: #fff;\n      }\n\n\n      /* --- input --- */\n\n      input {\n        left: 0;\n        right: auto;\n        position: absolute;\n        top: 0;\n        width: 68px;\n        height: 48px;\n        margin: 0;\n        opacity: 0;\n        cursor: pointer;\n        pointer-events: auto;\n      }\n\n      :host(.checked) input {\n        transform: translateX(-20px);\n      }\n\n\n\n      /* --- ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.16);\n      }\n\n      :host(.mdw-primary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n\n      :host(.mdw-error) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.16);\n      }\n\n      .mdw-switch-ripple {\n        position: absolute;\n        left: 0;\n        top: 0;\n        right: 0;\n        bottom: 0;\n        border-radius: 50%;\n        z-index: 1;\n        pointer-events: none;\n      }\n    '}}),customElements.define("mdw-tab-body",class extends g{constructor(){super(),this.cloneTemplate()}connectedCallback(){this.parentNode.registerBody(this)}disconnectedCallback(){this.parentNode.unregisterBody(this)}addSlot(){this.shadowRoot.querySelector("mdw-tab-body-content").insertAdjacentHTML("beforeend","<slot></slot>")}removeSlot(){this.shadowRoot.querySelector("slot").remove()}activate(){this.addSlot(),this.classList.add("mdw-active")}deactivate(){this.removeSlot(),this.classList.remove("mdw-active")}template(){return"\n      <mdw-tab-body-content>\n        \x3c!-- slot is added dynamicly --\x3e\n      </mdw-tab-body-content>\n    "}styles(){return"\n      mdw-tab-body-content {\n        height: 100%;\n        overflow: auto;\n      }\n    "}}),customElements.define("mdw-tab-button",class extends g{constructor(){super(),this.bound_click=this.click.bind(this),this.cloneTemplate()}connectedCallback(){this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:this}),this.parentNode.registerTab(this),this.addEventListener("click",this.bound_click)}disconnectedCallback(){this.ripple.destroy(),this.parentNode.unregisterTab(this),this.removeEventListener("click",this.bound_click)}get indicator(){return this.shadowRoot.querySelector(".mdw-tab-button-indicator__content")}click(e){this.parentNode.tabClick(this)}activate(){clearTimeout(this._animationTimer),this.indicator.style.transform="",this._runNextAnimationFrame(()=>{this._animationTimer=setTimeout(()=>{this.classList.add("mdw-active")},180)})}deactivate(e){clearTimeout(this._animationTimer),this.indicator.style.transform=`translateX(${e.toString()}px)`,this._animationTimer=setTimeout(()=>{this.classList.remove("mdw-active")},200)}_runNextAnimationFrame(e){cancelAnimationFrame(this._animationFrame),this._animationFrame=requestAnimationFrame(()=>{this._animationFrame=0,clearTimeout(this._animationTimer),this._animationTimer=setTimeout(e,0)})}template(){return'\n      <span class="text"><slot></slot></span>\n      <span class="mdw-tab-button-indicator">\n        <span class="mdw-tab-button-indicator__content mdw-tab-button-indicator__content--underline"></span>\n      </span>\n      <div class="mdw-ripple mdw-tab-button-ripple"></div>\n    '}styles(){return"\n      :host(.mdw-show-spinner) span.text {\n        opacity: 0;\n      }\n\n      /* Add this to button or creat a new componenet mdw-tab */\n      .mdw-tab-button-indicator {\n        display: flex;\n        position: absolute;\n        top: 0;\n        left: 0;\n        width: 100%;\n        height: 100%;\n        pointer-events: none;\n        z-index: 1;\n      }\n\n      :host(.mdw-active) .mdw-tab-button-indicator .mdw-tab-button-indicator__content {\n        transition: transform .2s cubic-bezier(.4,0,.2,1);\n      }\n\n      .mdw-tab-button-indicator__content {\n        opacity: 0;\n        transform-origin: left;\n      }\n\n      :host(.mdw-active) .mdw-tab-button-indicator__content {\n        opacity: 1;\n        transform: translateX(0);\n      }\n\n      .mdw-tab-button-indicator .mdw-tab-button-indicator__content--underline {\n        align-self: flex-end;\n        width: 100%;\n        background-color: var(--mdw-theme-primary);\n        height: 2px;\n      }\n\n\n      /* --- Ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n      }\n\n      .mdw-tab-button-ripple,\n      .mdw-tab-button-focus-overlay {\n        border-radius: inherit;\n        position: absolute;\n        top: 0;\n        left: 0;\n        bottom: 0;\n        right: 0;\n        pointer-events: none;\n      }\n\n      :host(.mdw-active) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n    "}}),customElements.define("mdw-tabs-bar",class extends g{constructor(){super(),this._activeTab=0,this.tabIdCounter=0,this._contentElements=[],this.cloneTemplate()}registerTab(e){e.setAttribute("tab-id",this.tabIdCounter),0===this.tabIdCounter&&(this.activeTab=e,e.activate()),this.tabIdCounter++}unregisterTab(e){}registerContent(e){this._contentElements.push(e),e.changeTab(this.activeTab.getAttribute("tab-id"))}unregisterContent(e){this._contentElements=this._contentElements.filter(t=>t!=e)}tabClick(e){const t=parseInt(e.getBoundingClientRect().x-this.activeTab.getBoundingClientRect().x);this.activeTab.deactivate(t),this.activeTab=e,this.activeTab.activate(),this._contentElements.forEach(e=>e.changeTab(this.activeTab.getAttribute("tab-id")))}get activeTab(){return this._activeTab}set activeTab(e){this._activeTab=e}get internalStylesFile(){return"./internal.css"}template(){return"\n      <mdw-tabs-bar-scroller>\n        <mdw-tabs-bar-scroller-area>\n          <mdw-tabs-bar-scroller-content>\n            <slot></slot>\n          </mdw-tabs-bar-scroller-content>\n        </mdw-tabs-bar-scroller-area>\n      </mdw-tabs-bar-scroller>\n    "}styles(){return"\n      mdw-tabs-bar-scroller {\n        display: block;\n        overflow-y: hidden;\n      }\n\n      mdw-tabs-bar-scroller-area {\n        display: flex;\n        /* overflow-x: scroll; */\n      }\n\n      mdw-tabs-bar-scroller-content {\n        position: relative;\n        display: flex;\n        flex: 1 0 auto;\n        transform: none;\n        will-change: transform;\n      }\n\n      ::slotted(mdw-button) {\n        flex: 1 0 auto;\n      }\n    "}}),customElements.define("mdw-tabs-content",class extends g{constructor(){super(),this._bodies=[]}connectedCallback(){this.tabsBar.registerContent(this)}disconnectedCallback(){this.tabsBar&&this.tabsBar.unregisterContent(this)}get tabsBar(){return document.body.querySelector(`mdw-tabs-bar#${this.getAttribute("tabs-id")}`)}registerBody(e){this._bodies.push(e),void 0!==this._wiatForBodyActiveId&&this._bodies.length===this._wiatForBodyActiveId+1&&(this._activeBody=e,e.activate(),this._wiatForBodyActiveId=void 0)}unregisterBody(e){this._bodies=this._bodies.filter(t=>t!=e)}changeTab(e){this._bodies.length?(this._activeBody&&this._activeBody.deactivate(),this._activeBody=this._bodies[e],this._activeBody.activate()):this._wiatForBodyActiveId=parseInt(e)}}),customElements.define("mdw-textfield",class extends g{constructor(){super(),this.classList.add("mdw-no-animation"),this.bound_onFocus=this.onFocus.bind(this),this.bound_onBlur=this.onBlur.bind(this),this.bound_onInput=this.onInput.bind(this)}connectedCallback(){this.compose(),this.checkForValue(),requestAnimationFrame(()=>{this.classList.remove("mdw-no-animation")}),this.input.addEventListener("focus",this.bound_onFocus),this.input.addEventListener("blur",this.bound_onBlur),this.input.addEventListener("input",this.bound_onInput),this.classList.toggle("mdw-invalid",!this.valid)}disconnectedCallback(){this.input.removeEventListener("focus",this.bound_onFocus),this.input.removeEventListener("blur",this.bound_onBlur),this.input.removeEventListener("input",this.bound_onInput)}compose(){this.classList.add("mdw-upgraded"),this.isTextarea()&&this.classList.add("mdw-textarea"),this.outlined&&(this.insertAdjacentHTML("beforeend",this.outlinedHTML),this.setNotchWidth()),this.querySelector(".mdw-line-ripple")||this.insertAdjacentHTML("beforeend",this.lineRippleHTML),this.isTrailingIcon()&&this.classList.add("mdw-trailing-icon")}checkForValue(){this.classList.toggle("not-empty",!!this.input.value.length)}onFocus(){this.setNotchWidth()}onBlur(){this.classList.toggle("not-empty",!!this.input.value.length),this.classList.toggle("mdw-invalid",!this.valid)}onInput(){this.classList.toggle("mdw-invalid",!this.valid)}setNotchWidth(){this.outlined&&(this.notch.style.width=this.labelWidth+"px")}isTrailingIcon(){return!!this.iconElement&&[...this.children].indexOf(this.iconElement)>1}isTextarea(){return!!this.querySelector("textarea")}clear(){this.input.value=""}get valid(){return this.input.validity.valid}get outlined(){return[].slice.apply(this.classList||[]).includes("mdw-outlined")}get input(){return this.inputType_||(this.inputType_=this.querySelector("input")?"input":"textarea"),this.querySelector(this.inputType_)}get notch(){return this.querySelector(".mdw-outlined-notch")}get label(){return this.querySelector("label")}get labelWidth(){return.95*this.label.offsetWidth}get helperTextElement(){return this.querySelector("mdw-textfield-helper")}get iconElement(){return this.querySelector("mdw-icon")}get outlinedHTML(){return'\n      <div class="mdw-outlined-border-container">\n        <div class="mdw-outlined-leading"></div>\n        <div class="mdw-outlined-notch"></div>\n        <div class="mdw-outlined-trailing"></div>\n      </div>\n    '}get lineRippleHTML(){return'<div class="mdw-line-ripple"></div>'}}),customElements.define("mdw-top-app-bar",class extends g{constructor(){super(),this._throttledScrollHandler=a.rafThrottle(this._scrollHandler),this._throttledResizeHandler=a.rafThrottle(this._resizeHandler),this.bound_throttledScrollHandler=this._throttledScrollHandler.bind(this),this.bound_throttledResizeHandler=this._throttledResizeHandler.bind(this),document.body.classList.add("mdw-has-top-app-bar")}connectedCallback(){this._isProminent=this.hasAttribute("mdw-prominent"),this._isFixed=this.hasAttribute("mdw-fixed"),this._isShrink=this.hasAttribute("mdw-shrink"),this.parentNode&&"HEADER"===this.parentNode.nodeName&&(this.parentNode.classList.add("mdw-top-app-bar"),this._isProminent&&this.parentNode.classList.add("mdw-prominent"),this._isShrink&&this.parentNode.classList.add("mdw-shrink"),this._isFixed&&this.parentNode.classList.add("mdw-fixed")),this._isShrink&&(this._animationElements=[...this.querySelectorAll("[mdw-animation-property]")||[]].map(e=>{const t=parseFloat(e.getAttribute("mdw-animation-start")||0),n=parseFloat(e.getAttribute("mdw-animation-end")||0),i=e.getAttribute("mdw-animation-property").split(":");return{element:e,property:i[0],valueWrapper:i[1]||"",start:t,end:n,range:Math.abs(t-n)}}),this._scrollTarget=this._getScrollTarget(),this._lastScrollPosition=this._getViewportScrollY(),this._topAppBarHeight=this.clientHeight+6,this._scrollHandler(),this._createObserver(),this._scrollTarget.addEventListener("scroll",this.bound_throttledScrollHandler)),(this._isFixed||this._isShrink)&&(this._page=document.querySelector("mdw-page"),this._page&&(this.style.width=`${this._page.offsetWidth}px`)),window.addEventListener("resize",this.bound_throttledResizeHandler)}disconnectedCallback(){this._observer&&(this._observer.unobserve(this),this._observer.disconnect(),this._observer=void 0),this._scrollTarget&&this._scrollTarget.removeEventListener("scroll",this.bound_throttledScrollHandler),window.removeEventListener("resize",this.throttledResizeHandler)}notContextual(){this.removeAttribute("mdw-contextual")}contextual(){this.setAttribute("mdw-contextual","")}_getAllFixedSections(){return this.querySelectorAll("section[mdw-fixed]")}_getScrollTarget(){if("HEADER"===this.parentNode.nodeName){const e=this.parentNode.parentNode.querySelector("header + mdw-content:not([mdw-no-scroll])");if(e)return e}const e=document.querySelector("mdw-page-content");if(e&&e.contains(this))return e;const t=document.querySelector("mdw-content");return t&&t.contains(this)?t:window}_getViewportScrollY(){return this._scrollTarget[this._scrollTarget===window?"pageYOffset":"scrollTop"]}_scrollHandler(){const e=Math.max(this._getViewportScrollY(),0);let t;if(this._isProminent){const n=this._topAppBarHeight/2;t=e<=n?e:n}else t=e<=this._topAppBarHeight?e:this._topAppBarHeight;this.style.transform=`translateY(-${t}px)`,this._getAllFixedSections().forEach(e=>{e.style.transform=`translateY(${t}px)`})}_resizeHandler(){(this._isFixed||this._isShrink)&&(this._page=document.querySelector("mdw-page"),this._page&&(this.style.width=`${this._page.offsetWidth}px`))}_createObserver(){0!==this._animationElements.length&&(this._observer=new IntersectionObserver(this._handleIntersect.bind(this),{root:null,rootMargin:"0px",threshold:this._buildThresholdList()}),this._observer.observe(this))}_buildThresholdList(){let e=[];for(let t=1;t<=68;t++){let n=t/68;e.push(n)}return this._intersectionThresholds=e,this._intersectionThresholds}_handleIntersect(e,t){e.forEach(e=>{let t;this._isProminent&&(t=-e.boundingClientRect.top/(this._topAppBarHeight/2)),this._animationElements.forEach(e=>this._animationValue(t,e))})}_animationValue(e,{element:t,property:n,valueWrapper:i,start:s,end:o,range:r}){t.style[n]=i.replace("#",s-e*r)}}),customElements.define("mdw-bound-property",class extends g{constructor(){super(),this._property=this.innerHTML,this.innerHTML=""}connectedCallback(){const e=this;this._value=window.activePage[this._property],Object.defineProperty(window.activePage,this._property,{configurable:!0,enumerable:!0,get:()=>e._value,set(t){e._value=t,e.innerHTML=t}})}});const T=new class{constructor(){this._templates={},this._loadedTemplates={}}registerOnce(e,t){this._templates[e]||this.register(e,t)}async registerAndLoad(e,t){this._templates[e]||(this.register(e,t),t.includes(".html")&&await this.loadHtml(t))}register(e,t){if(!e)throw Error("requires id");if(!t)throw Error("requires templateString");if(this._templates[e])throw Error(`id "${e}" already taken`);this._templates[e]=t}unregister(e){this._templates[e]=void 0}isString(e){return"string"==typeof e&&(!e.includes(".html")&&e.includes("<"))}async get(e,t){if(!this._templates[e])throw Error(`no template found with id: ${e}`);const n=this._templates[e];return"function"==typeof n?n(t):n.includes(".html")?await this.loadHtml(n):n}async loadHtml(e){if(this._loadedTemplates[e])return this._loadedTemplates[e];const t=await fetch(e),n=await t.text();return this._loadedTemplates[e]=n,n}};window.MDWTemplate=T;var S=T;customElements.define("mdw-template",class extends g{constructor(){super();const e=this.templateId;e&&S.get(e).then(e=>{this.insertAdjacentHTML("beforeend",e)});const t=this.templateUrl;t&&S.loadHtml(t).then(e=>{this.insertAdjacentHTML("beforeend",e)})}get templateId(){return this.getAttribute("template-id")}get templateUrl(){return this.getAttribute("template-url")}async show(e){const t=await S.get(e);this.innerHTML=t}});const A=["panel","sheetBottom","sheetSide"],P=["scale","height"],M=["top","center"];class R{constructor({id:e,component:t,template:n,animation:i}){this._id=e,this._component=t,this._template=n,this._animation=i}get id(){return this._id}get component(){return this._component}get element(){return this._element}async open(){const e=document.querySelector("mdw-page");switch(e?e.insertAdjacentHTML("afterEnd",this._template):document.body.insertAdjacentHTML("beforeend",this._template),this._element=document.querySelector(`#${this.id}`),this.component){case"panel":this._element.setAnimation(this._animation),this.element.open();break;case"sheetBottom":case"sheetSide":this.element.open()}window._activeSurface&&await window._activeSurface.close(),window._activeSurface=this}async close(){switch(this.component){case"panel":case"sheetBottom":case"sheetSide":await this.element.close()}this.element.remove(),window._activeSurface=void 0}}const D=new class{constructor(){this._defaultMobileComponent="panel",this._defaultDesktopComponent="sheetSide"}setDefaultMobileComponent(e){this._validateComponent(e),this._defaultMobileComponent=e}setDefaultDesktopComponent(e){this._validateComponent(e),this._defaultDesktopComponent=e}async open({template:e,templateData:t,animation:n,animationTarget:i,component:s,mobileComponent:o,desktopComponent:r}){(await this.create({template:e,templateData:t,animation:n,animationTarget:i,component:s,mobileComponent:o,desktopComponent:r})).open()}async create({template:e,templateData:t,animation:n,animationTarget:i,component:s,mobileComponent:o,desktopComponent:r}){s||(s=this._autoSelectComponent(o,r)),this._validateComponent(s),"panel"===s&&(n||(n=this._autoSelectAnimation(i)),this._validateAnimation(n));const d=a.uid("surface"),l=S.isString(e)?e:await S.get(e,t);let c;switch(s){case"panel":c=this._buildPanel({id:d,animation:n,templateString:l});break;case"sheetBottom":c=this._buildSheetBottom({id:d,templateString:l});break;case"sheetSide":c=this._buildSheetSide({id:d,templateString:l})}return new R({id:d,component:s,template:c,animation:n})}close(){window._activeSurface&&window._activeSurface.close()}_autoSelectComponent(e,t){return a.isMobile?e||this._defaultMobileComponent:t||this._defaultDesktopComponent}_validateComponent(e){if(!A.includes(e))throw Error(`type must be one of these: ${A.join(", ")}`)}_autoSelectAnimation(e){return{type:"height",origin:"center",fullscreen:!0,target:e}}_validateAnimation(e){if(!P.includes(e.type))throw Error(`animation.type must be one of these: ${P.join(", ")}`);if(e.origin&&!M.includes(e.origin))throw Error(`animation.type must be one of these: ${M.join(", ")} or not defined`)}_buildPanel({id:e,templateString:t}){return`\n      <mdw-panel id="${e}">\n        ${t}\n      </mdw-panel>\n    `}_buildSheetBottom({id:e,templateString:t}){return`\n      <mdw-sheet-bottom id="${e}">\n        ${t}\n      </mdw-sheet-bottom>\n    `}_buildSheetSide({id:e,templateString:t}){return`\n      <mdw-sheet-side id="${e}" class="mdw-hide" mdw-modal mdw-no-backdrop>\n        ${t}\n      </mdw-sheet-side>\n    `}};window.MDWSurface=D;var H=D}]);
+!function(e){var t={};function n(i){if(t[i])return t[i].exports;var s=t[i]={i:i,l:!1,exports:{}};return e[i].call(s.exports,s,s.exports,n),s.l=!0,s.exports}n.m=e,n.c=t,n.d=function(e,t,i){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:i})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var i=Object.create(null);if(n.r(i),Object.defineProperty(i,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var s in e)n.d(i,s,function(t){return e[t]}.bind(null,s));return i},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=0)}([function(e,t,n){"use strict";n.r(t),n.d(t,"MDWDialog",(function(){return k})),n.d(t,"MDWSnackbar",(function(){return C})),n.d(t,"MDWTemplate",(function(){return S})),n.d(t,"MDWSurface",(function(){return H})),n.d(t,"MDWDrag",(function(){})),n.d(t,"MDWSwipe",(function(){}));const i=function(){if("undefined"!=typeof FORCE_MOBILE&&"true"===FORCE_MOBILE)return!0;var e=!1;return function(t){(/(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows ce|xda|xiino/i.test(t)||/1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i.test(t.substr(0,4)))&&(e=!0)}(navigator.userAgent||navigator.vendor||window.opera),e}(),o=function(){if("undefined"!=typeof FORCE_MOBILE&&"true"===FORCE_MOBILE)return!0;var e=!1;return function(t){(/(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows ce|xda|xiino|android|ipad|playbook|silk/i.test(t)||/1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i.test(t.substr(0,4)))&&(e=!0)}(navigator.userAgent||navigator.vendor||window.opera),e}();const r=new class{constructor(){this._uid=1,this._setupTransitionEvent(),this._setTransformPropertyName(),this.isPhone=i,this.isPhoneAndTablet=o,this.routeChangeCallbacks=new Set,this.onReady(()=>{this.isMobile?document.body.classList.add("mdw-is-mobile"):document.body.classList.remove("mdw-is-mobile"),window.addEventListener("hashchange",this.onRouteChange)})}onReady(e){document.body?e():setTimeout(()=>{this.onReady(e)},0)}uid(e="id"){return`${e}_${this._uid++}`}get isMobile(){return this.isPhoneAndTablet}lockPageScroll(){(document.body.classList.contains("prevent-over-scroll")?document.querySelector("mdw-page > mdw-content"):document.body).style.overflow="hidden"}unlockPageScroll(){(document.body.classList.contains("prevent-over-scroll")?document.querySelector("mdw-page > mdw-content"):document.body).style.overflow=""}disableUserSelect(){document.body.classList.add("mdw-no-select")}enableUserSelect(){document.body.classList.remove("mdw-no-select")}debounce(e,t){let n;return function(){const i=arguments,s=this;clearTimeout(n),n=setTimeout(()=>{n=void 0,e.apply(s,i)},t||10)}}throttle(e,t){let n;return function(){const i=arguments,s=this;n||(n=!0,e.apply(s,i),setTimeout(()=>{n=!1},t))}}rafThrottle(e){let t;return function(){const n=arguments,i=this;t||(t=!0,e.apply(i,n),requestAnimationFrame(()=>{t=!1}))}}querySlotted(e,t){if(!e)throw Error("requires either component");if(!t)throw Error("requires selector");return e.shadowRoot.querySelector("slot")?e.shadowRoot.querySelector("slot").assignedNodes().find(e=>!!e.matches&&e.matches(t)):null}querySlottedAll(e,t){if(!e)throw Error("requires either component");if(!t)throw Error("requires selector");return e.shadowRoot.querySelector("slot").assignedNodes({flatten:!0}).reduce((e,n)=>n.querySelectorAll?e.concat([...n.querySelectorAll(t)]):e,[])}slottedChildren(e){if(!e)throw Error("requires either component");return e.shadowRoot.querySelector("slot").assignedNodes()}get transitionEventName(){return this.transitionEventName_}get transformPropertyName(){return this.transformPropertyName_}addBackdrop(e,t,n={sheet:!1}){const i=this.uid();e.insertAdjacentHTML("afterend",`<div id="${i}" class="mdw-backdrop"></div>`);const s=document.querySelector(`#${i}`);return!0===n.sheet&&s.classList.add("mdw-sheet-backdrop"),t&&s.addEventListener("click",t),{remove(){t&&s.removeEventListener("click",t),s.remove()}}}get searchParamters(){return this._extractSearchParameters(this._clean(window.location.href)).split(",").filter(e=>!!e).reduce((e,t)=>{const n=t.split("=");return e[n[0]]=n[1],e},{})}setSearchParamter(e,t){const n=this.searchParamters;null==t?delete n[e]:n[e]=t;let i=window.location.href.split("?")[0];Object.keys(n).length>0&&(i+="?"+Object.keys(n).map(e=>`${e}=${n[e]}`).join(",")),window.history.pushState({path:i},"",i)}removeSearchParamter(e){this.setSearchParamter(e,void 0)}addOnRouteChange(e){return this.routeChangeCallbacks.add(e),()=>{this.routeChangeCallbacks.delete(e)}}removeOnRouteChange(e){this.routeChangeCallbacks.has(e)&&this.routeChangeCallbacks.delete(e)}_routeChange({oldUrl:e,newURL:t}={oldUrl:void 0,newURL:void 0}){this.routeChangeCallbacks.forEach(n=>n({oldUrl:e,newURL:t}))}_clean(e){return e instanceof RegExp?s:e.replace(/\/+$/,"").replace(/^\/+/,"/")}_extractSearchParameters(e){return e.split(/\?(.*)?$/).slice(1).join("")}_setupTransitionEvent(){const e=document.createElement("fakeelement"),t={transition:"transitionend",OTransition:"oTransitionEnd",MozTransition:"transitionend",WebkitTransition:"webkitTransitionEnd"};for(let n in t)void 0!==e.style[n]&&(this.transitionEventName_=t[n])}_setTransformPropertyName(e=!1){if(void 0===this.transformPropertyName_||e){const e=document.createElement("div");this.transformPropertyName_="transform"in e.style?"transform":"webkitTransform"}}};window.MDWUtils=r;var a=r;const d=new Map,l={start:"start",move:"move",end:"end"};function c(e,t){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");if("function"!=typeof t)throw Error("callback must be a function");const n=new m(e,t);n.addEvents(),d.get(e)||d.set(e,new Map),d.get(e).set(t,n)}function h(e,t){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");const n=d.get(e);if(n)if(t){const e=n.get(t);e&&e.removeEvents(),n.delete(t)}else n.forEach(e=>e.removeEvents()),d.delete(e)}class m{constructor(e,t){this.element=e,this.callback=t,this.hasPointerEvent=!!window.PointerEvent,this.bound_handleGestureStart=this.handleGestureStart.bind(this),this.bound_handleGestureMove=this.handleGestureMove.bind(this),this.bound_handleGestureEnd=this.handleGestureEnd.bind(this),this.callbackThrottle=a.rafThrottle(e=>{e.distance=this.getDistance(e),e.direction=this.getDirection(this.lastDistance,e.distance),this.lastDistance=e.distance,this.callback(e)})}addEvents(){this.disableTouchEvents(),a.isMobile?this.element.addEventListener("touchstart",this.bound_handleGestureStart,!1):this.element.addEventListener("mousedown",this.bound_handleGestureStart)}disable(){this.removeEvents()}enable(){this.addEvents()}enableTouchEvents(){this.element.style["touch-action"]=""}disableTouchEvents(){this.element.style["touch-action"]="none"}removeEvents(){this.enableTouchEvents(),a.isMobile?(this.element.removeEventListener("touchstart",this.bound_handleGestureStart),this.element.removeEventListener("touchmove",this.bound_handleGestureMove),this.element.removeEventListener("touchend",this.bound_handleGestureEnd),this.element.removeEventListener("touchcancel",this.bound_handleGestureEnd)):(this.element.removeEventListener("mousedown",this.bound_handleGestureStart),window.removeEventListener("mousemove",this.bound_handleGestureMove),window.removeEventListener("mouseup",this.bound_handleGestureEnd))}handleGestureStart(e){"mousedown"===e.type&&0!==e.button||(e.state="start",a.isMobile?(this.element.addEventListener("touchmove",this.bound_handleGestureMove,!1),this.element.addEventListener("touchend",this.bound_handleGestureEnd,!1),this.element.addEventListener("touchcancel",this.bound_handleGestureEnd,!1)):(window.addEventListener("mousemove",this.bound_handleGestureMove),window.addEventListener("mouseup",this.bound_handleGestureEnd)),this.startTime=Date.now(),this.initialTouchPos=this.getClientXY(e),this.lastDistance=this.getDistance(e),this.moved=!1)}handleGestureMove(e){e.state=this.moved?"move":"start",this.moved=!0,this.initialTouchPos&&this.callbackThrottle(e)}handleGestureEnd(e){if(e.state="end",a.isMobile?(this.element.removeEventListener("touchmove",this.bound_handleGestureMove),this.element.removeEventListener("touchend",this.bound_handleGestureEnd),this.element.removeEventListener("touchcancel",this.bound_handleGestureEnd)):(window.removeEventListener("mousemove",this.bound_handleGestureMove),window.removeEventListener("mouseup",this.bound_handleGestureEnd)),!1!==this.moved){if(this.endTime=Date.now(),e.runTime=this.endTime-this.startTime,e.distance=this.getDistance(e),e.direction=this.getDirection({x:0,y:0},e.distance),e.velocity=this.getVelocity(e.distance,e.runTime),void 0===e.clientX){const t=this.getClientXY(e);e.clientX=t.x,e.clientY=t.y}this.callback(e)}}getDistance(e){const t=this.getClientXY(e);return{x:t.x-this.initialTouchPos.x,y:t.y-this.initialTouchPos.y}}getDirection(e,t){const n=t.x>e.x?1:t.x===e.x?0:-1,i=t.y>e.y?1:t.y===e.y?0:-1;return{x:n,y:i,xDescription:0===n?"none":1===n?"right":"left",yDescription:0===i?"none":1===i?"down":"up"}}getClientXY(e){return{x:e.targetTouches&&e.targetTouches.length?e.targetTouches[0].clientX:e.changedTouches&&e.changedTouches.length?e.changedTouches[0].clientX:e.clientX,y:e.targetTouches&&e.targetTouches.length?e.targetTouches[0].clientY:e.changedTouches&&e.changedTouches.length?e.changedTouches[0].clientY:e.clientY}}getVelocity(e,t){return{x:e.x/t,y:e.y/t}}}window.MDWDrag={states:l,addDragListener:c,removeDragListener:h,enableDragListenerForElement:function(e){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");const t=d.get(e);t&&t.forEach(e=>e.enable())},disableDragListenerForElement:function(e){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");const t=d.get(e);t&&t.forEach(e=>e.disable())}};const u=new Map;class p{constructor(e,t){this.bound_dragEvent=this.dragEvent.bind(this),this.element=e,this.callback=t}addEvents(){c(this.element,this.bound_dragEvent)}removeEvents(){h(this.element,this.bound_dragEvent)}dragEvent(e){"end"===e.state&&(Math.abs(e.velocity.x)>1.4||Math.abs(e.velocity.y)>1.4)&&this.callback(e)}}function w(e,t){const n=new p(e,t);n.addEvents(),u.get(e)||u.set(e,new Map),u.get(e).set(t,n)}function b(e,t){if(!(e instanceof HTMLElement))throw Error("element must be an instance HTMLElement");const n=u.get(e);if(n)if(t){const e=n.get(t);e&&e.removeEvents(),n.delete(t)}else n.forEach(e=>e.removeEvents()),u.delete(e)}window.MDWSwipe={addSwipeListener:w,removeSwipeListener:b};new class{constructor(){this.routes={},this.classReference={},this.intercepter=void 0,this.pageClassnameRegex=/class\s(.*)\sextends/,this.bound_resolve=this._resolve.bind(this),this.PARAMETER_REGEXP=/([:*])(\w+)/g,this.WILDCARD_REGEXP=/\*/g,this.REPLACE_VARIABLE_REGEXP="([^/]+)",this.REPLACE_WILDCARD="(?:.*)",this.FOLLOWED_BY_SLASH_REGEXP="(?:/$|$)",this.MATCH_REGEXP_FLAGS="",this._transitionPages=!1,this.bound_onTransitionComplete=this._onTransitionComplete.bind(this),this.__mutationObserver=new MutationObserver(()=>{window.activePage.connectedCallback&&window.activePage.connectedCallback(),this._stopWatchingForConnect()})}init(){window.addEventListener("hashchange",this.bound_resolve),window.addEventListener("DOMContentLoaded",()=>{this._resolve(void 0,!0)})}interceptRouteChange(e){"function"==typeof e&&(this.intercepter=e)}get transitionPages(){return this._transitionPages}set transitionPages(e){this._transitionPages=!!e}get path(){let e=window.location.hash.replace(/.*#/,"");return e.indexOf("?")>-1&&(e=e.split("?")[0]),"/"!==e.charAt(0)&&(e="/"+e),e}get urlParameters(){const e=this._match(this.path);return e?e.params:{}}get searchParamters(){return this._extractSearchParameters(this._clean(window.location.href)).split(",").filter(e=>!!e).reduce((e,t)=>{const n=t.split("=");return e[n[0]]=n[1],e},{})}set hash(e){window.location=`#/${e}`}setSearchParamter(e,t){const n=this.searchParamters;null==t?delete n[e]:n[e]=t;let i=window.location.href.split("?")[0];Object.keys(n).length>0&&(i+="?"+Object.keys(n).map(e=>`${e}=${n[e]}`).join(",")),window.history.pushState({path:i},"",i)}removeSearchParamter(e){this.setSearchParamter(e,void 0)}addTransitionCSS(){this._addTransitionCSSAdded||(document.body.insertAdjacentHTML("beforebegin","<style>\n      page-container {\n        display: flex;\n      }\n      page-container.in-transition {\n        overflow-x: hidden;\n      }\n      page-render-block {\n        width: 100%;\n        flex-shrink: 0;\n        opacity: 1;\n      }\n      page-render-block.before-transition-page-out {\n        pointer-events: none;\n        user-select: none;\n      }\n      page-render-block.before-transition-page-in {\n        transform: scale(0.9) translateX(-100%);\n        opacity: 0;\n      }\n      page-render-block.transition-page-out {\n        transform: scale(0.9);\n        opacity: 0;\n        transition: opacity .16s linear,\n                    transform .26s cubic-bezier(0,0,.2,1);\n      }\n      page-render-block.transition-page-in {\n        transform: scale(1) translateX(-100%);\n        transform-origin: -50% 0;\n        opacity: 1;\n        transition: opacity .18s linear,\n                    transform .26s cubic-bezier(0,0,.2,1);\n      }\n    </style>"),this._addTransitionCSSAdded=!0)}addPageHideCSS(){this._addPageHideCSSAdded||(document.body.insertAdjacentHTML("beforebegin","<style>\n      .mdw-hide-non-page-container {\n        display: none;\n      }\n    </style>"),this._addPageHideCSSAdded=!0)}addPageClass(e,t){const n=this.getClassName(e,t);t&&this.addPageClassPath(e,t),(e.routes||[]).forEach(i=>{if(t!==i){if(this.routes[i])throw Error(`Path already exists: ${i}`);this.classReference[n]=e,this.routes[i]=n}})}addPageClassPath(e,t){const n=this.getClassName(e,t);if(this.routes[t])throw Error(`Path already exists: ${optionalPath}`);this.classReference[n]=e,this.routes[t]=n}getClassName(e,t){const n=this.pageClassnameRegex.exec(e);return n?n[1]:t.split("/").pop().replace(".js","")}setRoot(e){if(this.routes[e]){const t=this.classReference[this.routes[e]];this.addPageClassPath(t,"/")}else this.addPageClassPath(e,"/")}set404({Class:e}){e&&(this._notFoundRouteClass=e)}showPageOnly(){const e=document.querySelector("page-container"),t=document.documentElement;let n,i=e,s=e;for(;i.parentNode&&i.parentNode!==t;){for(i=i.parentNode,n=i.firstChild;n;)1===n.nodeType&&n!==s&&n.classList.add("mdw-hide-non-page-container"),n=n.nextSibling;s=i}this.addPageHideCSS(),this._isShowingPageOnly=!0}undoShowPageOnly(){!0===this._isShowingPageOnly&&([...document.querySelectorAll(".mdw-hide-non-page-container")||[]].forEach(e=>e.classList.remove("mdw-hide-non-page-container")),this._isShowingPageOnly=!1)}_resolve(e,t=!1){const{oldURL:n,newURL:i}=e||{};if(!1===t&&void 0!==n&&n===i)return;const s=this.intercepter?this.intercepter(i,n):void 0;if(s&&s.then&&"function"==typeof s.then&&console.error("you cannot return a Promise to the router.intercepter callback. Expecting either true or false"),!1===s)return void window.history.go();const o=this.path,r=this._match(o);if(!r)return this._notFoundRouteClass?this._changePage(this._notFoundRouteClass):console.warn("no page found and no default not found page setup");let a=o;if(t&&this._pageIsPreRendered())return;let d=this._extractSearchParameters(this._clean(window.location.href));if(d&&(a+=`?${d}`),void 0!==n){const e=n.length>i.length?n.replace(i,""):i.replace(n,"");if(""===e||"#/"===e)return}return this._changePage(r)}_watchForConnect(){const e=document.querySelector("page-render-block:not(.previous)");this.__mutationObserver.observe(e,{childList:!0})}_stopWatchingForConnect(){this.__mutationObserver.disconnect()}resetPageScroll(){const e=document.querySelector("mdw-page mdw-page-content");if(e&&e.scrollTop>0)return e.scrollTop=0;const t=document.querySelector("mdw-page mdw-content");if(t&&t.scrollTop>0)return t.scrollTop=0;const n=document.querySelector("mdw-page");if(n&&n.scrollTop>0)return n.scrollTop=0;const i=document.querySelector("body");if(i.scrollTop>0)return i.scrollTop=0;const s=document.documentElement;return s.scrollTop>0?s.scrollTop=0:void 0}_changePage({Class:e}){if(!e)throw Error("no class found");const t=document.querySelector("page-container");if(!t)throw Error("<page-container> required for router to work");this._stopWatchingForConnect(),this.undoShowPageOnly();const n=document.querySelector("page-render-block");if(!n)return t.appendChild(document.createElement("page-render-block")),window.activePage=new e,this._watchForConnect(),window.activePage.render(),void this.resetPageScroll();if(!this._transitionPages)return window.activePage.disconnectedCallback(),window.activePage=new e,this._watchForConnect(),window.activePage.render(),void this.resetPageScroll();n.classList.add("previous"),n.classList.add("before-transition-page-out"),window.activePage._disableRender=!0,window.activePage.disconnectedCallback();const i=document.createElement("page-render-block");i.classList.add("before-transition-page-in"),n.insertAdjacentElement("afterend",i);const s=new e;window.activePage=s,this._watchForConnect(),s.render();const o=document.querySelector("title");o&&(o.innerText=s.title),t.classList.add("in-transition"),n.classList.add("transition-page-out"),i.classList.add("transition-page-in"),n.addEventListener("transitionend",this.bound_onTransitionComplete),i.addEventListener("transitionend",this.bound_onTransitionComplete)}_onTransitionComplete({target:e}){e.removeEventListener("transitionend",this.bound_onTransitionComplete),e.classList.contains("transition-page-out")?e.remove():(e.classList.remove("before-transition-page-in"),e.classList.remove("transition-page-in")),document.querySelector("page-render-block.previous")||document.querySelector("page-render-block.next")||document.querySelector("page-container").classList.remove("in-transition")}_pageIsPreRendered(){const e=document.querySelector("page-render-block");return!!(e&&e.children.length>0)}_clean(e){return e instanceof RegExp?s:e.replace(/\/+$/,"").replace(/^\/+/,"/")}_extractSearchParameters(e){return e.split(/\?(.*)?$/).slice(1).join("")}_match(e){let t=this._findMatchedRoutes(e);return!!t.length&&(1===t.length?t[0]:t.sort((e,t)=>t.params?1:-1)[0])}_findMatchedRoutes(e){return Object.keys(this.routes).map(t=>{const n=this.routes[t],{regexp:i,paramNames:s}=this._replaceDynamicURLParts(this._clean(t)),o=e.replace(/^\/+/,"/").match(i),r=this._regExpResultToParams(o,s),a=this.classReference[n];return!!o&&{match:o,route:t,params:r,className:n,Class:a}}).filter(e=>e&&""!==e.match[0])}_replaceDynamicURLParts(e){let t=[],n="";return n=e instanceof RegExp?e:new RegExp(this._clean(e).replace(this.PARAMETER_REGEXP,(e,n,i)=>(t.push(i),this.REPLACE_VARIABLE_REGEXP)).replace(this.WILDCARD_REGEXP,this.REPLACE_WILDCARD)+this.FOLLOWED_BY_SLASH_REGEXP,this.MATCH_REGEXP_FLAGS),{regexp:n,paramNames:t}}_regExpResultToParams(e,t){return 0===t.length?null:e?e.slice(1,e.length).reduce((e,n,i)=>(null===e&&(e={}),e[t[i]]=decodeURIComponent(n),e),null):null}};class g extends HTMLElement{constructor(){super()}get _templateId(){return`${this.nodeName.toLowerCase()}--template`}cloneTemplate(e){let t=document.getElementById(this._templateId);t||(t=this._createTemplate());const n=t.content,i=this.shadowRoot?this.shadowRoot:this.attachShadow({mode:"open"}),s=n.cloneNode(!0);e&&(!this.__isBuildProcess&&this.beforeRender&&this.beforeRender(),s.querySelector("render-block").innerHTML=this.template()),i.appendChild(s),!this.__isBuildProcess&&this.afterRender&&this.afterRender()}_createTemplate(){const e=document.createElement("template");return e.setAttribute("id",this._templateId),e.innerHTML=`\n        <style>\n          ${this.styles()}\n        </style>\n        <render-block>\n          ${this.template()}\n        </render-block>\n    `,document.body.insertAdjacentElement("beforeend",e),e}connectedCallback(){!this.__isBuildProcess&&this.addEvents&&this.addEvents()}disconnectedCallback(){!this.__isBuildProcess&&this.removeEvents&&this.removeEvents()}render(){if(this.__isBuildProcess)return;const e=this.shadowRoot.querySelector("render-block");if(!e)throw Error("Could not find <render-block>");this.removeEvents&&this.removeEvents(),this.beforeRender&&this.beforeRender(),e.innerHTML=this.template(),this.afterRender&&this.afterRender(),this.addEvents&&this.addEvents()}beforeRender(){}afterRender(){}addEvents(){}removeEvents(){}styles(){}externalStyles(){}template(){}}const _=new class{constructor(){this.queue=[]}add(e,t){this.queue.push({el:e,resolver:t}),this.handleQueue()}remove(e){this.current&&this.current.el===e?(this.current.resolver(!1),e._dissmiss()):this.queue=this.queue.filter(t=>t.el!==e)}accept(e){this.current&&this.current.el===e?(this.current.resolver(!0),e._dissmiss()):this.queue=this.queue.filter(t=>t.el!==e)}handleQueue(){0!==this.queue.length&&(this.current||(this.current=this.queue.shift(),this.current.el._show(),this.current.el.addEventListener("close",()=>{this.current=void 0,setTimeout(()=>{this.handleQueue()},300)})))}create({message:e,dismissLabel:t="dismiss",acceptLabel:n=null,template:i,parent:s}){if(!e&&!i)throw Error("Either `message` or `template` is required");if(!i&&!t&&!n)throw Error("When not using a `template` you are required to provide either a `dismissLabel` or an `acceptLabel`");const o=a.uid();i||(i=this.template(e,t,n,o));let r=s||document.querySelector("mdw-page > mdw-top-app-bar");r||(r=document.querySelector("mdw-page")),r||(r=document.querySelector("body"));let d,l=void 0;"MDW-TOP-APP-BAR"===r.nodeName?(r.insertAdjacentHTML("afterend",i),l=document.querySelector(`mdw-banner#${o}`)):(r.insertAdjacentHTML("afterbegin",i),l=document.querySelector(`mdw-banner#${o}`));const c=new Promise(e=>{d=e});return this.add(l,d),c}template(e,t,n,i){return`\n      <mdw-banner id="${i}" class="mdw-elevation-1">\n        <div>${e}</div>\n        <div>\n          ${t?`<mdw-button onclick="${i}.dismiss()" class="mdw-secondary">${t}</mdw-button>`:""}\n          ${n?`<mdw-button onclick="${i}.accept()" class="mdw-secondary">${n}</mdw-button>`:""}\n        </div>\n      </mdw-banner>\n    `}};window.MDWBanner=_;var v=_;customElements.define("mdw-banner",class extends g{constructor(){super()}connectedCallback(){this.style.marginBottom=`-${this.clientHeight+1}px`}show(){v.add(this)}dismiss(){v.remove(this)}accept(){v.accept(this)}_show(){this.classList.add("mdw-show")}_dissmiss(){const e=this;e.addEventListener(a.transitionEventName,(function t(){e.removeEventListener(a.transitionEventName,t),e.remove()})),this.classList.add("mdw-dismiss"),this.dispatchClose()}dispatchClose(){this.dispatchEvent(new CustomEvent("close"))}}),customElements.define("mdw-bottom-navigation",class extends g{constructor(){super(),!o&&this.hasAttribute("mdw-mobile-only")?(this.style.display="none",this.style.pointerEvents="none"):(this.bound_routeChange=this.routeChange.bind(this),document.body.classList.add("mdw-has-bottom-navigation"))}connectedCallback(){window.addEventListener("hashchange",this.bound_routeChange),window.addEventListener("DOMContentLoaded",this.bound_routeChange)}disconnectedCallback(){window.removeEventListener("hashchange",this.bound_routeChange),window.removeEventListener("DOMContentLoaded",this.bound_routeChange)}get path(){let e=window.location.hash.replace(/.*#/,"");return e.indexOf("?")>-1&&(e=e.split("?")[0]),"/"!==e.charAt(0)&&(e="/"+e),e}routeChange(){this.querySelectorAll(".mdw-current-link").forEach(e=>e.classList.remove("mdw-current-link"));let e=this.querySelectorAll(`[href="#${this.path}"]`);e&&0!==e.length||(e=this.querySelectorAll(`[alt-href="#${this.path}"]`)),e.forEach(e=>e.classList.add("mdw-current-link"))}});class f{constructor(e={}){if(this.RIPPLE_FADE_IN_DURATION=200,this.RIPPLE_FADE_OUT_DURATION=150,this.RIPPLE_STATE={FADING_IN:"FADING_IN",VISIBLE:"VISIBLE",FADING_OUT:"FADING_OUT",HIDDEN:"HIDDEN"},!e.element)throw Error("requires config.element");if(!e.triggerElement)throw Error("requires config.triggerElement");this.element=e.element,this.triggerElement=[].concat(e.triggerElement).filter(e=>!!e),this.centered=!!e.centered,this.speedFactor=e.speedFactor||1,this.radius=e.radius,this.color=e.color||null,this.persistent=!!e.persistent,this.activeRipples=new Set,this.isMousedown=!1,this.bound_mousesdown_=this.mousesdown_.bind(this),this.bound_mouseup_=this.mouseup_.bind(this),this.bound_mouseleave_=this.mouseleave_.bind(this),this.triggerElement.forEach(e=>{e.addEventListener("mousedown",this.bound_mousesdown_)})}destroy(){this.triggerElement.forEach(e=>{e.removeEventListener("mousedown",this.bound_mousesdown_)})}mousesdown_(e){this.isMousedown=!0,this.triggerElement.forEach(e=>{e.addEventListener("mouseup",this.bound_mouseup_),e.addEventListener("mouseleave",this.bound_mouseleave_)}),this.fadeInRipple(e.pageX,e.pageY)}mouseup_(e){this.isMousedown=!1,this.activeRipples.forEach(e=>{e.config.persistent||e.state!==this.RIPPLE_STATE.VISIBLE||e.fadeOut()}),this.triggerElement.forEach(e=>{e.removeEventListener("mouseup",this.bound_mouseup_),e.removeEventListener("mouseleave",this.bound_mouseleave_)})}mouseleave_(){this.isMousedown&&this.mouseup_()}fadeInRipple(e,t){const n=this.element.getBoundingClientRect();if(this.centered)e=n.left+n.width/2,t=n.top+n.height/2;else{const n=this.getViewportScrollPosition();e-=n.left,t-=n.top}const i=this.radius||this.distanceToFurthestCorner(e,t,n),s=this.RIPPLE_FADE_IN_DURATION*(1/this.speedFactor),o=e-n.left,r=t-n.top,a=document.createElement("div");a.classList.add("mdw-ripple-element"),a.style.left=`${o-i}px`,a.style.top=`${r-i}px`,a.style.height=`${2*i}px`,a.style.width=`${2*i}px`,a.style.backgroundColor=this.color,a.style.transitionDuration=`${s}ms`,this.element.appendChild(a),this.enforceStyleRecalculation(a),a.style.transform="scale(1)";let d={config:{centered:this.centered,tiggerElement:this.triggerElement,speedFactor:this.speedFactor,radius:i,color:this.color,persistent:this.persistent,duration:s},element:a,fadeOut:()=>l(),state:this.RIPPLE_STATE.FADING_IN};const l=()=>{this.fadeOutRipple(d)};this.activeRipples.add(d),setTimeout(()=>{d.state=this.RIPPLE_STATE.VISIBLE,this.persistent||this.isMousedown||d.fadeOut()},s)}fadeOutRipple(e){if(!this.activeRipples.delete(e))return;const t=e.element;t.style.transitionDuration=`${this.RIPPLE_FADE_OUT_DURATION}ms`,t.style.opacity="0",e.state=this.RIPPLE_STATE.FADING_OUT,setTimeout(()=>{e.state=this.RIPPLE_STATE.HIDDEN,t.parentNode.removeChild(t)},this.RIPPLE_FADE_OUT_DURATION)}distanceToFurthestCorner(e,t,n){const i=Math.max(Math.abs(e-n.left),Math.abs(e-n.right)),s=Math.max(Math.abs(t-n.top),Math.abs(t-n.bottom));return Math.sqrt(i*i+s*s)}getViewportScrollPosition(){const e=document.documentElement.getBoundingClientRect();return{top:-e.top||document.body.scrollTop||window.scrollY||document.documentElement.scrollTop||0,left:-e.left||document.body.scrollLeft||window.scrollX||document.documentElement.scrollLeft||0}}enforceStyleRecalculation(e){window.getComputedStyle(e).getPropertyValue("opacity")}}customElements.define("mdw-button",class extends g{constructor(){super(),this.bound_asyncClick=this.asyncClick.bind(this),this.bound_hrefClick=this.hrefClick.bind(this),this.bound_checkHREFActive=this.checkHREFActive.bind(this),this.cloneTemplate(),this.setupAsync(),this.connectHREF()}connectedCallback(){this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:this})}disconnectedCallback(){this.ripple.destroy(),this.removeEventListener("click",this.bound_asyncClick),this.removeEventListener("click",this.bound_hrefClick),window.removeEventListener("hashchange",this.bound_checkHREFActive)}get spinnerContainer(){return this._spinnerContainer||(this._spinnerContainer=this.shadowRoot.querySelector(".mdw-spinner-container")),this._spinnerContainer}get pending(){return this._pending}setupAsync(){this.hasAttribute("mdw-async")&&this.addEventListener("click",this.bound_asyncClick)}resolve(){!1!==this._pending&&(this._pending=!1,this.hideSpinner())}asyncClick(e){!0!==this._pending&&(this._pending=!0,this.showSpinner())}showSpinner(){this._showSpinner=!0,this.classList.add("mdw-show-spinner");const e=this.classList.contains("mdw-primary")||this.classList.contains("mdw-secondary")||this.classList.contains("mdw-error");this.spinnerContainer.innerHTML=`<mdw-circular-progress mdw-mode="indeterminate" mdw-diameter="24" class="${e?"mdw-white":"mdw-grey"}" style="position: absolute; left: calc(50% - 12px); top: 6px;"></mdw-circular-progress>`}hideSpinner(){this._showSpinner=!1,this.classList.remove("mdw-show-spinner"),this.spinnerContainer.innerHTML=""}connectHREF(){this.hasAttribute("href")&&(this.checkHREFActive(),window.addEventListener("hashchange",this.bound_checkHREFActive),this.addEventListener("click",this.bound_hrefClick))}checkHREFActive(){if(!this.hasAttribute("href"))return;const e=document.location.href,t=document.location.hash;e===this.getAttribute("href")||e===this.getAttribute("href-alt")||t===this.getAttribute("href")||t===this.getAttribute("href-alt")?this.setAttribute("active","active"):this.removeAttribute("active")}hrefClick(){"_blank"!==this.getAttribute("target")?document.location.href=this.getAttribute("href"):window.open(this.getAttribute("href"),"_blank")}template(){return'\n      <span class="text"><slot></slot></span>\n      <span class="mdw-spinner-container"></span>\n      <div class="mdw-ripple mdw-button-ripple"></div>\n    '}styles(){return'\n      :host {\n        user-select: none;\n        -webkit-user-select: none;\n        align-items: center;\n        border: none;\n        box-sizing: border-box;\n        display: inline-flex;\n        font-size: 0.875rem;\n        font-weight: 500;\n        justify-content: center;\n        letter-spacing: 0.08929em;\n        outline: none;\n        overflow: hidden;\n        position: relative;\n        text-decoration: none;\n        text-transform: uppercase;\n        vertical-align: middle;\n        will-change: transform, opacity;\n        margin: 0;\n        background-color: transparent;\n        white-space: nowrap;\n\n        border-radius: 4px;\n        line-height: 2.25rem;\n        padding: 0 8px 0 8px;\n        height: 36px;\n        min-width: 64px;\n\n        color: var(--mdw-theme-on-primary);\n      }\n\n      :host-context(.mdw-density-comfortable),\n      :host(.mdw-density-comfortable) {\n        height: 28px;\n        margin-top: 0;\n        margin-bottom: 0;\n        padding: 0 8px 0 8px;\n      }\n\n      :host-context(.mdw-density-compact),\n      :host(.mdw-density-compact) {\n        height: 24px;\n        margin-top: 0;\n        margin-bottom: 0;\n        padding: 0 4px 0 4px;\n      }\n\n      :host(.mdw-icon):host-context(.mdw-density-comfortable),\n      :host(.mdw-density-comfortable.mdw-icon) {\n        height: 28px;\n        width: 28px;\n        margin-top: 0;\n        margin-bottom: 0;\n      }\n\n      :host(.mdw-icon):host-context(.mdw-density-compact),\n      :host(.mdw-density-compact.mdw-icon) {\n        height: 24px;\n        width: 24px;\n        margin-top: 0;\n        margin-bottom: 0;\n      }\n\n      :host-context(.mdw-density-comfortable) .mdw-spinner-container mdw-circular-progress,\n      :host(.mdw-density-comfortable) .mdw-spinner-container mdw-circular-progress {\n        top: 2px !important;\n      }\n\n      :host-context(.mdw-density-compact) .mdw-spinner-container mdw-circular-progress,\n      :host(.mdw-density-compact) .mdw-spinner-container mdw-circular-progress {\n        top: 0 !important;\n      }\n\n\n      :host::before,\n      :host::after {\n        position: absolute;\n        border-radius: 50%;\n        opacity: 0;\n        pointer-events: none;\n        content: "";\n        top: calc(50% - 100%);\n        left: calc(50% - 100%);\n        width: 200%;\n        height: 200%;\n        background-color: var(--mdw-theme-foreground);\n      }\n\n      :host::before {\n        z-index: 1;\n        transition: opacity 15ms linear,\n                    background-color 15ms linear;\n      }\n\n      :host(:hover) {\n        cursor: pointer;\n      }\n\n      :host([disabled]) {\n        background-color: transparent !important;\n        color: var(--mdw-theme-text-disabled-on-background);\n        cursor: default;\n        pointer-events: none;\n      }\n      :host::-moz-focus-inner {\n        padding: 0;\n        border: 0;\n      }\n\n      :host(:active) {\n        outline: none;\n      }\n\n      :host(.mdw-contained),\n      :host(.mdw-raised),\n      :host(.mdw-unelevated) {\n        background-color: var(--mdw-theme-background);\n        padding: 0 16px 0 16px;\n      }\n\n      :host(.mdw-contained)::before,\n      :host(.mdw-raised)::before,\n      :host(.mdw-unelevated)::before {\n        opacity: 0.08;\n      }\n\n      :host(.mdw-contained),\n      :host(.mdw-raised) {\n        box-shadow: 0px 3px 1px -2px rgba(0, 0, 0, 0.2), 0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 1px 5px 0px rgba(0, 0, 0, 0.12);\n        transition: box-shadow 280ms cubic-bezier(0.4, 0, 0.2, 1);\n      }\n\n      :host(.mdw-contained:hover),\n      :host(.mdw-contained:focus),\n      :host(.mdw-raised:hover),\n      :host(.mdw-raised:focus) {\n        box-shadow: 0px 2px 4px -1px rgba(0, 0, 0, 0.2), 0px 4px 5px 0px rgba(0, 0, 0, 0.14), 0px 1px 10px 0px rgba(0, 0, 0, 0.12);\n      }\n\n      :host(.mdw-contained:active),\n      :host(.mdw-raised:active) {\n        box-shadow: 0px 5px 5px -3px rgba(0, 0, 0, 0.2), 0px 8px 10px 1px rgba(0, 0, 0, 0.14), 0px 3px 14px 2px rgba(0, 0, 0, 0.12);\n      }\n\n      :host(.mdw-contained[disabled]),\n      :host(.mdw-raised[disabled]) {\n        box-shadow: 0px 0px 0px 0px rgba(0, 0, 0, 0.2), 0px 0px 0px 0px rgba(0, 0, 0, 0.14), 0px 0px 0px 0px rgba(0, 0, 0, 0.12);\n      }\n\n      :host(.mdw-outlined) {\n        border-color: rgba(0, 0, 0, 0.37);\n        line-height: inherit;\n        border-style: solid;\n        padding: 0 14px 0 14px;\n        border-width: 2px;\n      }\n\n      :host(.mdw-shaped) {\n        border-radius: var(--mdw-theme--button-shape-radius, 18px);\n      }\n\n      :host(.mdw-icon) {\n        border-radius: 50%;\n        min-width: 0;\n        width: 48px;\n        height: 48px;\n        padding: 12px;\n        line-height: 19px;\n      }\n      \n      :host(.mdw-icon) span.text {\n        display: flex;\n        flex-direction: column;\n        align-items: center;\n        font-size: 0.75rem;\n        text-transform: none;\n      }\n\n      :host(.mdw-bottom-navigation) {\n        border-radius: 50%;\n        min-width: 0;\n        max-width: 100px;\n        width: 56px;\n        height: 56px;\n        padding: 28px;\n      }\n\n      :host(:not(.mdw-icon)) .text ::slotted(mdw-icon) {\n        display: inline-block;\n        height: 28px;\n        vertical-align: middle;\n      }\n\n      /*\n      :host(.mdw-icon) ::slotted(mdw-icon) {\n        line-height: 19px;\n        font-weight: normal;\n        font-style: normal;\n        font-size: 24px;\n        letter-spacing: normal;\n        text-transform: none;\n        display: inline-block;\n        white-space: nowrap;\n        word-wrap: normal;\n      }\n\n      ::slotted(mdw-icon) {\n        width: 18px;\n        height: 18px;\n        font-size: 18px;\n        margin-left: -4px;\n        margin-right: 2px;\n        vertical-align: top;\n        line-height: 36px;\n      }\n      */\n\n      :host ::slotted(svg.mdw-icon) {\n        fill: currentColor;\n      }\n\n\n      /* primary */\n      :host(.mdw-primary) {\n        color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-primary.mdw-contained),\n      :host(.mdw-primary.mdw-raised),\n      :host(.mdw-primary.mdw-unelevated)  {\n        background-color: var(--mdw-theme-primary);\n        color: var(--mdw-theme-text-primary-on-background);\n      }\n\n      :host(.mdw-primary.mdw-outlined) {\n        border-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-primary)::before,\n      :host(.mdw-primary)::after {\n        background-color: var(--mdw-theme-primary);\n      }\n\n\n      /* secondary */\n\n      :host(.mdw-secondary) {\n        color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-secondary.mdw-contained),\n      :host(.mdw-secondary.mdw-raised),\n      :host(.mdw-secondary.mdw-unelevated) {\n        background-color: var(--mdw-theme-secondary);\n        color: var(--mdw-theme-text-primary-on-background);\n      }\n\n      :host(.mdw-secondary.mdw-outlined) {\n        border-color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-secondary)::before,\n      :host(.mdw-secondary)::after {\n        background-color: var(--mdw-theme-secondary);\n      }\n\n      /* error */\n\n      :host(.mdw-error) {\n        color: var(--mdw-theme-error);\n      }\n\n      :host(.mdw-error.mdw-contained),\n      :host(.mdw-error.mdw-raised),\n      :host(.mdw-error.mdw-unelevated) {\n        background-color: var(--mdw-theme-error);\n        color: var(--mdw-theme-text-primary-on-background);\n      }\n\n      :host(.mdw-error.mdw-outlined) {\n        border-color: var(--mdw-theme-error);\n      }\n\n      :host(.mdw-error)::before,\n      :host(.mdw-error)::after {\n        background-color: var(--mdw-theme-error);\n      }\n\n      :host(:not(.mdw-bottom-navigation):hover)::before {\n        opacity: 0.04;\n      }\n\n      :host(.mdw-show-spinner) span.text {\n        opacity: 0;\n      }\n\n      /* --- Ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-on-primary--rgb), 0.16);\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n      }\n\n      .mdw-button-ripple,\n      .mdw-button-focus-overlay {\n        border-radius: inherit;\n        position: absolute;\n        top: 0;\n        left: 0;\n        bottom: 0;\n        right: 0;\n        pointer-events: none;\n      }\n\n\n      :host(.mdw-primary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n\n      :host(.mdw-primary.mdw-contained) .mdw-ripple-element,\n      :host(.mdw-primary.mdw-raised) .mdw-ripple-element,\n      :host(.mdw-primary.mdw-unelevated) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n      :host(.mdw-secondary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.16);\n      }\n\n      :host(.mdw-secondary.mdw-contained) .mdw-ripple-element,\n      :host(.mdw-secondary.mdw-raised) .mdw-ripple-element,\n      :host(.mdw-secondary.mdw-unelevated) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n      :host(.error) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.16);\n      }\n\n      :host(.error.mdw-contained) .mdw-ripple-element,\n      :host(.error.mdw-raised) .mdw-ripple-element,\n      :host(.error.mdw-unelevated) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n\n      /* bottom navigation */\n      :host(.mdw-bottom-navigation) span.text {\n        display: flex;\n        flex-direction: column;\n        align-items: center;\n        font-size: 12px;\n        text-transform: none;\n        line-height: 12px;\n      }\n    '}}),customElements.define("mdw-card",class extends g{constructor(){super(),this.classList.add("mdw-elevation-1")}}),customElements.define("mdw-checkbox",class extends g{constructor(){super(),this.bound_handleClick=this.handleClick.bind(this),this.cloneTemplate(),this._defaultIcons(),this.state="unchecked"}connectedCallback(){this.hasAttribute("indeterminate")&&(this.indeterminate=!0),this.hasAttribute("checked")&&(this.checked=!0),this.hasAttribute("mdw-no-ripple")||(this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:[this],centered:!0})),this.addEventListener("click",this.bound_handleClick)}disconnectedCallback(){this.removeEventListener("click",this.bound_handleClick),this.ripple.destroy()}static get observedAttributes(){return["checked","indeterminate"]}attributeChangedCallback(e,t,n){this[e]=n}get state(){return this._state}set state(e){["checked","unchecked","indeterminate"].includes(e)||console.error(`ivalid state fro checkbox "${e}". Only excepts "checked", "unchecked", "indeterminate"`),this.setAttribute("mdw-state",e);const t=this._state!==e;this._state=e,!0===t&&this.dispatchEvent(new Event("change"))}get checked(){return"checked"===this.state}set checked(e){""===e&&(e=!0),this.state=!0===e?"checked":"unchecked"}get indeterminate(){return this.input.indeterminate}set indeterminate(e){""===e&&(e=!0),!0===e&&(this.state="indeterminate")}toggle(){this.checked=!this.checked}handleClick(){let e=!1;switch(this.state){case"indeterminate":this.state="checked",e=!0;break;case"checked":this.state="unchecked",e=!0;break;case"unchecked":this.state="checked",e=!0}}_defaultIcons(){0===this.children.length&&this.insertAdjacentHTML("afterbegin","\n        <mdw-icon mdw-checked>check_box</mdw-icon>\n        <mdw-icon mdw-unchecked>check_box_outline_blank</mdw-icon>\n        <mdw-icon mdw-indeterminate>indeterminate_check_box</mdw-icon>\n      ")}template(){return'\n      <slot></slot>\n      <div class="mdw-ripple mdw-checkbox-ripple"></div>\n    '}styles(){return'\n      :host {\n        --mdw-checkbox-size: 24px;\n        -webkit-tap-highlight-color: rgba(0, 0, 0, 0);\n        -webkit-box-sizing: content-box;\n        box-sizing: content-box;\n        cursor: pointer;\n        display: inline-block;\n        position: relative;\n        vertical-align: bottom;\n        white-space: nowrap;\n        cursor: pointer;\n\n        width: var(--mdw-checkbox-size);\n        height: var(--mdw-checkbox-size);\n        line-height: 0;\n        padding: 8px 0;\n      }\n\n      :host(.mdw-no-padding) {\n        padding: 0;\n      }\n\n      :host([disabled]) {\n        pointer-events: none;\n        cursor: not-allowed;\n        opacity: 0.5;\n      }\n\n      :host-context(.mdw-density-comfortable),\n      :host(.mdw-density-comfortable) {\n        padding: 9px;\n        margin: 0;\n      }\n\n      :host-context(.mdw-density-compact),\n      :host(.mdw-density-compact) {\n        padding: 5px;\n        margin: 0;\n      }\n\n      ::slotted(mdw-icon) {\n        position: absolute;\n        user-select: none;\n        outline: none;\n        font-size: var(--mdw-checkbox-size);\n      }\n\n      :host(.mdw-large) {\n        --mdw-checkbox-size: 48px;\n      }\n\n      :host(.mdw-primary) ::slotted(mdw-icon) {\n        color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-secondary) ::slotted(mdw-icon) {\n        color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-error) ::slotted(mdw-icon) {\n        color: var(--mdw-theme-error);\n      }\n      \n\n      /* --- state: checked --- */\n      :host([mdw-state="checked"]) ::slotted(mdw-icon[mdw-unchecked]),\n      :host([mdw-state="checked"]) ::slotted(mdw-icon[mdw-indeterminate]) {\n        display: none;\n      }\n\n      /* --- state: unchecked --- */\n      :host([mdw-state="unchecked"]) ::slotted(mdw-icon[mdw-checked]),\n      :host([mdw-state="unchecked"]) ::slotted(mdw-icon[mdw-indeterminate]) {\n        display: none;\n      }\n\n      /* --- state: indeterminate --- */\n      :host([mdw-state="indeterminate"]) ::slotted(mdw-icon[mdw-checked]),\n      :host([mdw-state="indeterminate"]) ::slotted(mdw-icon[mdw-unchecked]) {\n        display: none;\n      }\n\n\n      /* --- Ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n        background-color: rgba(var(--mdw-theme-on-background--rgb), 0.16);\n      }\n\n      .mdw-checkbox-ripple {\n        position: absolute;\n        left: 0;\n        top: 0;\n        right: 0;\n        bottom: 0;\n        border-radius: 50%;\n        z-index: 1;\n        pointer-events: none;\n        overflow: visible;\n      }\n\n      :host(.mdw-primary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n\n      :host(.mdw-secondary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.16);\n      }\n\n      :host(.mdw-error) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.16);\n      }\n    '}}),customElements.define("mdw-circular-progress",class extends g{constructor(){super(),this.insertedDiameters=[],this.cloneTemplate()}connectedCallback(){this.diameter=this.getAttribute("mdw-diameter")||100,this.render(),this.style.width=this.style.height=this.diameter+"px",this.value&&(this.value=this.value)}static get observedAttributes(){return["value"]}attributeChangedCallback(e,t,n){this[e]=n}get diameter(){return this._diameter}set diameter(e){this._diameter=parseInt((""+e).replace("px","")),this.insertedDiameters[this._diameter]||(this.insertedDiameters.push(this._diameter),this.shadowRoot.querySelector("style").sheet.insertRule(this._getAnimationText(),0))}get svg(){return this._svg||(this._svg=this.shadowRoot.querySelector("svg")),this._svg}get strokeWidth(){return this._strokeWidth||this.diameter/10}set strikeWidth(e){this._strokeWidth=parseInt((""+e).replace("px",""))}get value(){return this.getAttribute("value")}set value(e){this._value=Math.max(0,Math.min(100,parseInt((""+e).replace("px","")))),void 0!==this.diameter&&(this.circle.style.strokeDashoffset=this._strokeCircumference*(100-this._value)/100+"px")}get mode(){return"determinate"===this.getAttribute("mdw-mode")?"determinate":"indeterminate"}get circle(){return this._circle||(this._circle=this.shadowRoot.querySelector("circle")),this._circle}get _circleRadius(){return(this.diameter-10)/2}get _circleStrokeWidth(){return this.strokeWidth/this.diameter*100}get _strokeCircumference(){return 2*Math.PI*this._circleRadius}get INDETERMINATE_ANIMATION_TEMPLATE(){return"\n     @keyframes mat-progress-spinner-stroke-rotate-DIAMETER {\n        0%      { stroke-dashoffset: START_VALUE;  transform: rotate(0); }\n        12.5%   { stroke-dashoffset: END_VALUE;    transform: rotate(0); }\n        12.5001%  { stroke-dashoffset: END_VALUE;    transform: rotateX(180deg) rotate(72.5deg); }\n        25%     { stroke-dashoffset: START_VALUE;  transform: rotateX(180deg) rotate(72.5deg); }\n        25.0001%   { stroke-dashoffset: START_VALUE;  transform: rotate(270deg); }\n        37.5%   { stroke-dashoffset: END_VALUE;    transform: rotate(270deg); }\n        37.5001%  { stroke-dashoffset: END_VALUE;    transform: rotateX(180deg) rotate(161.5deg); }\n        50%     { stroke-dashoffset: START_VALUE;  transform: rotateX(180deg) rotate(161.5deg); }\n        50.0001%  { stroke-dashoffset: START_VALUE;  transform: rotate(180deg); }\n        62.5%   { stroke-dashoffset: END_VALUE;    transform: rotate(180deg); }\n        62.5001%  { stroke-dashoffset: END_VALUE;    transform: rotateX(180deg) rotate(251.5deg); }\n        75%     { stroke-dashoffset: START_VALUE;  transform: rotateX(180deg) rotate(251.5deg); }\n        75.0001%  { stroke-dashoffset: START_VALUE;  transform: rotate(90deg); }\n        87.5%   { stroke-dashoffset: END_VALUE;    transform: rotate(90deg); }\n        87.5001%  { stroke-dashoffset: END_VALUE;    transform: rotateX(180deg) rotate(341.5deg); }\n        100%    { stroke-dashoffset: START_VALUE;  transform: rotateX(180deg) rotate(341.5deg); }\n      }\n    "}_getAnimationText(){return this.INDETERMINATE_ANIMATION_TEMPLATE.replace(/START_VALUE/g,`${.95*this._strokeCircumference}`).replace(/END_VALUE/g,`${.2*this._strokeCircumference}`).replace(/DIAMETER/g,`${this.diameter}`)}template(){return`\n      <svg style="width: ${this.diameter}px; height: ${this.diameter}px;">\n        <circle\n          cx="50%"\n          cy="50%"\n          r="${this._circleRadius}"\n          style="\n            animation-name: mat-progress-spinner-stroke-rotate-${this.diameter};\n            stroke-dasharray: ${this._strokeCircumference}px;\n            stroke-width: ${this._circleStrokeWidth}%;\n          "\n          ></circle>\n      </svg>\n    `}styles(){return"\n      :host {\n        display: block;\n        position: relative;\n      }\n\n      svg {\n        position: absolute;\n        transform: rotate(-90deg);\n        top: 0;\n        left: 0;\n        transform-origin: center;\n        overflow: visible;\n      }\n\n      circle {\n        fill: transparent;\n        transform-origin: center;\n        transition: stroke-dashoffset 225ms linear;\n        stroke: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-white) circle {\n        stroke: white;\n      }\n\n      :host(.mdw-grey) circle {\n        stroke: grey;\n      }\n\n      :host(.mdw-secondary) circle {\n        stroke: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-error) circle {\n        stroke: var(--mdw-theme-error);\n      }\n\n      :host([mdw-mode='indeterminate']) {\n        animation: mat-progress-spinner-linear-rotate 2000ms linear infinite;\n      }\n\n      :host([mdw-mode='indeterminate']) circle {\n        transition-property: stroke;\n        animation-duration: 4000ms;\n        animation-timing-function: cubic-bezier(0.35, 0, 0.25, 1);\n        animation-iteration-count: infinite;\n      }\n\n      @keyframes mat-progress-spinner-linear-rotate {\n        0%       { transform: rotate(0deg); }\n        100%     { transform: rotate(360deg); }\n      }\n\n      @keyframes mat-progress-spinner-stroke-rotate-100 {\n        0% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotate(0);\n        }\n        12.5% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotate(0);\n        }\n        12.5001% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotateX(180deg) rotate(72.5deg);\n        }\n        25% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotateX(180deg) rotate(72.5deg);\n        }\n        25.0001% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotate(270deg);\n        }\n        37.5% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotate(270deg);\n        }\n        37.5001% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotateX(180deg) rotate(161.5deg);\n        }\n        50% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotateX(180deg) rotate(161.5deg);\n        }\n        50.0001% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotate(180deg);\n        }\n        62.5% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotate(180deg);\n        }\n        62.5001% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotateX(180deg) rotate(251.5deg);\n        }\n        75% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotateX(180deg) rotate(251.5deg);\n        }\n        75.0001% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotate(90deg);\n        }\n        87.5% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotate(90deg);\n        }\n        87.5001% {\n          stroke-dashoffset: 56.5486677px;\n          transform: rotateX(180deg) rotate(341.5deg);\n        }\n        100% {\n          stroke-dashoffset: 268.606171575px;\n          transform: rotateX(180deg) rotate(341.5deg);\n        }\n      }\n    "}});const y=new class{constructor(){this.currentDialog=null}open({title:e,message:t,okLabel:n,cancelLabel:i,position:s="center center",clickOutsideClose:o=!1}){return new Promise(r=>{const d=a.uid("dialog"),l=this.template({id:d,title:e,message:t,okLabel:n,cancelLabel:i,position:s});document.body.insertAdjacentHTML("beforeend",l);const c=document.querySelector(`#${d}`),h=e=>{r(e.detail.ok),c.removeEventListener("close",h),c.remove(),this.currentDialog=null};c.addEventListener("close",h),c.clickOutsideClose=o,this.currentDialog=c,requestAnimationFrame(()=>{c.open()})})}removeCurrent(){this.currentDialog.close()}template({id:e,title:t,message:n,okLabel:i,cancelLabel:s,position:o}){return`\n      <mdw-dialog id="${e}">\n        <mdw-panel mdw-position="${o}">\n          <mdw-dialog-container>\n            ${t?`<mdw-dialog-title>${t}</mdw-dialog-title>`:""}\n            <mdw-dialog-content>${n}</mdw-dialog-content>\n            <mdw-dialog-actions>\n              ${s?`<mdw-button class="mdw-error" onclick="${e}.close(false)">${s}</mdw-button>`:""}\n              ${i?`<mdw-button onclick="${e}.close(true)">${i}</mdw-button>`:""}\n            </mdw-dialog-actions>\n          </mdw-dialog-container>\n        </mdw-panel>\n      </mdw-dialog>\n    `}};window.MDWDialog=y;var k=y;customElements.define("mdw-dialog",class extends g{constructor(){super(),this.bound_onPanelClose=this.onPanelClose.bind(this),this.clickOutsideClose_=!1}disconnectedCallback(){this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose),this.panel.remove(),this.backdrop&&(this.backdrop.remove(),this.backdrop=void 0)}get panel(){return this._panel||(this._panel=this.querySelector("mdw-panel")),this._panel}get position(){return this._position||"center center"}set position(e){this._position=e}get clickOutsideClose(){return this.clickOutsideClose_}set clickOutsideClose(e){this.clickOutsideClose_=e}open(e=!1){this._fromService=e,this.panel.hoistToBody(),this.panel.setPosition(this.position),this.panel.addEventListener("MDWPanel:closed",this.bound_onPanelClose),this.backdrop=a.addBackdrop(this.panel,()=>{!0===this.clickOutsideClose&&this.close()}),requestAnimationFrame(()=>{this.panel.open()})}close(e){this.panel.close(),this.dispatchClose(e)}onPanelClose(){if(!this._fromService)return this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose),void(this.backdrop&&(this.backdrop.remove(),this.backdrop=void 0));this.remove()}dispatchClose(e=!1){this.dispatchEvent(new CustomEvent("close",{detail:{ok:e}}))}}),customElements.define("mdw-fab",class extends g{constructor(){super(),this.bound_asyncClick=this.asyncClick.bind(this),this.cloneTemplate(),this.setupAsync()}connectedCallback(){this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:this})}disconnectedCallback(){this.ripple.destroy(),this.removeEventListener("click",this.bound_asyncClick)}template(){return'\n      <span class="text"><slot></slot></span>\n      <span class="mdw-spinner-container"></span>\n      <div class="mdw-ripple mdw-fab-ripple"></div>\n    '}get dense(){return this.classList.contains("mdw-dense")}get spinnerContainer(){return this._spinnerContainer||(this._spinnerContainer=this.shadowRoot.querySelector(".mdw-spinner-container")),this._spinnerContainer}set disabled(e){e||""===e?this.setAttribute("disabled","disabled"):this.removeAttribute("disabled")}get pending(){return this.pending_}setupAsync(){this.hasAttribute("mdw-async")&&this.addEventListener("click",this.bound_asyncClick)}resolve(){!1!==this.pending_&&(this.pending_=!1,this.hideSpinner())}asyncClick(e){!0!==this.pending_&&(this.pending_=!0,this.showSpinner())}get spinnerStyle(){return this.dense?"position: absolute; left: calc(50% - 12px); top: 8px;":"position: absolute; left: calc(50% - 16px); top: 12px;"}get spinnerDiameter(){return this.dense?24:32}showSpinner(){this._showSpinner=!0,this.classList.add("mdw-show-spinner");const e=this.classList.contains("mdw-primary")||this.classList.contains("mdw-secondary")||this.classList.contains("mdw-error");this.spinnerContainer.innerHTML=`<mdw-circular-progress mdw-mode="indeterminate" mdw-diameter="${this.spinnerDiameter}" class="${e?"mdw-white":"mdw-grey"}" style="${this.spinnerStyle}"></mdw-circular-progress>`}hideSpinner(){this._showSpinner=!1,this.classList.remove("mdw-show-spinner"),this.spinnerContainer.innerHTML=""}styles(){return"\n      :host(.mdw-show-spinner) span.text {\n        opacity: 0;\n      }\n\n      :host-context(.mdw-density-comfortable) .mdw-spinner-container mdw-circular-progress {\n        top: 3px !important;\n      }\n\n      :host-context(.mdw-density-compact) .mdw-spinner-container mdw-circular-progress {\n        top: 0 !important;\n      }\n\n      /* --- Ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n      }\n\n      .mdw-fab-ripple,\n      .mdw-fab-focus-overlay {\n        border-radius: inherit;\n        position: absolute;\n        top: 0;\n        left: 0;\n        bottom: 0;\n        right: 0;\n        pointer-events: none;\n      }\n\n\n      :host(.mdw-primary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n      :host(.mdw-secondary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n\n      :host(.mdw-error) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n      }\n    "}}),customElements.define("mdw-icon",class extends g{constructor(){super(),this.cloneTemplate(),this.hasAttribute("mdw-src")&&this.render()}get src(){return this.getAttribute("mdw-src")}styles(){return"\n      :host {\n        font-family: 'Material Icons';\n        font-weight: normal;\n        font-style: normal;\n        font-size: 24px;\n        line-height: 1;\n        letter-spacing: normal;\n        text-transform: none;\n        display: inline-block;\n        white-space: nowrap;\n        word-wrap: normal;\n        direction: ltr;\n        font-feature-settings: 'liga';\n        -webkit-font-feature-settings: 'liga';\n        -webkit-font-smoothing: antialiased;\n      }\n\n      :host img {\n        width: 24px;\n        height: 24px;\n      }\n\n\n      :host(.mdw-primary) {\n        color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-secondary) {\n        color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-error) {\n        color: var(--mdw-theme-error);\n      }\n    "}template(){const e=this.src;return e?`<img src="${e}"></img>`:"<slot></slot>"}}),customElements.define("mdw-linear-progress",class extends g{constructor(){super(),this.cloneTemplate()}connectedCallback(){null===this.percent&&this.classList.add("mdw-query")}static get observedAttributes(){return["mdw-percent"]}attributeChangedCallback(e,t,n){switch(e){case"mdw-percent":this.percent=n}}get bar(){return this._bar||(this._bar=this.shadowRoot.querySelector(".mdw-bar")),this._bar}get percent(){return this.getAttribute("mdw-percent")}set percent(e){e<0&&(e=0),e>100&&(e=100),this.bar.style.width=`${e}%`}template(){return'\n      <div class="mdw-bar"></div>\n    '}styles(){return"\n      :host {\n        display: block;\n        position: relative;\n        width: 100%;\n        height: 4px;\n        padding-top: 0;\n        margin-bottom: 0;\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.18);\n      }\n\n      .mdw-bar {\n        position: absolute;\n        left: 0;\n        top: 0;\n        bottom: 0;\n        width: 100%;\n        height: 4px;\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-white) {\n        background-color: rgba(255, 255, 255, 0.18);\n      }\n\n      :host(.mdw-white) .mdw-bar {\n        background-color: white;\n      }\n\n      :host(.mdw-grey) {\n        background-color: rgba(50, 50, 50, 0.18);\n      }\n\n      :host(.mdw-grey) .mdw-bar {\n        background-color: grey;\n      }\n\n      :host(.mdw-secondary) {\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.18);\n      }\n\n      :host(.mdw-secondary) .mdw-bar {\n        background-color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-error) {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.18);\n      }\n\n      :host(.mdw-error) .mdw-bar {\n        background-color: var(--mdw-theme-error);\n      }\n\n\n      :host(.mdw-query) .mdw-bar {\n        transition: all 0.2s linear;\n        animation: query .8s infinite cubic-bezier(0.390, 0.575, 0.565, 1.000);\n      }\n\n      @keyframes query {\n        0% {\n          opacity: 1;\n          transform: translateX(35%) scale(.3, 1);\n        }\n        100% {\n          opacity: 0;\n          transform: translateX(-50%) scale(0, 1);\n        }\n      }\n    "}}),customElements.define("mdw-list",class extends g{constructor(){super(),this.selectedIndexes_=[]}static get observedAttributes(){return["mdw-select","mdw-select-onclick"]}attributeChangedCallback(e,t,n){switch(e){case"mdw-select":this.selectType=n;break;case"mdw-select-onclick":this.selectOnclick=null!==n}}set selectOnclick(e){this.selectOnclick_=e}get selectOnclick(){return this.selectOnclick_}set selectType(e){["single","multiple"].includes(e)||console.warn('mdw-list[mdw-select] attribute - only accepts "single" or "multiple"'),this.selectType_=e}get selectType(){return this.selectType_}get selected(){return[].concat(this.selectedIndexes_)}deselectAll(){[...this.querySelectorAll("mdw-list-item")].forEach(e=>e.deselect()),this.selectedIndexes_=[]}itemSelected(e){const t=Array.prototype.indexOf.call(this.children,e)-1;if("single"===this.selectType_){const e=[...this.children];this.selectedIndexes_.forEach(t=>e[t].deselect()),this.selectedIndexes_=[]}this.selectedIndexes_.push(t),this.handleChange()}itemDeselected(e){const t=Array.prototype.indexOf.call(this.children,e)-1;this.selectedIndexes_.splice(this.selectedIndexes_.indexOf(t),1),this.handleChange()}handleChange(){this.dispatchEvent(new CustomEvent("change",this))}closeExpanded(){this.dispatchEvent(new CustomEvent("MDWList:closeExpanded",this))}}),customElements.define("mdw-list-item",class extends g{constructor(){super(),this.bound_hrefClick=this.hrefClick.bind(this),this.bound_onSelect=this.onSelect.bind(this),this.bound_onclickSelect=this.onclickSelect.bind(this),this.bound_checkHREFCurrent=this.checkHREFCurrent.bind(this)}get list(){return this.parentNode}get expanded(){return this.querySelector("mdw-list-item-expanded")}get key(){return this.getAttribute("mdw-key")}isSelect(){return["single","multiple"].includes(this.list.selectType)}selectOnclick(){return!!this.list.selectOnclick}connectedCallback(){this.connectRipple(),this.connectHREF(),this.connectSelect();const e=this.expanded;e&&requestAnimationFrame(()=>{e.listItem=this})}disconnectedCallback(){this.ripple&&this.ripple.destroy(),this.removeEventListener("click",this.bound_hrefClick),this._selectEl&&this._selectEl.removeEventListener("change",this.bound_onSelect),this.removeEventListener("click",this.bound_onclickSelect),window.removeEventListener("hashchange",this.bound_checkHREFCurrent)}expand(){const e=this.expanded;e&&e.open()}connectRipple(){const e=this.querySelector(".mdw-ripple");e&&(this.ripple=new f({element:e,triggerElement:this}),this.classList.add("mdw-has-ripple"))}connectHREF(){this.hasAttribute("href")&&(this.checkHREFCurrent(),window.addEventListener("hashchange",this.bound_checkHREFCurrent),this.addEventListener("click",this.bound_hrefClick))}checkHREFCurrent(){if(!this.hasAttribute("href"))return;const e=document.location.href,t=document.location.hash;e===this.getAttribute("href")||e===this.getAttribute("href-alt")||t===this.getAttribute("href")||t===this.getAttribute("href-alt")?this.classList.add("mdw-current-link"):this.classList.remove("mdw-current-link")}hrefClick(){"_blank"!==this.getAttribute("target")?document.location.href=this.getAttribute("href"):window.open(this.getAttribute("href"),"_blank")}onSelect(e){e.target.checked?this.list.itemSelected(this):this.list.itemDeselected(this)}onclickSelect(e){this.selectOnclick()&&e.target!==this._selectEl&&(this._selectEl.checked=!this._selectEl.checked)}connectSelect(){this.isSelect()&&(this._selectEl=this.querySelector("mdw-checkbox"),this._selectEl&&this._selectEl.addEventListener("change",this.bound_onSelect),this.selectOnclick()&&this.addEventListener("click",this.bound_onclickSelect))}deselect(){this._selectEl.checked=!1}}),customElements.define("mdw-menu",class extends g{constructor(){super(),this.bound_onClick=this.onClick.bind(this),this.bound_onPanelClick=this.onPanelClick.bind(this),!0===a.isMobile?this.createSheet():this.createPanel()}connectedCallback(){this.button.addEventListener("click",this.bound_onClick),!0!==a.isMobile&&(this.classList.add("mdw-panel--container"),this.panel.classList.add("mdw-menu"))}onClick(){!0!==a.isMobile?(this.panel.setPosition(this.panelPosition),this.panel.autoPosition(),this.panel.clickBodyToClose(),this.panel.open(!0),this.panel.addEventListener("click",this.bound_onPanelClick)):this.sheet.open()}onPanelClick(){this.panel.close()}set panelPosition(e){this.panelPosition_=e}get panelPosition(){return this.panelPosition_||"inner-top inner-left"}get button(){return this.children[0]}get contentElement(){return this.querySelector("mdw-menu-content")}get panel(){return this.querySelector("mdw-panel")}get sheet(){return this.querySelector("mdw-sheet")}createSheet(){this.insertAdjacentHTML("beforeend",`\n      <mdw-sheet mdw-modal>\n        <mdw-sheet-content>\n          ${this.contentElement.innerHTML}\n        </mdw-sheet-content>\n      </mdw-sheet>\n    `),this.contentElement.remove()}createPanel(){this.contentElement&&(this.insertAdjacentHTML("beforeend",`\n      <mdw-panel>\n        ${this.contentElement.innerHTML}\n      </mdw-panel>\n    `),this.contentElement.remove())}}),customElements.define("mdw-navigation-rail",class extends g{constructor(){super(),o&&this.hasAttribute("mdw-desktop-only")?(this.style.display="none",this.style.pointerEvents="none"):(this.bound_routeChange=this.routeChange.bind(this),document.body.classList.add("mdw-has-navigation-rail"))}connectedCallback(){window.addEventListener("hashchange",this.bound_routeChange),window.addEventListener("DOMContentLoaded",this.bound_routeChange)}disconnectedCallback(){window.removeEventListener("hashchange",this.bound_routeChange),window.removeEventListener("DOMContentLoaded",this.bound_routeChange)}get path(){let e=window.location.hash.replace(/.*#/,"");return e.indexOf("?")>-1&&(e=e.split("?")[0]),"/"!==e.charAt(0)&&(e="/"+e),e}routeChange(){this.querySelectorAll(".mdw-current-link").forEach(e=>e.classList.remove("mdw-current-link"));let e=this.querySelectorAll(`[href="#${this.path}"]`);e&&0!==e.length||(e=this.querySelectorAll(`[alt-href="#${this.path}"]`)),e.forEach(e=>e.classList.add("mdw-current-link"))}}),customElements.define("mdw-panel-draggable-header",class extends g{constructor(){super(),this.bound_onDrag=this.onDrag.bind(this),this.cloneTemplate()}get draggableElement(){return this.shadowRoot.querySelector(".mdw-panel-draggable-header")}get panel(){return this.parentNode}connectedCallback(){c(this.draggableElement,this.bound_onDrag)}disconnectedCallback(){h(this.draggableElement,this.bound_onDrag)}onDrag(e){switch(e.state){case l.start:this._initialLeft=parseInt((this.panel.style.left||"0").replace("px","")),this._initialTop=parseInt((this.panel.style.top||"0").replace("px",""));break;case l.move:const t=this.panel;t.style.left=`${this._initialLeft+e.distance.x}px`,t.style.top=`${this._initialTop+e.distance.y}px`}}template(){return'\n      <div class="mdw-panel-draggable-header">\n        <slot></slot>\n        <mdw-icon>close</mdw-icon>\n      </div>\n    '}styles(){return"\n      .mdw-panel-draggable-header {\n        display: flex;\n        align-items: center;\n        justify-content: space-between;\n        cursor: move;\n        padding: 12px;\n        user-select: none;\n        -webkit-user-select: none;\n      }\n      mdw-icon {\n        cursor: pointer;\n      }\n    "}}),customElements.define("mdw-panel",class extends g{constructor(){super(),this.FOCUSABLE_ELEMENTS=["button:not(:disabled)",'[href]:not([aria-disabled="true"])',"input:not(:disabled)","select:not(:disabled)","textarea:not(:disabled)",'[tabindex]:not([tabindex="-1"]):not([aria-disabled="true"])'].join(", "),this._clickOutsideClose=!1,this._boundHandleBodyClick=this._handleBodyClick.bind(this),this._boundHandleKeydown=this._handleKeydown.bind(this),this.bound_close=this.close.bind(this),this._clickOutsideCloseIgnorElement=[],this._autoPosition=!1,this._animationConfig={type:"scale",opacity:!0},this.bound_onOpenTransitionEnd=this.onOpenAnimationEnd.bind(this)}connectedCallback(){this.classList.add("mdw-upgraded"),this.transformPropertyName=a.transformPropertyName}disconnectedCallback(){this.removeBodyClickEvent_(),this.removeKeydownEvent_(),clearTimeout(this._openAnimationEndTimerId),clearTimeout(this._closeAnimationEndTimerId),cancelAnimationFrame(this._animationRequestId)}static get observedAttributes(){return["mdw-position"]}attributeChangedCallback(e,t,n){switch(e){case"mdw-position":this._position=n}}set clickOutsideClose(e){this._clickOutsideClose=e}set setQuickOpen(e){this._isQuickOpen=e}get position(){return this._position}fullscreen(){this.classList.add("mdw-fullscreen")}setPosition(e){const t=e.split(" ");this._position=`${t[0]||"top"} ${t[1]||"left"}`,this.setAttribute("mdw-position",this._position),this._positionSet=!0}autoPosition(){this._autoPosition=!0}setAnimation(e){this._animationConfig=e}clickBodyToClose(){this._clickOutsideClose=!0}isOpen(){return this._isOpen}onOpenAnimationEnd(){this.style.transition="",this.style.transformOrigin="",this.style.overflow="",this.style.maxHeight="",this.classList.remove("mdw-panel--animating-open"),this.removeEventListener("transitionend",this.bound_onOpenTransitionEnd),this.notifyOpen()}open(e){void 0!==e&&(this._clickOutsideClose=e);const t=this.querySelectorAll(this.FOCUSABLE_ELEMENTS);this._firstFocusableElement=t[0],this._lastFocusableElement=t[t.length-1],this.saveFocus(),this._isQuickOpen?(this.classList.add("mdw-open"),this._isHoisted?this.setHoistedPosition():this.setPositionStyle()):(this.prepareAnimation(),this._isHoisted?this.setHoistedPosition():this.setPositionStyle(),this._animationRequestId=this._runNextAnimationFrame(()=>{switch(this._animationConfig.fullscreen&&this.classList.add("mdw-fullscreen"),this._animationConfig.type){case"height":this.style.transition="max-height .22s cubic-bezier(0,0,.2,1), transform .22s cubic-bezier(0,0,.2,1), opacity .22s linear",this.style.maxHeight=this.classList.contains("mdw-fullscreen")?"100%":`${this.scrollHeight}px`,this.style.transform="";break;case"scale":default:this.style.transition="transform .1s cubic-bezier(0,0,.2,1), opacity 0.1s linear",this.style.transform=""}this.style.opacity=1,this.addEventListener("transitionend",this.bound_onOpenTransitionEnd)})),this._addBodyClickEvent(),this._addKeydownEvent(),this.addEventListener("MDWPanel:close",this.bound_close),this._isOpen=!0}prepareAnimation(){switch(this.classList.add("mdw-open"),this.classList.add("mdw-panel--animating-open"),this._animationConfig.target&&this._animationConfig.fullscreen&&(this.style.width="100%"),this._animationConfig.type){case"height":switch(this.style.overflow="hidden",this.style.maxHeight=this._animationConfig.target?`${this._animationConfig.target.offsetHeight}px`:"0",this._animationConfig.origin){case"center":let e=this.classList.contains("mdw-fullscreen")?window.innerHeight/2:this.scrollHeight/2;this._animationConfig.target&&(e=this._animationConfig.target.getBoundingClientRect().y),this.style.transform=`translateY(${e}px)`;break;case"top":default:this._animationConfig.target&&(e=this._animationConfig.target.offsetTop,this.style.transform=`translateY(${e}px)`)}break;case"scale":default:this.style.transform="scale(0.9)",this.style.transformOrigin=this._animationConfig.origin||"center"}this._animationConfig.opacity&&(this.style.opacity=0)}async close(e){return new Promise(t=>{e&&e.stopPropagation(),this.removeEventListener("MDWPanel:close",this.bound_close),this._isQuickOpen?(this.classList.remove("mdw-open"),this.resetPosition(),t()):(this.classList.add("mdw-panel--animating-closed"),this.removeBodyClickEvent_(),this._animationRequestId=this._runNextAnimationFrame(()=>{this.classList.remove("mdw-open"),this._closeAnimationEndTimerId=setTimeout(()=>{this._closeAnimationEndTimerId=0,this.classList.remove("mdw-panel--animating-closed"),this.resetPosition(),this.notifyClose(),t()},75)})),this.removeKeydownEvent_(),this._isOpen=!1;const n=this.isFocused(),i=document.activeElement&&this.contains(document.activeElement);(n||i)&&this.restoreFocus()})}_runNextAnimationFrame(e){cancelAnimationFrame(this._animationFrame),this._animationFrame=requestAnimationFrame(()=>{this._animationFrame=0,clearTimeout(this._animationTimer),this._animationTimer=setTimeout(e,0)})}isFocused(){return document.activeElement===this}saveFocus(){this._previousFocus=document.activeElement}restoreFocus(){this.contains(document.activeElement)&&this._previousFocus&&this._previousFocus.focus&&this._previousFocus.focus()}focusFirstElement(){this._firstFocusableElement&&this._firstFocusableElement.focus&&this._firstFocusableElement.focus()}focusLastElement(){this._lastFocusableElement&&this._lastFocusableElement.focus&&this._lastFocusableElement.focus()}isFirstElementFocused(){this._firstFocusableElement&&(this._firstFocusableElement,document.activeElement)}isLastElementFocused(){this._lastFocusableElement&&(this._lastFocusableElement,document.activeElement)}_addBodyClickEvent(){this._clickOutsideClose&&setTimeout(()=>{this.hasBodyEvent=!0,document.body.addEventListener("click",this._boundHandleBodyClick)},0)}removeBodyClickEvent_(){this.hasBodyEvent&&document.body.removeEventListener("click",this._boundHandleBodyClick),this.hasBodyEvent=!1}_addKeydownEvent(){this.hasKeydownEvent=!0,document.body.addEventListener("keydown",this._boundHandleKeydown)}removeKeydownEvent_(){this.hasKeydownEvent&&document.body.removeEventListener("keydown",this._boundHandleKeydown),this.hasKeydownEvent=!1}ignoreElementOnClickToClose(e){this._clickOutsideCloseIgnorElement.push(e)}_handleBodyClick(e){const t=e.target;this._clickOutsideCloseIgnorElement.includes(t)||this.contains(t)||(this.removeBodyClickEvent_(),this.close())}_handleKeydown(e){const{key:t,keyCode:n,shiftKey:i}=e,s="Tab"===t||9===n;"Escape"===t||27===n?this.close():s&&(this.isLastElementFocused()&&!i?(this.focusFirstElement(),e.preventDefault()):this.isFirstElementFocused()&&i&&(this.focusLastElement(),e.preventDefault()))}notifyClose(){this.dispatchEvent(new Event("MDWPanel:closed",this))}notifyOpen(){this.dispatchEvent(new Event("MDWPanel:opened"),this)}hoistToBody(e){this._isHoisted||(this._container=e||this.parentNode,document.body.appendChild(this),this.classList.add("mdw-panel-hoisted"),this._isHoisted=!0)}_autoPositionHoisted(){if(!this._autoPosition)return;const e=window.innerHeight,t=(this.getBoundingClientRect(),this.offsetHeight);let n=this.offsetTop;if(n+t>e&&t<=e){let i=n-(e-t);i>20?i+=10:i/=2,n-=i}this.style.top=`${n}px`}setHoistedPosition(){const e=this._container.getBoundingClientRect();if(this.style.top=`${e.top}px`,this.style.left=`${e.left}px`,this._positionSet){let t=0,n=0;this.style.top=`${t}px`,this.style.left=`${n}px`,setTimeout(()=>{const{clientWidth:i,clientHeight:s}=document.documentElement,o=this.offsetHeight,r=this.offsetWidth,a=(this.position||" ").split(" "),d=a[0],l=a[1];switch(d){case"top":t=0;break;case"inner-top":t=e.y+12;break;case"bottom":t=s;break;case"center":t=s/2-o/2;break;case"inner-bottom":t=s-o-12}switch(l){case"left":n=0;break;case"inner-left":n=e.x+12;break;case"right":n=i;break;case"inner-right":n=i-r-12;break;case"center":n=i/2-r/2}this.style.width=`${this.width}px`,this.style.top=`${t}px`,this.style.left=`${n}px`},0)}else this._autoPositionHoisted()}setPositionStyle(e){e?this._parentOverride=e:this._parentOverride&&(e=this._parentOverride);this.position;let t=0,n=0;if(e)t=e.offsetWidth,n=e.offsetHeight;else{let e=this.parentNode;"MDW-SNACKBAR"===e.nodeName&&(e=e.parentNode);const i=e.getBoundingClientRect();t=i.width,n=i.height}const i=this.offsetWidth,s=this.offsetHeight,o=(this.position||" ").split(" "),r=o[0],a=o[1];let d=0,l=0;switch(r){case"top":d=-s;break;case"bottom":d=n;break;case"center":d=n/2-s/2;break;case"inner-bottom":d=n-s}switch(a){case"left":l=-i;break;case"right":l=t;break;case"inner-right":l=t-i;break;case"center":l=t/2-i/2}if(this._autoPosition){const{clientWidth:e,clientHeight:o}=document.documentElement,{x:r,y:a}=this.getBoundingClientRect();a+s>o&&(d=n-s),r+i>e&&(l=t-i)}this.style.top=`${parseInt(d)}px`,this.style.left=`${parseInt(l)}px`}resetPosition(){this.style.top="",this.style.left=""}}),customElements.define("mdw-radio",class extends g{constructor(){super(),this.insertAdjacentHTML("beforeend",this.rippleTemplate())}connectedCallback(){this.parentNode.initialValue===this.value&&(this.input.checked=!0),this.input.hasAttribute("type")||this.input.setAttribute("type","radio"),this.input.hasAttribute("name")||this.input.setAttribute("name",this.name),this.ripple=new f({element:this.querySelector(".mdw-ripple"),triggerElement:[this.input],radius:20,centered:!0})}disconnectedCallback(){this.ripple.destroy()}get value(){return this.input.value}get input(){return this.querySelector("input")}get name(){return this.parentNode&&this.parentNode.hasAttribute("name")?this.name_=this.parentNode.getAttribute("name"):this.hasAttribute("name")&&(this.name_=this.getAttribute("name")),this.name_||(this.name_=a.uid(),this.parentNode?this.parentNode.setAttribute("name",this.name_):this.setAttribute("name",this.name_)),this.name_}rippleTemplate(){return'\n      <div class="mdw-radio-background">\n        <div class="mdw-radio__outer-circle"></div>\n        <div class="mdw-radio__inner-circle"></div>\n      </div>\n      <div class="mdw-ripple mdw-radio-ripple"></div>\n    '}}),customElements.define("mdw-radio-group",class extends g{constructor(){super(),this.initialValue=this.getAttribute("mdw-value")}}),customElements.define("mdw-select",class extends g{constructor(){super(),this.setupLabel_(),this._isEnhanced&&this.prepareEnhance_(),this.classList.add("mdw-no-animation"),this.cloneTemplate(!0),this.bound_onFocus=this.onFocus.bind(this),this.bound_onBlur=this.onBlur.bind(this),this.bound_onClick=this.onClick.bind(this),this.bound_onChange=this.onChange.bind(this),this.bound_onPanelClick=this.onPanelClick.bind(this),this.bound_onKeyDown=this.onKeyDown.bind(this)}connectedCallback(){this._isEnhanced?(this._selected&&(this.value=this._selected.value),this.shadowRoot.querySelector("render-block").addEventListener("click",this.bound_onClick),document.body.addEventListener("keydown",this.bound_onKeyDown)):(this.selectElement.addEventListener("focus",this.bound_onFocus),this.selectElement.addEventListener("blur",this.bound_onBlur),this.selectElement.addEventListener("change",this.bound_onChange)),this.onChange(),setTimeout(()=>{this.classList.remove("mdw-no-animation"),this._isEnhanced&&(this.panel.style.minWidth=`${this.offsetWidth}px`)},0)}disconnectedCallback(){this._isEnhanced?(this.shadowRoot.querySelector("render-block").removeEventListener("click",this.bound_onClick),document.body.removeEventListener("keydown",this.bound_onKeyDown)):(this.selectElement.removeEventListener("focus",this.bound_onFocus),this.selectElement.removeEventListener("blur",this.bound_onBlur),this.selectElement.removeEventListener("change",this.bound_onChange))}get value(){return this._isEnhanced?this._value:this.selectElement.value||this._value}set value(e){this._value=e,this.onChange(),this.dispatchEvent(new Event("change"))}get selectElement(){return a.querySlotted(this,"select")}get label(){return this.shadowRoot.querySelector("label")}get labelWidth(){return.9*this.label.offsetWidth}get enhacedElementId(){return this._enhacedElementId||(this._enhacedElementId=`select-enhanced-${a.uid()}`),this._enhacedElementId}get panel(){return document.querySelector(`#${this.enhacedElementId}`)}get sheet(){return document.querySelector(`#${this.enhacedElementId}`)}get _isEnhanced(){return null!==this.getAttribute("mdw-enhanced")}get outlined(){return[].slice.apply(this.classList||[]).includes("mdw-outlined")}get notch(){return this._notch||(this._notch=this.shadowRoot.querySelector(".mdw-outlined-notch")),this._notch}setupLabel_(){const e=this.querySelector("label");e&&(this._labelText=e.innerText,e.remove())}prepareEnhance_(){this._optionsMap=[...this.querySelectorAll("option")].map(e=>({text:e.innerText,value:e.value,selected:e.hasAttribute("selected")})),this._selected=this._optionsMap.filter(({selected:e})=>!0===e)[0]||{text:"",value:""};const e=this.querySelector("select");if(e){const t=e.getAttribute("onchange");t&&this.setAttribute("onchange",t),e.remove()}a.isMobile?this.prepareSheet_():this.preparePanel_()}preparePanel_(){const e=`\n      <mdw-panel id="${this.enhacedElementId}" mdw-position="bottom inner-left" class="mdw-panel-hoisted">\n        <mdw-list>\n          ${this._optionsMap.map(({text:e,value:t,selected:n})=>`\n            <mdw-list-item value="${t}"${n?" selected":""}>${e}</mdw-list-item>\n          `).join("\n")}\n        </mdw-list>\n      </mdw-panel>\n    `;document.body.insertAdjacentHTML("beforeend",e);const t=this.panel;t.setAnimation({target:this.querySelector("select"),type:"scale",origin:"top",opacity:!0}),t.hoistToBody(this)}prepareSheet_(){const e=`\n      <mdw-sheet mdw-modal id=${this.enhacedElementId}>\n        <mdw-sheet-content>\n          <mdw-list>\n            ${this._optionsMap.map(({text:e,value:t,selected:n})=>`\n              <mdw-list-item value="${t}"${n?" selected":""}>${e}</mdw-list-item>\n            `).join("\n")}\n          </mdw-list>\n        </mdw-sheet-content>\n      </mdw-sheet>\n    `;document.body.insertAdjacentHTML("beforeend",e)}onFocus(){this.classList.add("mdw-focused"),this.outlined&&(this.notch.style.width=this.labelWidth+"px")}onBlur(){this.classList.remove("mdw-focused"),this.classList.toggle("mdw-not-empty",this.value),this._isEnhanced&&(a.isMobile?(this.sheet.removeEventListener("MDWSheet:closed",this.bound_onBlur),this.sheet.removeEventListener("click",this.bound_onPanelClick)):(this.panel.removeEventListener("MDWPanel:closed",this.bound_onBlur),this.panel.removeEventListener("click",this.bound_onPanelClick))),a.unlockPageScroll()}onChange(){this.value&&this.label?(this.label.classList.add("mdw-select--float-above"),this.label.classList.remove("mdw-empty-no-float"),this.outlined&&(this.notch.style.width=this.labelWidth+"px")):(this.label.classList.remove("mdw-select--float-above"),this.label.classList.add("mdw-empty-no-float"),this.outlined&&(this.notch.style.width="0"))}onClick(e){if(this._focusIndex,this.onFocus(),a.isMobile){const e=this.sheet;e.open(),e.addEventListener("MDWSheet:closed",this.bound_onBlur),e.addEventListener("click",this.bound_onPanelClick);const t=e.querySelector(".mdw-focused");t&&t.classList.remove("mdw-focused");const n=e.querySelector("[selected]");n&&n.classList.add("mdw-focused")}else{const e=this.panel;e.autoPosition(),e.open(!0),e.addEventListener("MDWPanel:closed",this.bound_onBlur),e.addEventListener("click",this.bound_onPanelClick);const t=e.querySelector(".mdw-focused");t&&t.classList.remove("mdw-focused");const n=e.querySelector("[selected]");n&&n.classList.add("mdw-focused")}a.lockPageScroll()}onPanelClick(e){if(!e.target.hasAttribute("value"))return;this.value=e.target.getAttribute("value"),this.setSelectedText(e.target.innerText);const t=this.panel.querySelector("[selected]");t&&t.removeAttribute("selected"),e.target.setAttribute("selected",""),this.panel.close()}setSelectedText(e){this.shadowRoot.querySelector(".mdw-select__selected-text").innerText=e}get internalStylesFile(){return"./internal.css"}template(){return`\n      <i class="mdw-select__icon"></i>\n      ${this._isEnhanced?`\n        <div class="mdw-select__selected-text">${this._selected.text}</div>\n      `:"<slot></slot>"}\n      <label>${this._labelText}</label>\n      ${this.outlined?"":'<div class="mdw-line-ripple"></div>'}\n      ${this.outlined?'\n        <div class="mdw-outlined-border-container">\n          <div class="mdw-outlined-leading"></div>\n          <div class="mdw-outlined-notch"></div>\n          <div class="mdw-outlined-trailing"></div>\n        </div>\n      ':""}\n    `}onKeyDown(e){if(this.panel.isOpen())switch(e.keyCode){case 40:case 39:this.focusNext(),e.preventDefault();break;case 38:case 37:this.focusPrevious(),e.preventDefault();break;case 13:this.selectFocused(),e.preventDefault();break;default:if(e.keyCode>=31||e.keyCode<=90){const t=this.keyboardSearchNodes(e.keyCode);void 0!==t&&this.selectNode(t),e.stopPropagation(),e.preventDefault()}}}keyboardSearchNodes(e){void 0!==this._clearSearchTimeout&&clearTimeout(this._clearSearchTimeout),this._clearSearchTimeout=setTimeout(()=>{this._clearSearchTimeout=void 0,this._keyboardSearchStr="",this._keyboardOptionNames=void 0},300),void 0===this._keyboardSearchStr&&(this._keyboardSearchStr=""),this._keyboardSearchStr+=String.fromCharCode(e);const t=new RegExp("^"+this._keyboardSearchStr,"i");this._keyboardOptionNames||(this._keyboardOptionNames=[...this.panel.querySelectorAll("mdw-list-item")].map(e=>e.innerText));const n=this._keyboardOptionNames.length;let i=0;for(;i<n;){if(t.test(this._keyboardOptionNames[i]))return i;i+=1}}selectNode(e){const t=[...this.panel.querySelectorAll("mdw-list-item")];this._focusIndex=e,this._focusedOption&&this._focusedOption.classList.remove("mdw-focused"),this._focusedOption=t[this._focusIndex],this._focusedOption.classList.add("mdw-focused")}focusNext(){if(!this.panel.isOpen())return;const e=[...this.panel.querySelectorAll("mdw-list-item")];if(void 0===this._focusIndex){const t=e.findIndex(e=>e.classList.contains("mdw-focused"));t>=0&&(this._focusedOption=e[t]),this._focusIndex=t<=0?1:t+1}else this._focusIndex+=1;this._focusIndex>e.length-1&&(this._focusIndex=e.length-1),this._focusedOption&&this._focusedOption.classList.remove("mdw-focused"),this._focusedOption=e[this._focusIndex],this._focusedOption.classList.add("mdw-focused")}focusPrevious(){if(!this.panel.isOpen())return;const e=[...this.panel.querySelectorAll("mdw-list-item")];void 0===this._focusIndex?this._focusIndex=0:this._focusIndex-=1,this._focusIndex<=0&&(this._focusIndex=0),this._focusedOption&&this._focusedOption.classList.remove("mdw-focused"),this._focusedOption=e[this._focusIndex],this._focusedOption.classList.add("mdw-focused")}selectFocused(){if(!this.panel.isOpen())return;const e=[...this.panel.querySelectorAll("mdw-list-item")];(null==this._focusIndex||this._focusIndex>e.length-1)&&(this._focusIndex=0),this.onPanelClick({target:e[this._focusIndex]})}styles(){return"\n      :host-context(.mdw-density-comfortable) .mdw-select__icon {\n        bottom: 15px;\n      }\n\n      :host-context(.mdw-density-comfortable) .mdw-select__selected-text {\n        height: 48px;\n        padding-top: 16px;\n      }\n\n      :host(.mdw-outlined):host-context(.mdw-density-comfortable) .mdw-select__icon {\n        bottom: 20px;\n      }\n      \n      :host(.mdw-outlined):host-context(.mdw-density-comfortable) label {\n        top: 18px;\n      }\n\n      :host-context(.mdw-density-compact) .mdw-select__icon {\n        bottom: 12px;\n      }\n\n      :host(.mdw-outlined):host-context(.mdw-density-compact) .mdw-select__icon {\n        top: 18px;\n      }\n\n      :host(.mdw-outlined):host-context(.mdw-density-compact) label {\n        top: 12px;\n      }\n\n      :host-context(.mdw-density-compact) .mdw-select__selected-text {\n        height: 40px;\n        line-height: 1.1rem;\n      }\n\n      :host-context(.mdw-density-compact) ::slotted(select) {\n        height: 40px;\n        padding-top: 12px;\n      }\n\n      :host-context(.mdw-density-compact) label {\n        top: 16px;\n      }\n\n      :host(.mdw-outlined.mdw-focused):host-context(.mdw-density-compact) label,\n      :host(.mdw-outlined):host-context(.mdw-density-compact) label.mdw-select--float-above {\n        transform: translateY(-100%) scale(0.75);\n      }\n\n      ::slotted(label.mdw-empty-no-float) {\n        transform: none;\n      }\n\n      :host(.mdw-focused) .mdw-select__icon {\n        transform: rotate(180deg) translateY(-5px);\n        transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1);\n      }\n\n      .mdw-select__icon {\n        transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1);\n        pointer-events: none;\n        position: absolute;\n        bottom: 23px;\n        left: auto;\n        right: 8px;\n        width: 0;\n        height: 0;\n        border-left: 5px solid transparent;\n        border-right: 5px solid transparent;\n        border-top: 5px solid var(--mdw-theme-on-secondary);\n      }\n\n      ::slotted(select:focus) .mdw-select__icon,\n      :host(.mdw-focused:focus) .mdw-select__icon {\n        transform: rotate(180deg) translateY(-5px);\n        transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1);\n      }\n\n      :host(:not(.mdw-select--disabled)) ::slotted(select),\n      :host(:not(.mdw-select--disabled)) .mdw-select__selected-text {\n        border-bottom-color: rgba(var(--mdw-theme-on-background--rgb), 0.54);\n        color: var(--mdw-theme-on-primary);\n      }\n\n      :host(.mdw-focused:not(.mdw-select--disabled)) ::slotted(select),\n      :host(.mdw-focused:not(.mdw-select--disabled)) .mdw-select__selected-text,\n      :host(:not(.mdw-select--disabled)) ::slotted(select:focus),\n      :host(.mdw-focused:focus:not(.mdw-select--disabled)) .mdw-select__selected-text {\n        border-bottom: 2px solid;\n        border-bottom-color: var(--mdw-theme-primary);\n        height: calc(100% + 1px); /* add 1px to height so the text does not get pushed up by border size change */\n      }\n\n      :host(.mdw-outlined) ::slotted(select),\n      :host(.mdw-outlined) .mdw-select__selected-text {\n        border: none;\n      }\n\n      ::slotted(select),\n      .mdw-select__selected-text {\n        position: absolute;\n        padding: 20px 52px 4px 16px;\n        font-family: Roboto,sans-serif;\n        font-size: 1rem;\n        line-height: 1.75rem;\n        font-weight: 400;\n        letter-spacing: .009375em;\n        text-decoration: inherit;\n        text-transform: inherit;\n        box-sizing: border-box;\n        width: 100%;\n        height: 56px;\n        border: none;\n        border-bottom: 1px solid;\n        outline: none;\n        background-color: transparent;\n        color: inherit;\n        white-space: nowrap;\n        cursor: pointer;\n        appearance: none;\n        -webkit-appearance: none;\n        -moz-appearance: none;\n      }\n\n      /* outlined */\n      :host(.mdw-outlined) ::slotted(select),\n      :host(.mdw-outlined) .mdw-select__selected-text {\n        padding: 12px 52px 12px 16px;\n        display: flex;\n        border: none;\n        background-color: transparent;\n        z-index: 1;\n      }\n      :host(.mdw-outlined) ::slotted(select) {\n        border-radius: 4px;\n      }\n\n      ::slotted(select) {\n        border-radius: 4px 4px 0 0;\n      }\n\n      :host([dir=rtl]) ::slotted(select),\n      ::slotted(select[dir=rtl]),\n      :host([dir=rtl]) .mdw-select__selected-text,\n      .mdw-select__selected-text[dir=rtl] {\n        padding-left: 52px;\n        padding-right: 16px;\n      }\n\n\n      label {\n        font-size: 1rem;\n        line-height: 1.75rem;\n        font-weight: 400;\n        letter-spacing: 0.009375em;\n        text-decoration: inherit;\n        text-transform: inherit;\n        position: absolute;\n        left: 0;\n        transform-origin: left top;\n        transition: transform 150ms cubic-bezier(0.4, 0, 0.2, 1), color 150ms cubic-bezier(0.4, 0, 0.2, 1);\n        line-height: 1.15rem;\n        text-align: left;\n        text-overflow: ellipsis;\n        white-space: nowrap;\n        cursor: text;\n        overflow: hidden;\n        will-change: transform;\n        transform: none;\n        pointer-events: none;\n        color: rgba(var(--mdw-theme-on-background--rgb), .6);\n        z-index: 1;\n\n        left: 16px;\n        right: initial;\n        top: 21px;\n      }\n\n      :host(.mdw-focused) label {\n        color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-no-animation) label {\n        transition: none;\n      }\n\n      label:not(.mdw-empty-no-float) {\n        transform: translateY(-70%) scale(0.75);\n      }\n\n      ::slotted(select:focus) + label,\n      label.mdw-select--float-above {\n        transform: translateY(-70%) scale(0.75);\n      }\n\n      :host(.mdw-outlined.mdw-focused) label,\n      :host(.mdw-outlined) label.mdw-select--float-above {\n        transform: translateY(-132%) scale(0.75);\n      }\n\n      :host(.mdw-select--with-leading-icon) label {\n        left: 48px;\n        right: initial;\n      }\n\n      :host(.mdw-outlined) label {\n        left: 15px;\n        right: initial;\n        top: 18px;\n      }\n\n      :host(.mdw-outlined.mdw-select--with-leading-icon) label {\n        left: 36px;\n        right: initial;\n      }\n\n      :host(.mdw-outlined.mdw-select--with-leading-icon) label.mdw-select--float-above {\n        left: 36px;\n        right: initial;\n      }\n\n      .mdw-outlined-border-container {\n        display: -ms-flexbox;\n        display: flex;\n        position: absolute;\n        top: 0;\n        right: 0;\n        left: 0;\n        -webkit-box-sizing: border-box;\n        box-sizing: border-box;\n        width: 100%;\n        max-width: 100%;\n        height: 100%;\n        text-align: left;\n        pointer-events: none;\n      }\n\n      .mdw-outlined-leading {\n        border-radius: 4px 0 0 4px;\n        border-left: 1px solid;\n        border-right: none;\n        width: 12px;\n      }\n\n      .mdw-outlined-notch {\n        -ms-flex: 0 0 auto;\n        flex: 0 0 auto;\n        width: auto;\n        max-width: calc(100% - 12px * 2);\n      }\n\n      .mdw-outlined-trailing {\n        border-left: none;\n        border-right: 1px solid;\n        border-radius: 0 4px 4px 0;\n        -ms-flex-positive: 1;\n        flex-grow: 1;\n      }\n\n      .mdw-outlined-leading,\n      .mdw-outlined-notch,\n      .mdw-outlined-trailing {\n        box-sizing: border-box;\n        height: 100%;\n        border-top: 1px solid;\n        border-bottom: 1px solid;\n        pointer-events: none;\n\n        border-color: rgba(var(--mdw-theme-on-background--rgb), 0.54);\n      }\n\n      .mdw-outlined-notch {\n        border-top: none;\n      }\n\n      :host(.mdw-focused) .mdw-outlined-leading,\n      :host(.mdw-focused) .mdw-outlined-notch,\n      :host(.mdw-focused) .mdw-outlined-trailing,\n      ::slotted(select:focus) .mdw-outlined-leading,\n      ::slotted(select:focus) .mdw-outlined-notch,\n      ::slotted(select:focus) .mdw-outlined-trailing {\n        border-width: 2px;\n        border-color: var(--mdw-theme-primary);\n      }\n\n      :host(.invalid) .mdw-outlined-leading,\n      :host(.invalid) .mdw-outlined-notch,\n      :host(.invalid) .mdw-outlined-trailing {\n        border-color: var(--mdw-theme-error);\n      }\n    "}}),customElements.define("mdw-sheet-bottom-header",class extends g{constructor(){super(),this.parentNode._registerHeader&&this.parentNode._registerHeader(this),this.sheetInstance=this.parentNode,this.bound_close=this.close.bind(this),this.bound_toTop=this.toTop.bind(this),this.innerHTML=this.template(this.isModal,this.title,this.innerHTML);const e=this.querySelector("mdw-header");e&&(this.classList.add("has-standard-header"),e.setAttribute("class",this.classList.toString()))}connectedCallback(){this.closeButton&&this.closeButton.addEventListener("click",this.bound_close),this._hasMdwHeader&&this.mdwHeader.addEventListener("click",this.bound_toTop)}disconnectedCallback(){this.closeButton&&this.closeButton.removeEventListener("click",this.bound_close),this._hasMdwHeader&&this.mdwHeader.removeEventListener("click",this.bound_toTop)}get closeButton(){return this.querySelector("#mdw-sheet-close-action")}get mdwHeader(){return this.querySelector("mdw-header")}get isModal(){return"modal"===this.parentNode.type}get title(){return this._title?this._title:this.hasAttribute("mdw-title")?this.getAttribute("mdw-title"):""}set title(e){this._title=e}close(){if(this.isModal)return this.parentNode.close();this.parentNode.exitFullscreen()}disable(){this.classList.add("mdw-disabled")}show(){this.classList.add("mdw-show"),this.isShowing=!0}hide(){this.classList.remove("mdw-show"),this.isShowing=!1}toTop(){console.log("to top"),this.parentNode._transitionToNearestPosition()}template(e,t,n){const i=(new DOMParser).parseFromString(n,"text/html").querySelector("mdw-sheet-top-bar");return i&&(n=n.replace(i.outerHTML,"")),n&&(this._hasMdwHeader=!0),`\n      ${i?i.outerHTML:`\n        <mdw-sheet-top-bar>\n          <mdw-button id="mdw-sheet-close-action" class="mdw-icon">\n            ${e?"<mdw-icon>close</mdw-icon>":"<mdw-icon>keyboard_arrow_down</mdw-icon>"}\n          </mdw-button>\n          ${t}\n        </mdw-sheet-top-bar>\n      `}\n\n      ${n&&`<mdw-header>\n\n          ${n}\n        </mdw-header>`}\n    `}});class E{constructor(e){this.componentElement=e,this.isAnchored=e.hasAttribute("mdw-anchored")}get initialPosition(){return this.componentElement._headerHeight+window.innerHeight/4}get minimizedPosition(){return this.componentElement._headerHeight}get clientPosition(){return window.innerHeight/4}get isDraggable(){return this.componentElement.contentHeight>this.clientPosition}registerHeader(e){this.headerElement=e,this.headerElement.title=this.title}setupHeader(){const e=this.componentElement.querySelector("mdw-header");this.componentElement.insertAdjacentHTML("afterbegin",`<mdw-sheet-bottom-header mdw-title="${this.componentElement.title}" class="${e.classList.toString()}">${e&&e.innerHTML}</mdw-sheet-bottom-header>`),e&&e.remove()}addBackdrop(){}removeBackdrop(){}handleOnMove({position:e,isAtTop:t,isAboveTop:n,targetingTop:i,targetingInitial:s}){i||t||n?this.headerElement.show():this.headerElement.hide(),t||n?this.headerElement.classList.add("mdw-fullscreen"):this.headerElement.classList.remove("mdw-fullscreen"),n?this.headerElement.classList.add("mdw-is-above-top"):this.headerElement.classList.remove("mdw-is-above-top")}}class x{constructor(e){this.componentElement=e}get initialPosition(){return this.minimizedPosition}get minimizedPosition(){const e=this.componentElement.contentHeight;return Math.min(e,this.clientPosition)}get clientPosition(){return window.innerHeight/2}get isDraggable(){return this.componentElement.contentHeight>this.clientPosition}registerHeader(e){this.headerElement=e,this.headerElement.title=this.title}setupHeader(){this.componentElement.insertAdjacentHTML("afterbegin",`<mdw-sheet-bottom-header mdw-title="${this.componentElement.title}"></mdw-sheet-bottom-header>`)}addBackdrop(){this.backdrop=a.addBackdrop(this.componentElement,()=>{this.componentElement.close()})}removeBackdrop(){this.backdrop&&this.backdrop.remove(),this.backdrop=void 0}handleOnMove({position:e,isAtTop:t,isAboveTop:n,targetingTop:i,targetingInitial:s}){i||t||n?this.headerElement.show():this.headerElement.hide(),t||n?this.headerElement.classList.add("mdw-fullscreen"):this.headerElement.classList.remove("mdw-fullscreen"),n?this.headerElement.classList.add("mdw-is-above-top"):this.headerElement.classList.remove("mdw-is-above-top")}}customElements.define("mdw-sheet-bottom",class extends g{constructor(){switch(super(),this.type){case"modal":this._helpers=new x(this);break;default:this._helpers=new E(this)}this.bound_onTransitionEnd=this._onTransitionEnd.bind(this),this.bound_onTransitionEndClose=this._onTransitionEndClose.bind(this),this.bound_onDrag=this._onDrag.bind(this),this._helpers.setupHeader(),this._setupOverScroll()}get title(){return this.getAttribute("mdw-title")}get type(){return this.hasAttribute("mdw-modal")?"modal":"standard"}get contentElement(){return this.querySelector("mdw-content")}get contentHeight(){return this.contentElement.offsetHeight}get _isDraggable(){return this._helpers.isDraggable}get _headerHeight(){return this._helpers.headerElement.offsetHeight}get _maxScroll(){return this.contentHeight+this._headerHeight}get _topPosition(){const e=window.innerHeight,t=this.contentHeight,n=this._headerHeight;return"standard"===this.type?e-(55-n):e-(t>=e-n?55-n:e-n-t)}get _scrollDistanceRemaining(){return this._maxScroll-this._currentPosition}get _initialPosition(){return this._helpers.initialPosition}get _minimizedPosition(){return this._helpers.minimizedPosition}get _isAnchored(){return this.hasAttribute("mdw-anchored")}disconnectedCallback(){this.classList.contains("mdw-dragging")&&(a.unlockPageScroll(),a.disableUserSelect()),this._cancelTransitions()}open(){this._cancelTransitions(),this.classList.add("mdw-show"),this.style.height=`${this.contentHeight+this._headerHeight}px`,this._positionBottom(),this._isAnchored?this._transitionToPosition(this._minimizedPosition):this._transitionToPosition(this._initialPosition),this._helpers.addBackdrop(),this._isDraggable&&(c(this.contentElement,this.bound_onDrag),this._helpers.headerElement&&c(this._helpers.headerElement,this.bound_onDrag))}async close(){return new Promise(e=>{h(this.contentElement,this.bound_onDrag),this._helpers.headerElement&&h(this._helpers.headerElement,this.bound_onDrag),this._cancelTransitions(),this.classList.add("mdw-animating-close"),this._positionBottom(),this.addEventListener("transitionend",()=>{this._cancelTransitions(),this.classList.remove("mdw-show"),this._helpers.removeBackdrop(),e()},{once:!0})})}minimize(){this._cancelTransitions(),this._transitionToPosition(this._minimizedPosition)}exitFullscreen(){this._cancelTransitions(),this._isAnchored?this._transitionToPosition(this._minimizedPosition):this._transitionToPosition(this._initialPosition)}toggle(){this.classList.contains("mdw-show")?this.close():this.open()}_registerHeader(e){this._helpers.registerHeader(e)}_positionTop(){this._setPosition(this._topPosition)}_positionBottom(){this._currentPosition=0,this.style.top="100%"}_positionInitial(){this._setPosition(this._initialPosition)}_positionMinimized(){this._setPosition(this._minimizedPosition)}_setPosition(e){const t=this._maxScroll;let n=0;if(e>t){const i=100;n=i*Math.log(e-t+i)-i*Math.log(i),e=t}const i=window.innerHeight+(this._headerHeight-55);this._isAtTop=e===i,this._isAboveTop=e>i,this._isAtOrAboveTop=this._isAtTop||this._isAboveTop,this._currentPosition=e,this.style.top=`calc(100% - ${e+n}px)`;const s=window.innerHeight-this._initialPosition,o=!this._isAtOrAboveTop&&e-this._initialPosition>=s/2,r=!this._isAtOrAboveTop&&e-this._initialPosition<s/2;this._helpers.handleOnMove({position:e,isAtTop:this._isAtTop,isAboveTop:this._isAboveTop,targetingTop:o,targetingInitial:r})}_transitionToNearestPosition(){const e=this._currentPosition;if(this._initialOffsetTop===this.offsetTop)return e===this._minimizedPosition?this._transitionToPosition(this._initialPosition):this._transitionToPosition(this._topPosition);const t=(this._topPosition-this._initialPosition)/2;e-this._initialPosition>=t?this._transitionToPosition(this._topPosition):this._transitionToPosition(this._initialPosition)}_transitionToPosition(e){this.classList.add("mdw-animating-open"),requestAnimationFrame(()=>{this._setPosition(e),this.addEventListener("transitionend",this.bound_onTransitionEnd)})}_cancelTransitions(){this.classList.remove("mdw-animating-open"),this.classList.remove("mdw-animating-close"),this.classList.remove("mdw-animating-scroll"),this.style.transitionDuration="",this.removeEventListener("transitionend",this.bound_onTransitionEnd),this.removeEventListener("transitionend",this.bound_onTransitionEndClose)}_onTransitionEnd(){this._initialOffsetTop=this.offsetTop,this._cancelTransitions()}_onTransitionEndClose(){this._cancelTransitions(),this.classList.remove("mdw-show"),this._helpers.removeBackdrop()}_onDrag(e){switch(e.state){case"start":this._cancelTransitions(),this._startPosition=this._currentPosition,this.classList.add("mdw-dragging"),a.lockPageScroll(),a.disableUserSelect();break;case"move":this._setPosition(this._startPosition-e.distance.y);break;case"end":this.classList.add("mdw-animating-scroll"),this._handleScrollEnd(e.velocity.y,e.direction.y),this.addEventListener("transitionend",this.bound_onTransitionEnd),this.classList.remove("mdw-dragging"),a.unlockPageScroll(),a.enableUserSelect()}}_handleScrollEnd(e,t){if(!this._isAtOrAboveTop&&e<-1.1)this._positionTop();else{if(!this._isAtOrAboveTop&&e>.7)return"modal"===this.type?this.close():this.minimize();{const n=Math.abs(e)/3;let i=this._scrollDistanceRemaining*n*-t;i>this._scrollDistanceRemaining&&(i=this._scrollDistanceRemaining);const s=i+this._currentPosition;if(this._isAtOrAboveTop)s<this._topPosition+80?this._positionTop():(this.style.transitionDuration=`${.5*n+.3}s`,this._setPosition(s));else{const e=(this._topPosition-this._initialPosition)/2;s-this._initialPosition>=e?this._positionTop():this._positionInitial()}}}}_setupOverScroll(){this.insertAdjacentHTML("beforeend",'<div class="mdw-sheet-bottom-over-scroll"></div>')}}),customElements.define("mdw-sheet-side",class extends g{constructor(){super(),this._useBackdrop=!this.hasAttribute("mdw-no-backdrop"),this.bound_onSwipe=this.onSwipe.bind(this),this.bound_routeChange=this.routeChange.bind(this)}connectedCallback(){this._isNavigationDrawer=this.classList.contains("mdw-navigation-drawer"),this._isNavigationDrawer&&document.body.classList.add("mdw-has-navigation-drawer"),a.isMobile&&this.setAttribute("mdw-modal",""),this._isNavigationDrawer&&this.isModal&&document.body.classList.add("mdw-navigation-drawer-modal"),this.isModal&&this._isNavigationDrawer&&!this.isHidden&&!this.classList.contains("mdw-show")?this.classList.add("mdw-hide"):this._isNavigationDrawer?document.body.classList.add("mdw-navigation-drawer-open"):this.isModal&&!this.isHidden&&this._useBackdrop&&(this._backdrop=a.addBackdrop(this,()=>this.close(),{sheet:!0})),this._isNavigationDrawer&&(window.addEventListener("hashchange",this.bound_routeChange),window.addEventListener("DOMContentLoaded",this.bound_routeChange))}disconnectedCallback(){this._backdrop&&this._backdrop.remove(),b(document.body,this.bound_onSwipe),this._isNavigationDrawer&&(window.removeEventListener("hashchange",this.bound_routeChange),window.removeEventListener("DOMContentLoaded",this.bound_routeChange))}get isModal(){return this.hasAttribute("mdw-modal")}get isHidden(){return this.classList.contains("mdw-hide")}get isLeft(){return this.classList.contains("mdw-left")||this.classList.contains("mdw-navigation-drawer")}set useBackdrop(e){this._useBackdrop=!!e}get path(){let e=window.location.hash.replace(/.*#/,"");return e.indexOf("?")>-1&&(e=e.split("?")[0]),"/"!==e.charAt(0)&&(e="/"+e),e}routeChange(){this.querySelectorAll(".mdw-current-link").forEach(e=>e.classList.remove("mdw-current-link"));let e=this.querySelectorAll(`[href="#${this.path}"]`);e&&0!==e.length||(e=this.querySelectorAll(`[alt-href="#${this.path}"]`)),e.forEach(e=>e.classList.add("mdw-current-link"))}open(){console.log(this.classList.contains("mdw-hide")),setTimeout(()=>{this.classList.remove("mdw-hide"),this.classList.add("mdw-show"),this.isModal&&(w(document.body,this.bound_onSwipe),this._useBackdrop&&(this._backdrop=a.addBackdrop(this,()=>this.close(),{sheet:!0}))),this._isNavigationDrawer&&document.body.classList.add("mdw-navigation-drawer-open")},10)}async close(){return new Promise(e=>{this.classList.remove("mdw-show"),this.classList.add("mdw-hide"),this._backdrop&&this._backdrop.remove(),b(document.body,this.bound_onSwipe),this._isNavigationDrawer&&document.body.classList.remove("mdw-navigation-drawer-open"),setTimeout(()=>{e()},200)})}toggle(){this.isHidden?this.open():this.close()}onSwipe(e){this.isLeft?-1===e.direction.x&&e.velocity.x<-.8&&this.close():1===e.direction.x&&e.velocity.x>.8&&this.close()}}),customElements.define("mdw-slider",class extends g{constructor(){super(),this.cloneTemplate(),this.bound_onMouseDown=this.onMouseDown.bind(this),this.bound_onMouseUp=this.onMouseUp.bind(this),this.bound_onMouseMove=this.onMouseMove.bind(this),this.bound_onMouseEnter=this.onMouseEnter.bind(this),this.bound_onMouseLeave=this.onMouseLeave.bind(this),this.bound_trackClick=this.trackClick.bind(this),this.bound_onDrag=this.onDrag.bind(this)}connectedCallback(){this.value=this.attrValue,this.thumbContainer.style.left=`${(this.attrValue-this.min)/this.range*this.offsetWidth}px`,this.notchContainer.style.marginLeft=`-${this.offsetWidth-(this.attrValue-this.min)/this.range*this.offsetWidth}px`,this.throttled_dispatchChange=a.rafThrottle(this.dispatchChange),c(this.thumb,this.bound_onDrag)}disconnectedCallback(){this.track.removeEventListener("click",this.bound_trackClick),h(this.thumb,this.bound_onDrag)}static get observedAttributes(){return["value","min","max","step"]}attributeChangedCallback(e,t,n){this[e]=n,["min","max","step"].includes(e)&&this.render()}get min(){return this.min_||0}set min(e){this.min_=parseFloat(e)}get max(){return this.max_||100}set max(e){this.max_=parseFloat(e)}get range(){return this.max-this.min}get step(){return this.step_}set step(e){this.step_=parseFloat(e)}get stepCount(){return this.step?Math.floor(this.range/this.step):0}get attrValue(){let e=parseFloat(this.getAttribute("value")||0);return e<this.min&&(e=this.min),e}get value(){const{width:e}=this.getBoundingClientRect(),t=(this.thumbContainer.style.left||"0px").replace("px","")/e,n=this.range;return this.value_=this.min+t*n,(""+this.step).includes(".")||(this.value_=parseInt(this.value_)),this.value_||0}set value(e){this.value_=parseFloat(e)}get thumb(){return this.shadowRoot.querySelector(".mdw-slider__thumb-hover")}get thumbContainer(){return this.thumbContainer_||(this.thumbContainer_=this.shadowRoot.querySelector(".mdw-slider__thumb-container")),this.thumbContainer_}get notchContainer(){return this.notchContainer_||(this.notchContainer_=this.shadowRoot.querySelector(".mdw-slider__notch-container")),this.notchContainer_}get track(){return this.shadowRoot.querySelector(".mdw-slider__track-container")}trackClick(e){const{left:t,width:n}=this.getBoundingClientRect();let i=e.layerX;e.clientX<t&&(i=0),i>n&&(i=n),this.thumbContainer.style.left=`${this.snap(i,n)}px`,this.notchContainer.style.marginLeft=`-${this.offsetWidth-this.snap(i,n)}px`,this.dispatchChange()}onDrag(e){switch(e.state){case"start":this.classList.add("mdw-pressed"),this.initialX_=parseInt((this.thumbContainer.style.left||"0px").replace("px",""));break;case"move":const{left:t,width:n}=this.getBoundingClientRect();let i=e.distance.x+this.initialX_;i<0&&(i=0),i>n&&(i=n),this.thumbContainer.style.left=`${this.snap(i,n)}px`,this.notchContainer.style.marginLeft=`-${this.offsetWidth-this.snap(i,n)}px`,this.throttled_dispatchChange();break;case"end":this.classList.remove("mdw-pressed")}}onMouseDown(e){this.classList.add("mdw-pressed"),document.addEventListener("mouseup",this.bound_onMouseUp),document.addEventListener("mousemove",this.bound_onMouseMove)}onMouseUp(e){this.classList.remove("mdw-pressed"),document.removeEventListener("mouseup",this.bound_onMouseUp),document.removeEventListener("mousemove",this.bound_onMouseMove)}onMouseMove(e){const{left:t,width:n}=this.getBoundingClientRect();let i=e.layerX;e.clientX<t&&(i=0),i>n&&(i=n),this.thumbContainer.style.left=`${this.snap(i,n)}px`,this.notchContainer.style.marginLeft=`-${this.offsetWidth-this.snap(i,n)}px`,this.throttled_dispatchChange()}onMouseEnter(e){this.classList.add("mdw-hover"),this.thumb.addEventListener("mouseleave",this.bound_onMouseLeave)}onMouseLeave(e){this.classList.remove("mdw-hover"),this.thumb.removeEventListener("mouseleave",this.bound_onMouseLeave)}snap(e,t){if(!this.step)return e;const n=e/t,i=this.range,s=n*i;return(s-s%this.step)/i*t}dispatchChange(){this.dispatchEvent(new Event("change",this))}template(){return`\n      <div class="mdw-slider__track-container">\n        <div class="mdw-slider__track"></div>\n\n        <div class="mdw-slider__notch-container">\n          <div class="mdw-slider__notch-pre-container">\n            ${[...new Array(this.stepCount)].map(e=>'<div class="mdw-slider__notch"></div>').join("\n")}\n          </div>\n\n          <div class="mdw-slider__notch-post-container">\n            ${[...new Array(this.stepCount)].map(e=>'<div class="mdw-slider__notch"></div>').join("\n")}\n          </div>\n        </div>\n      </div>\n      <div class="mdw-slider__thumb-container">\n        <div class="mdw-slider__thumb"></div>\n        <div class="mdw-slider__thumb-hover"></div>\n      </div>\n    `}styles(){return"\n      .mdw-slider__track-container {\n        position: absolute;\n        top: 50%;\n        width: 100%;\n        height: 10px;\n        margin-top: -6px;\n        overflow: hidden;\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      .mdw-slider__track {\n        position: absolute;\n        width: 100%;\n        height: 2px;\n        top: 50%;\n        user-select: none;\n        -webkit-user-select: none;\n        /* background-color: var(--mdw-theme-secondary); */\n      }\n\n      /* :host(.mdw-primary) .mdw-slider__track {\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-error) .mdw-slider__track {\n        background-color: var(--mdw-theme-error);\n      } */\n\n\n      .mdw-slider__thumb-container {\n        position: absolute;\n        top: 50%;\n        left: 0;\n        user-select: none;\n        -webkit-user-select: none;\n        z-index: 2;\n      }\n\n      .mdw-slider__thumb {\n        box-sizing: border-box;\n        width: 12px;\n        height: 12px;\n        border-radius: 50%;\n        margin-top: -50%;\n        z-index: 2;\n        background-color: var(--mdw-theme-secondary);\n        cursor: pointer;\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      :host(.mdw-primary) .mdw-slider__thumb {\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-error) .mdw-slider__thumb {\n        background-color: var(--mdw-theme-error);\n      }\n\n      .mdw-slider__thumb-hover {\n        position: absolute;\n        box-sizing: border-box;\n        top: -12px;\n        left: -6px;\n        width: 24px;\n        height: 24px;\n        border-radius: 50%;\n        transform-origin: center center;\n        transition: opacity .1s ease-out,fill .1s ease-out,\n                    transform .1s ease-out,fill .1s ease-out;\n        opacity: 0;\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.16);\n        cursor: pointer;\n        user-select: none;\n      }\n\n      :host(.mdw-primary) .mdw-slider__thumb-hover {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n\n      :host(.mdw-error) .mdw-slider__thumb-hover {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.16);\n      }\n\n      :host(.mdw-hover) .mdw-slider__thumb-hover {\n        opacity: 1;\n      }\n\n      :host(.mdw-pressed) .mdw-slider__thumb-hover {\n        transform: scale(1.8);\n      }\n\n\n\n\n      /* --- notches --- */\n\n      .mdw-slider__notch-container {\n        display: flex;\n        width: 200%;\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      .mdw-slider__notch-pre-container {\n        width: 100%;\n        height: 2px;\n        display: flex;\n        flex-direction: row;\n        margin-top: 5px;\n        overflow: hidden;\n        z-index: 1;\n        background-color: var(--mdw-theme-secondary);\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      :host(.mdw-primary) .mdw-slider__notch-pre-container {\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-error) .mdw-slider__notch-pre-container {\n        background-color: var(--mdw-theme-error);\n      }\n\n      .mdw-slider__notch-pre-container .mdw-slider__notch {\n        height: 2px;\n        flex: 1;\n        border-left: 3px solid rgba(255, 255, 255, 0.6);\n      }\n\n      .mdw-slider__notch-post-container {\n        width: 100%;\n        height: 2px;\n        display: flex;\n        flex-direction: row;\n        margin-top: 5px;\n        overflow: hidden;\n        z-index: 1;\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.5);\n        user-select: none;\n        -webkit-user-select: none;\n      }\n\n      :host(.mdw-primary) .mdw-slider__notch-post-container {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.5);\n      }\n\n      :host(.mdw-error) .mdw-slider__notch-post-container {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.5);\n      }\n\n      .mdw-slider__notch-post-container .mdw-slider__notch {\n        height: 2px;\n        flex: 1;\n        border-left: 3px solid rgba(0, 0, 0, 0.6);\n      }\n    "}});const L=new class{constructor(){this.queue=[]}add(e){this.queue.push({el:e}),this.handleQueue()}remove(e,t){this.current&&this.current.el===e?e._close(t):this.queue=this.queue.filter(t=>t.el!==e)}handleQueue(){0!==this.queue.length&&(this.current||(this.current=this.queue.shift(),this.current.el._open(),this.current.el.addEventListener("close",()=>{this.current=void 0,setTimeout(()=>{this.handleQueue()},300)})))}show({message:e,actionLabel:t,position:n}){return new Promise(i=>{const s=this.uid(),o=this.template({id:s,message:e,actionLabel:t});this.topLevelElement.insertAdjacentHTML("beforeend",o);const r=document.querySelector(`#${s}`),a=e=>{i(e.detail.ok),r.removeEventListener("close",a),r.remove()};r.addEventListener("close",a),n&&r.setPosition(n),r.show()})}get topLevelElement(){let e=document.body.querySelector("mdw-content");return e||(e=document.body.querySelector("mdw-body"),e||document.body)}uid(){return`snackbar_${parseInt(99999*Math.random())}`}template({id:e,message:t,actionLabel:n}){return`\n      <mdw-snackbar id="${e}">\n        <mdw-panel>\n          <mdw-snackbar-container>\n            <mdw-snackbar-content>${t}</mdw-snackbar-content>\n            <mdw-snackbar-actions>\n              ${n?`<mdw-button class="mdw-action-button">${n}</mdw-button>`:""}\n              <mdw-button onclick="${e}.close(true)" class="mdw-close-button mdw-icon">\n                <mdw-icon>close</mdw-icon>\n              </mdw-button>\n            </mdw-snackbar-actions>\n          </mdw-snackbar-container>\n        </mdw-panel>\n      </mdw-snackbar>\n    `}};window.MDWSnackbar=L;var C=L;customElements.define("mdw-snackbar",class extends g{constructor(){super(),this.bound_onPanelClose=this.onPanelClose.bind(this),this.panelId=`${this.getAttribute("id")}_panel`}connectedCallback(){this.querySelector("mdw-panel").setAttribute("id",`${this.panelId}`),this.hasBckdrop=!0,this.panel.clickOutsideClose=!1}disconnectedCallback(){this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose)}get panel(){return document.querySelector(`#${this.panelId}`)}get position(){return this.position_||"inner-bottom inner-left"}setPosition(e){const t=e.split(" ");this.position_=`${t[0]||"top"} ${t[1]||"left"}`,this.panel.setPosition(this.position)}show(){C.add(this)}close(e){C.remove(this,e)}_open(){this.panel.hoistToBody(this.parentNode),this.panel.setPosition(this.position),this.panel.autoPosition(),this.panel.open(),this.panel.addEventListener("MDWPanel:closed",this.bound_onPanelClose),this.autoCancelTimeout=setTimeout(()=>{this.close()},3e3)}_close(e){this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose),this.panel.close(),this.dispatchClose(e),clearTimeout(this.autoCancelTimeout),setTimeout(()=>{this.panel.remove()},200)}onPanelClose(){this.panel.removeEventListener("MDWPanel:closed",this.bound_onPanelClose)}dispatchClose(e=!1){this.dispatchEvent(new CustomEvent("close",{detail:{ok:e}}))}}),customElements.define("mdw-switch",class extends g{constructor(){super(),this.bound_onInputChange=this.onInputChange.bind(this),this.cloneTemplate()}connectedCallback(){this.input.addEventListener("change",this.bound_onInputChange),this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:[this.input],radius:20,centered:!0})}disconnectedCallback(){this.input.addEventListener("click",this.bound_click),this.ripple.destroy()}static get observedAttributes(){return["checked","disabled"]}attributeChangedCallback(e,t,n){this[e]=n}get input(){return this.input_||(this.input_=this.shadowRoot.querySelector("input")),this.input_}get checked(){return this.input.checked}set checked(e){""===e&&(e=!0),this.input.checked=e,this.updateCheckedClass()}set disabled(e){(e=!!e||""===e)?this.input.setAttribute("disabled","disabled"):this.input.removeAttribute("disabled")}updateCheckedClass(){this.checked?this.classList.add("checked"):this.classList.remove("checked")}dispatchChange(){this.dispatchEvent(new Event("change"))}onInputChange(e){this.updateCheckedClass(),this.dispatchChange()}template(){return'\n      <div class="mdw-track"></div>\n      <div class="mdw-thumb-underlay">\n        <div class="mdw-thumb">\n          <input type="checkbox" role="switch">\n          <div class="mdw-ripple mdw-switch-ripple"></div>\n        </div>\n      </div>\n    '}styles(){return'\n      .mdw-track {\n        box-sizing: border-box;\n        width: 32px;\n        height: 14px;\n        border: 1px solid;\n        border-radius: 7px;\n        opacity: .38;\n        transition: opacity 90ms cubic-bezier(.4,0,.2,1),\n                    background-color 90ms cubic-bezier(.4,0,.2,1),\n                    border-color 90ms cubic-bezier(.4,0,.2,1);\n      }\n\n      :host(:not(.checked)) .mdw-track {\n        background-color: rgba(var(--mdw-theme-on-background--rgb), 0.7);\n        border-color: rgba(var(--mdw-theme-on-background--rgb), 0.7);\n      }\n\n      :host(.checked) .mdw-track {\n        background-color: var(--mdw-theme-secondary);\n        border-color: var(--mdw-theme-secondary);\n        opacity: .54;\n      }\n\n      :host(.checked.mdw-primary) .mdw-track {\n        background-color: var(--mdw-theme-primary);\n        border-color: var(--mdw-theme-primary);\n      }\n\n      :host(.checked.mdw-error) .mdw-track {\n        background-color: var(--mdw-theme-error);\n        border-color: var(--mdw-theme-error);\n      }\n\n\n\n      /* --- thumb underlay --- */\n\n      .mdw-thumb-underlay {\n        -webkit-tap-highlight-color: rgba(0,0,0,0);\n        display: flex;\n        position: absolute;\n        will-change: transform,opacity;\n        left: -18px;\n        right: auto;\n        top: -17px;\n        align-items: center;\n        justify-content: center;\n        width: 48px;\n        height: 48px;\n        transform: translateX(0);\n        transition: transform 90ms cubic-bezier(.4,0,.2,1),\n                    background-color 90ms cubic-bezier(.4,0,.2,1),\n                    border-color 90ms cubic-bezier(.4,0,.2,1);\n      }\n\n      :host(.checked) .mdw-thumb-underlay {\n        transform: translateX(20px);\n      }\n\n      .mdw-thumb-underlay:after,\n      .mdw-thumb-underlay:before {\n        position: absolute;\n        border-radius: 50%;\n        opacity: 0;\n        pointer-events: none;\n        content: "";\n        background-color: var(--mdw-theme-secondary);\n      }\n\n      :host(.mdw-primary) .mdw-thumb-underlay:after,\n      :host(.mdw-primary) .mdw-thumb-underlay:before {\n        background-color: var(--mdw-theme-primary);\n      }\n\n      :host(.mdw-error) .mdw-thumb-underlay:after,\n      :host(.mdw-error) .mdw-thumb-underlay:before {\n        background-color: var(--mdw-theme-error);\n      }\n\n      .mdw-thumb-underlay:before {\n        transition: opacity 15ms linear,background-color 15ms linear;\n        z-index: 1;\n      }\n\n\n\n      /* --- thumb --- */\n\n      .mdw-thumb {\n        box-shadow: 0 3px 1px -2px rgba(0,0,0,.2),\n                    0 2px 2px 0 rgba(0,0,0,.14),\n                    0 1px 5px 0 rgba(0,0,0,.12);\n        box-sizing: border-box;\n        width: 20px;\n        height: 20px;\n        border: 10px solid;\n        border-radius: 50%;\n        pointer-events: none;\n        z-index: 1;\n      }\n\n      :host(.checked) .mdw-thumb {\n        background-color: var(--mdw-theme-secondary);\n        border-color: var(--mdw-theme-secondary);\n      }\n\n      :host(.checked.mdw-primary) .mdw-thumb {\n        background-color: var(--mdw-theme-primary);\n        border-color: var(--mdw-theme-primary);\n      }\n\n      :host(.checked.mdw-error) .mdw-thumb {\n        background-color: var(--mdw-theme-error);\n        border-color: var(--mdw-theme-error);\n      }\n\n      :host(:not(.checked)) .mdw-thumb {\n        background-color: #fff;\n        border-color: #fff;\n      }\n\n\n      /* --- input --- */\n\n      input {\n        left: 0;\n        right: auto;\n        position: absolute;\n        top: 0;\n        width: 68px;\n        height: 48px;\n        margin: 0;\n        opacity: 0;\n        cursor: pointer;\n        pointer-events: auto;\n      }\n\n      :host(.checked) input {\n        transform: translateX(-20px);\n      }\n\n\n\n      /* --- ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n        background-color: rgba(var(--mdw-theme-secondary--rgb), 0.16);\n      }\n\n      :host(.mdw-primary) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n\n      :host(.mdw-error) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-error--rgb), 0.16);\n      }\n\n      .mdw-switch-ripple {\n        position: absolute;\n        left: 0;\n        top: 0;\n        right: 0;\n        bottom: 0;\n        border-radius: 50%;\n        z-index: 1;\n        pointer-events: none;\n      }\n    '}}),customElements.define("mdw-tab-body",class extends g{constructor(){super(),this.cloneTemplate()}connectedCallback(){this.parentNode.registerBody(this)}disconnectedCallback(){this.parentNode.unregisterBody(this)}addSlot(){this.shadowRoot.querySelector("mdw-tab-body-content").insertAdjacentHTML("beforeend","<slot></slot>")}removeSlot(){this.shadowRoot.querySelector("slot").remove()}activate(){this.addSlot(),this.classList.add("mdw-active")}deactivate(){this.removeSlot(),this.classList.remove("mdw-active")}template(){return"\n      <mdw-tab-body-content>\n        \x3c!-- slot is added dynamicly --\x3e\n      </mdw-tab-body-content>\n    "}styles(){return"\n      mdw-tab-body-content {\n        height: 100%;\n        overflow: auto;\n      }\n    "}}),customElements.define("mdw-tab-button",class extends g{constructor(){super(),this.bound_click=this.click.bind(this),this.cloneTemplate()}connectedCallback(){this.ripple=new f({element:this.shadowRoot.querySelector(".mdw-ripple"),triggerElement:this}),this.parentNode.registerTab(this),this.addEventListener("click",this.bound_click)}disconnectedCallback(){this.ripple.destroy(),this.parentNode.unregisterTab(this),this.removeEventListener("click",this.bound_click)}get indicator(){return this.shadowRoot.querySelector(".mdw-tab-button-indicator__content")}click(e){this.parentNode.tabClick(this)}activate(){clearTimeout(this._animationTimer),this.indicator.style.transform="",this._runNextAnimationFrame(()=>{this._animationTimer=setTimeout(()=>{this.classList.add("mdw-active")},180)})}deactivate(e){clearTimeout(this._animationTimer),this.indicator.style.transform=`translateX(${e.toString()}px)`,this._animationTimer=setTimeout(()=>{this.classList.remove("mdw-active")},200)}_runNextAnimationFrame(e){cancelAnimationFrame(this._animationFrame),this._animationFrame=requestAnimationFrame(()=>{this._animationFrame=0,clearTimeout(this._animationTimer),this._animationTimer=setTimeout(e,0)})}template(){return'\n      <span class="text"><slot></slot></span>\n      <span class="mdw-tab-button-indicator">\n        <span class="mdw-tab-button-indicator__content mdw-tab-button-indicator__content--underline"></span>\n      </span>\n      <div class="mdw-ripple mdw-tab-button-ripple"></div>\n    '}styles(){return"\n      :host(.mdw-show-spinner) span.text {\n        opacity: 0;\n      }\n\n      /* Add this to button or creat a new componenet mdw-tab */\n      .mdw-tab-button-indicator {\n        display: flex;\n        position: absolute;\n        top: 0;\n        left: 0;\n        width: 100%;\n        height: 100%;\n        pointer-events: none;\n        z-index: 1;\n      }\n\n      :host(.mdw-active) .mdw-tab-button-indicator .mdw-tab-button-indicator__content {\n        transition: transform .2s cubic-bezier(.4,0,.2,1);\n      }\n\n      .mdw-tab-button-indicator__content {\n        opacity: 0;\n        transform-origin: left;\n      }\n\n      :host(.mdw-active) .mdw-tab-button-indicator__content {\n        opacity: 1;\n        transform: translateX(0);\n      }\n\n      .mdw-tab-button-indicator .mdw-tab-button-indicator__content--underline {\n        align-self: flex-end;\n        width: 100%;\n        background-color: var(--mdw-theme-primary);\n        height: 2px;\n      }\n\n\n      /* --- Ripple --- */\n\n      .mdw-ripple {\n        overflow: hidden;\n      }\n\n      .mdw-ripple.mdw-ripple-unbounded {\n        overflow: visible;\n      }\n\n      .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-background--rgb), 0.16);\n        position: absolute;\n        border-radius: 50%;\n        pointer-events: none;\n        transition: opacity, transform 0ms cubic-bezier(0, 0, 0.2, 1);\n        transform: scale(0);\n      }\n\n      .mdw-tab-button-ripple,\n      .mdw-tab-button-focus-overlay {\n        border-radius: inherit;\n        position: absolute;\n        top: 0;\n        left: 0;\n        bottom: 0;\n        right: 0;\n        pointer-events: none;\n      }\n\n      :host(.mdw-active) .mdw-ripple-element {\n        background-color: rgba(var(--mdw-theme-primary--rgb), 0.16);\n      }\n    "}}),customElements.define("mdw-tabs-bar",class extends g{constructor(){super(),this._activeTab=0,this.tabIdCounter=0,this._contentElements=[],this.cloneTemplate()}registerTab(e){e.setAttribute("tab-id",this.tabIdCounter),0===this.tabIdCounter&&(this.activeTab=e,e.activate()),this.tabIdCounter++}unregisterTab(e){}registerContent(e){this._contentElements.push(e),e.changeTab(this.activeTab.getAttribute("tab-id"))}unregisterContent(e){this._contentElements=this._contentElements.filter(t=>t!=e)}tabClick(e){const t=parseInt(e.getBoundingClientRect().x-this.activeTab.getBoundingClientRect().x);this.activeTab.deactivate(t),this.activeTab=e,this.activeTab.activate(),this._contentElements.forEach(e=>e.changeTab(this.activeTab.getAttribute("tab-id")))}get activeTab(){return this._activeTab}set activeTab(e){this._activeTab=e}get internalStylesFile(){return"./internal.css"}template(){return"\n      <mdw-tabs-bar-scroller>\n        <mdw-tabs-bar-scroller-area>\n          <mdw-tabs-bar-scroller-content>\n            <slot></slot>\n          </mdw-tabs-bar-scroller-content>\n        </mdw-tabs-bar-scroller-area>\n      </mdw-tabs-bar-scroller>\n    "}styles(){return"\n      mdw-tabs-bar-scroller {\n        display: block;\n        overflow-y: hidden;\n      }\n\n      mdw-tabs-bar-scroller-area {\n        display: flex;\n        /* overflow-x: scroll; */\n      }\n\n      mdw-tabs-bar-scroller-content {\n        position: relative;\n        display: flex;\n        flex: 1 0 auto;\n        transform: none;\n        will-change: transform;\n      }\n\n      ::slotted(mdw-button) {\n        flex: 1 0 auto;\n      }\n    "}}),customElements.define("mdw-tabs-content",class extends g{constructor(){super(),this._bodies=[]}connectedCallback(){this.tabsBar.registerContent(this)}disconnectedCallback(){this.tabsBar&&this.tabsBar.unregisterContent(this)}get tabsBar(){return document.body.querySelector(`mdw-tabs-bar#${this.getAttribute("tabs-id")}`)}registerBody(e){this._bodies.push(e),void 0!==this._wiatForBodyActiveId&&this._bodies.length===this._wiatForBodyActiveId+1&&(this._activeBody=e,e.activate(),this._wiatForBodyActiveId=void 0)}unregisterBody(e){this._bodies=this._bodies.filter(t=>t!=e)}changeTab(e){this._bodies.length?(this._activeBody&&this._activeBody.deactivate(),this._activeBody=this._bodies[e],this._activeBody.activate()):this._wiatForBodyActiveId=parseInt(e)}}),customElements.define("mdw-textfield",class extends g{constructor(){super(),this.classList.add("mdw-no-animation"),this.bound_onFocus=this.onFocus.bind(this),this.bound_onBlur=this.onBlur.bind(this),this.bound_onInput=this.onInput.bind(this)}connectedCallback(){this.compose(),this.checkForValue(),requestAnimationFrame(()=>{this.classList.remove("mdw-no-animation")}),this.input.addEventListener("focus",this.bound_onFocus),this.input.addEventListener("blur",this.bound_onBlur),this.input.addEventListener("input",this.bound_onInput),this.classList.toggle("mdw-invalid",!this.valid)}disconnectedCallback(){this.input.removeEventListener("focus",this.bound_onFocus),this.input.removeEventListener("blur",this.bound_onBlur),this.input.removeEventListener("input",this.bound_onInput)}compose(){this.classList.add("mdw-upgraded"),this.isTextarea()&&this.classList.add("mdw-textarea"),this.outlined&&(this.insertAdjacentHTML("beforeend",this.outlinedHTML),this.setNotchWidth()),this.querySelector(".mdw-line-ripple")||this.insertAdjacentHTML("beforeend",this.lineRippleHTML),this.isTrailingIcon()&&this.classList.add("mdw-trailing-icon")}checkForValue(){this.classList.toggle("not-empty",!!this.input.value.length)}onFocus(){this.setNotchWidth()}onBlur(){this.classList.toggle("not-empty",!!this.input.value.length),this.classList.toggle("mdw-invalid",!this.valid)}onInput(){this.classList.toggle("mdw-invalid",!this.valid)}setNotchWidth(){this.outlined&&(this.notch.style.width=this.labelWidth+"px")}isTrailingIcon(){return!!this.iconElement&&[...this.children].indexOf(this.iconElement)>1}isTextarea(){return!!this.querySelector("textarea")}clear(){this.input.value=""}get valid(){return this.input.validity.valid}get outlined(){return[].slice.apply(this.classList||[]).includes("mdw-outlined")}get input(){return this.inputType_||(this.inputType_=this.querySelector("input")?"input":"textarea"),this.querySelector(this.inputType_)}get notch(){return this.querySelector(".mdw-outlined-notch")}get label(){return this.querySelector("label")}get labelWidth(){return.95*this.label.offsetWidth}get helperTextElement(){return this.querySelector("mdw-textfield-helper")}get iconElement(){return this.querySelector("mdw-icon")}get outlinedHTML(){return'\n      <div class="mdw-outlined-border-container">\n        <div class="mdw-outlined-leading"></div>\n        <div class="mdw-outlined-notch"></div>\n        <div class="mdw-outlined-trailing"></div>\n      </div>\n    '}get lineRippleHTML(){return'<div class="mdw-line-ripple"></div>'}}),customElements.define("mdw-top-app-bar",class extends g{constructor(){super(),this._throttledScrollHandler=a.rafThrottle(this._scrollHandler),this._throttledResizeHandler=a.rafThrottle(this._resizeHandler),this.bound_throttledScrollHandler=this._throttledScrollHandler.bind(this),this.bound_throttledResizeHandler=this._throttledResizeHandler.bind(this),document.body.classList.add("mdw-has-top-app-bar")}connectedCallback(){this._isProminent=this.hasAttribute("mdw-prominent"),this._isFixed=this.hasAttribute("mdw-fixed"),this._isShrink=this.hasAttribute("mdw-shrink"),this.parentNode&&"HEADER"===this.parentNode.nodeName&&(this.parentNode.classList.add("mdw-top-app-bar"),this._isProminent&&this.parentNode.classList.add("mdw-prominent"),this._isShrink&&this.parentNode.classList.add("mdw-shrink"),this._isFixed&&this.parentNode.classList.add("mdw-fixed")),this._isShrink&&(this._animationElements=[...this.querySelectorAll("[mdw-animation-property]")||[]].map(e=>{const t=parseFloat(e.getAttribute("mdw-animation-start")||0),n=parseFloat(e.getAttribute("mdw-animation-end")||0),i=e.getAttribute("mdw-animation-property").split(":");return{element:e,property:i[0],valueWrapper:i[1]||"",start:t,end:n,range:Math.abs(t-n)}}),this._scrollTarget=this._getScrollTarget(),this._lastScrollPosition=this._getViewportScrollY(),this._topAppBarHeight=this.clientHeight+6,this._scrollHandler(),this._createObserver(),this._scrollTarget.addEventListener("scroll",this.bound_throttledScrollHandler)),(this._isFixed||this._isShrink)&&(this._page=document.querySelector("mdw-page"),this._page&&(this.style.width=`${this._page.offsetWidth}px`)),window.addEventListener("resize",this.bound_throttledResizeHandler)}disconnectedCallback(){this._observer&&(this._observer.unobserve(this),this._observer.disconnect(),this._observer=void 0),this._scrollTarget&&this._scrollTarget.removeEventListener("scroll",this.bound_throttledScrollHandler),window.removeEventListener("resize",this.throttledResizeHandler)}notContextual(){this.removeAttribute("mdw-contextual")}contextual(){this.setAttribute("mdw-contextual","")}_getAllFixedSections(){return this.querySelectorAll("section[mdw-fixed]")}_getScrollTarget(){if("HEADER"===this.parentNode.nodeName){const e=this.parentNode.parentNode.querySelector("header + mdw-content:not([mdw-no-scroll])");if(e)return e}const e=document.querySelector("mdw-page-content");if(e&&e.contains(this))return e;const t=document.querySelector("mdw-content");return t&&t.contains(this)?t:window}_getViewportScrollY(){return this._scrollTarget[this._scrollTarget===window?"pageYOffset":"scrollTop"]}_scrollHandler(){const e=Math.max(this._getViewportScrollY(),0);let t;if(this._isProminent){const n=this._topAppBarHeight/2;t=e<=n?e:n}else t=e<=this._topAppBarHeight?e:this._topAppBarHeight;this.style.transform=`translateY(-${t}px)`,this._getAllFixedSections().forEach(e=>{e.style.transform=`translateY(${t}px)`})}_resizeHandler(){(this._isFixed||this._isShrink)&&(this._page=document.querySelector("mdw-page"),this._page&&(this.style.width=`${this._page.offsetWidth}px`))}_createObserver(){0!==this._animationElements.length&&(this._observer=new IntersectionObserver(this._handleIntersect.bind(this),{root:null,rootMargin:"0px",threshold:this._buildThresholdList()}),this._observer.observe(this))}_buildThresholdList(){let e=[];for(let t=1;t<=68;t++){let n=t/68;e.push(n)}return this._intersectionThresholds=e,this._intersectionThresholds}_handleIntersect(e,t){e.forEach(e=>{let t;this._isProminent&&(t=-e.boundingClientRect.top/(this._topAppBarHeight/2)),this._animationElements.forEach(e=>this._animationValue(t,e))})}_animationValue(e,{element:t,property:n,valueWrapper:i,start:s,end:o,range:r}){t.style[n]=i.replace("#",s-e*r)}}),customElements.define("mdw-bound-property",class extends g{constructor(){super(),this._property=this.innerHTML,this.innerHTML=""}connectedCallback(){const e=this;this._value=window.activePage[this._property],Object.defineProperty(window.activePage,this._property,{configurable:!0,enumerable:!0,get:()=>e._value,set(t){e._value=t,e.innerHTML=t}})}});const T=new class{constructor(){this._loadedTemplates={}}async get(e,t){if(this.isString(e))return new Function(`return \`${e}\`;`).call(t);if(!this.isUrl(e))throw Error("Template must be a string or url");const n=await this.loadHtml(e);return new Function(`return \`${n}\`;`).call(t)}async loadHtml(e){if(this._loadedTemplates[e])return this._loadedTemplates[e];const t=await fetch(e),n=await t.text();return this._loadedTemplates[e]=n,n}isString(e){return!this.isUrl(e)&&("string"==typeof e&&e.includes("<"))}isUrl(e){return!!e.match(/.html$/)}};window.MDWTemplate=T;var S=T;customElements.define("mdw-template",class extends g{constructor(){super();const e=this.url;e&&S.get(e,window.activePage).then(e=>{console.log(e),this.insertAdjacentHTML("beforeend",e)})}get url(){return this.getAttribute("mdw-url")}});const A=["panel","sheetBottom","sheetSide"],P=["scale","height"],M=["top","center"];class R{constructor({id:e,component:t,template:n,animation:i}){this._id=e,this._component=t,this._template=n,this._animation=i}get id(){return this._id}get component(){return this._component}get element(){return this._element}async open(){const e=document.querySelector("mdw-page");switch(e?e.insertAdjacentHTML("afterEnd",this._template):document.body.insertAdjacentHTML("beforeend",this._template),this._element=document.querySelector(`#${this.id}`),this.component){case"panel":this._element.setAnimation(this._animation),this.element.open();break;case"sheetBottom":case"sheetSide":this.element.open()}window._activeSurface&&await window._activeSurface.close(),window._activeSurface=this}async close(){switch(this.component){case"panel":case"sheetBottom":case"sheetSide":await this.element.close()}this.element.remove(),window._activeSurface=void 0}}const D=new class{constructor(){this._defaultMobileComponent="panel",this._defaultDesktopComponent="sheetSide"}setDefaultMobileComponent(e){this._validateComponent(e),this._defaultMobileComponent=e}setDefaultDesktopComponent(e){this._validateComponent(e),this._defaultDesktopComponent=e}async open({template:e,templateData:t,animation:n,animationTarget:i,component:s,mobileComponent:o,desktopComponent:r}){(await this.create({template:e,templateData:t,animation:n,animationTarget:i,component:s,mobileComponent:o,desktopComponent:r})).open()}async create({template:e,templateData:t,animation:n,animationTarget:i,component:s,mobileComponent:o,desktopComponent:r}){s||(s=this._autoSelectComponent(o,r)),this._validateComponent(s),"panel"===s&&(n||(n=this._autoSelectAnimation(i)),this._validateAnimation(n));const d=a.uid("surface"),l=S.isString(e)?e:await S.get(e,t);let c;switch(s){case"panel":c=this._buildPanel({id:d,animation:n,templateString:l});break;case"sheetBottom":c=this._buildSheetBottom({id:d,templateString:l});break;case"sheetSide":c=this._buildSheetSide({id:d,templateString:l})}return new R({id:d,component:s,template:c,animation:n})}close(){window._activeSurface&&window._activeSurface.close()}_autoSelectComponent(e,t){return a.isMobile?e||this._defaultMobileComponent:t||this._defaultDesktopComponent}_validateComponent(e){if(!A.includes(e))throw Error(`type must be one of these: ${A.join(", ")}`)}_autoSelectAnimation(e){return{type:"height",origin:"center",fullscreen:!0,target:e}}_validateAnimation(e){if(!P.includes(e.type))throw Error(`animation.type must be one of these: ${P.join(", ")}`);if(e.origin&&!M.includes(e.origin))throw Error(`animation.type must be one of these: ${M.join(", ")} or not defined`)}_buildPanel({id:e,templateString:t}){return`\n      <mdw-panel id="${e}">\n        ${t}\n      </mdw-panel>\n    `}_buildSheetBottom({id:e,templateString:t}){return`\n      <mdw-sheet-bottom id="${e}">\n        ${t}\n      </mdw-sheet-bottom>\n    `}_buildSheetSide({id:e,templateString:t}){return`\n      <mdw-sheet-side id="${e}" class="mdw-hide" mdw-modal mdw-no-backdrop>\n        ${t}\n      </mdw-sheet-side>\n    `}};window.MDWSurface=D;var H=D}]);
