+++ src-docs/pages/documentation/page-layout/page.html
@@ -1,0 +1,100 @@
+<article class="page-article">
+  <h1 class="article-title">Page Layout</h1>
+  <h2 class="article-subtitle">Standard application layouts. All these layouts will automatically change correctly for mobile devices</h2>
+
+
+  <div class="links">
+    <div class="eyebrow">contents</div>
+    <anchor-link selector="#one" offset="64">Layout 1</anchor-link>
+    <anchor-link selector="#two" offset="64">Layout 2</anchor-link>
+    <anchor-link selector="#three" offset="64">Layout 3</anchor-link>
+    <anchor-link selector="#four" offset="64">Layout 4</anchor-link>
+  </div>
+
+  <mdw-row mdw-flex-position="space-around" class="mdw-flex-wrap">
+    <div class="mdw-flex">
+      <h6>One</h6>
+      <div style="margin-top: -40px; margin-bottom: 20px; height: 38px">Top bar covers the entire length of the screen</div>
+      <img src="./layouts/one.png" width="398" height="312" class="mdw-wrap"/>
+    </div>
+    <div class="mdw-flex">
+      <h6>Two</h6>
+      <div style="margin-top: -40px; margin-bottom: 20px; height: 38px">Navigation covers the full height of the view. The top bar covers the rest</div>
+      <img src="./layouts/two.png" width="398" height="312" class="mdw-wrap"/>
+    </div>
+  </mdw-row>
+  <mdw-row mdw-flex-position="space-around" class="mdw-flex-wrap">
+    <div class="mdw-flex">
+      <h6>Three</h6>
+      <div style="margin-top: -40px; margin-bottom: 20px; height: 38px">Side sheet covers the full height of the view. The top bar covers the rest</div>
+      <img src="./layouts/three.png" width="398" height="312" />
+    </div>
+    <div class="mdw-flex">
+      <h6>Four</h6>
+      <div style="margin-top: -40px; margin-bottom: 20px; height: 38px">The navigation and the side sheet cover the full height of the view. The top bar only covers the page content</div>
+      <img src="./layouts/four.png" width="398" height="312" />
+    </div>
+  </mdw-row>
+
+  <section>
+    <mdw-card id="one">
+      <div class="mdw-card__content">
+        <h6>One</h6>
+        <p>Top bar covers the entire length of the screen</p>
+      </div>
+
+      <div class="mdw-card__content--no-padding">
+        <monaco-editor id="editor-1" language="html"></monaco-editor>
+      </div>
+      
+      <div class="mdw-card__content">
+        <iframe style="width: 100%; height: 400px; border: 1px solid #ddd" src="layouts/one.html"></iframe>
+      </div>
+    </mdw-card>
+
+    <mdw-card id="two">
+      <div class="mdw-card__content">
+        <h6>Two</h6>
+        <p>Navigation covers the full height of the view. The top bar covers the rest</p>
+      </div>
+    
+      <div class="mdw-card__content--no-padding">
+        <monaco-editor id="editor-2" language="html"></monaco-editor>
+      </div>
+    
+      <div class="mdw-card__content">
+        <iframe style="width: 100%; height: 400px; border: 1px solid #ddd" src="layouts/two.html"></iframe>
+      </div>
+    </mdw-card>
+
+    <mdw-card id="three">
+      <div class="mdw-card__content">
+        <h6>Three</h6>
+        <p>Side sheet covers the full height of the view. The top bar covers the rest</p>
+      </div>
+    
+      <div class="mdw-card__content--no-padding">
+        <monaco-editor id="editor-3" language="html"></monaco-editor>
+      </div>
+    
+      <div class="mdw-card__content">
+        <iframe style="width: 100%; height: 400px; border: 1px solid #ddd" src="layouts/three.html"></iframe>
+      </div>
+    </mdw-card>
+
+    <mdw-card id="four">
+      <div class="mdw-card__content">
+        <h6>Four</h6>
+        <p>The navigation and the side sheet cover the full height of the view. The top bar only covers the page content</p>
+      </div>
+    
+      <div class="mdw-card__content--no-padding">
+        <monaco-editor id="editor-4" language="html"></monaco-editor>
+      </div>
+    
+      <div class="mdw-card__content">
+        <iframe style="width: 100%; height: 400px; border: 1px solid #ddd" src="layouts/four.html"></iframe>
+      </div>
+    </mdw-card>
+  </section>
+</article>
