+++ src-docs/public/layouts/style.css
@@ -1,0 +1,44 @@
+html {
+  font-family: Roboto, 'Helvetica Neue', sans-serif;
+  font-size: 16px;
+  margin: 0;
+}
+
+body {
+  display: flex;
+  flex-direction: column;
+  font-family: Roboto, 'Helvetica Neue', sans-serif;
+  margin: 0;
+  height: 100%;
+  width: 100%;
+  background-color: var(--mdw-theme-background);
+  color: var(--mdw-theme-on-primary);
+}
+
+mdw-row {
+  display: flex;
+  flex-direction: row;
+}
+
+mdw-column {
+  display: flex;
+  flex-direction: column;
+}
+
+mdw-scroll-container {
+  display: flex;
+  flex-direction: row;
+  position: relative;
+  overflow-y: hidden;
+  width: 100%;
+  height: 100%;
+}
+
+mdw-scroll-content {
+  display: block;
+  box-sizing: border-box;
+  height: 100%;
+  padding: 0 40px;
+  width: 100%;
+  overflow: auto;
+}
