+++ src-docs/public/layouts/three.html
@@ -1,0 +1,147 @@
+<!doctype html>
+<html lang="en" style="height: 100%">
+
+<head>
+  <meta http-equiv="Cache-Control" content="no-store" />
+  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Fira+Mono" />
+  <link href="/pax-components/theme.css" rel="stylesheet">
+  <link href="/pax-components/entry.css" rel="stylesheet">
+  <script src="/pax-components/entry.js"></script>
+  <link rel="stylesheet" href="style.css">
+</head>
+
+<body class="mdw-row">
+  <mdw-column>
+    <!-- page header -->
+    <mdw-top-app-bar>
+      <section>
+        <mdw-icon onclick="document.querySelector('.mdw-navigation-drawer').toggle()">menu</mdw-icon>
+      </section>
+    
+      <section class="mdw-flex">
+        <span class="mdw-title">Example</span>
+      </section>
+    
+      <section>
+        <mdw-icon onclick="document.querySelector('#sideSheet').toggle()">menu</mdw-icon>
+      </section>
+    </mdw-top-app-bar>
+
+    <!--- page-body -->
+    <mdw-scroll-container>
+      <!-- navigation -->
+      <mdw-sheet-side class="mdw-navigation-drawer">
+        <mdw-header>
+          <section>
+            <div class="mdw-title">Nav title</div>
+            <div class="mdw-subtitle">Pages</div>
+          </section>
+        </mdw-header>
+      
+        <mdw-content>
+          <mdw-list>
+            <mdw-list-item class="mdw-current-link">
+              <mdw-icon>inbox</mdw-icon>
+              Inbox
+            </mdw-list-item>
+            <mdw-list-item>
+              <mdw-icon>star</mdw-icon>
+              Two
+            </mdw-list-item>
+            <mdw-list-item>
+              <mdw-icon>star</mdw-icon>
+              Three
+            </mdw-list-item>
+            <mdw-list-item>
+              <mdw-icon>star</mdw-icon>
+              Four
+            </mdw-list-item>
+            <mdw-list-item>
+              <mdw-icon>inbox</mdw-icon>
+              Inbox
+            </mdw-list-item>
+            <mdw-list-item>
+              <mdw-icon>star</mdw-icon>
+              Two
+            </mdw-list-item>
+            <mdw-list-item>
+              <mdw-icon>star</mdw-icon>
+              Three
+            </mdw-list-item>
+            <mdw-list-item>
+              <mdw-icon>star</mdw-icon>
+              Four
+            </mdw-list-item>
+          </mdw-list>
+        </mdw-content>
+      </mdw-sheet-side>
+
+      <!-- page content -->
+      <mdw-scroll-content>
+        <p>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+          sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
+          enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+          aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
+          voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
+          occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
+          anim id est laborum.
+        </p>
+        <p>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+          sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
+          enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+          aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
+          voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
+          occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
+          anim id est laborum.
+        </p>
+        <p>
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+          sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
+          enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+          aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
+          voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
+          occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
+          anim id est laborum.
+        </p>
+      </mdw-scroll-content>
+    </mdw-scroll-container>
+  </mdw-column>
+
+  <!-- side sheet -->
+  <mdw-sheet-side id="sideSheet">
+    <mdw-header>
+      <div class="mdw-title mdw-flex">Title</div>
+      <mdw-button class="mdw-icon" onclick="sideSheet.close()">
+        <mdw-icon>close</mdw-icon>
+      </mdw-button>
+    </mdw-header>
+    <mdw-content>
+      Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+      sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
+      enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+      aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
+      voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
+      occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
+      anim id est laborum.
+      Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+      sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
+      enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+      aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
+      voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
+      occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
+      anim id est laborum.
+      Lorem ipsum dolor sit amet, consectetur adipiscing elit,
+      sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
+      enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut
+      aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in
+      voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
+      occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
+      anim id est laborum.
+    </mdw-content>
+  </mdw-sheet-side>
+</body>
+
+</html>
