+++ src-docs/public/variables.html
@@ -1,0 +1,1 @@
+<div>${this.variable}</div>
