+++ src/components/panel/external.scss
@@ -357,9 +357,6 @@ mdw-panel.mdw-open:not(.mdw-panel-hoisted):not(.mdw-upgraded)[mdw-position="cent
 }
 
 mdw-panel .mdw-options-list {
-  font-family: Roboto,sans-serif;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
   font-size: 1rem;
   line-height: 1.75rem;
   font-weight: 400;
