+++ src/core/headings-text.scss
@@ -1,9 +1,3 @@
-h1, h2, h3, h4, h5, h6, .mdw-headline1, .mdw-headline2, .mdw-headline3, .mdw-headline4, .mdw-headline5, .mdw-headline6,
-p, .mdw-subtitle, .mdw-subtitle2, .mdw-body1, .mdw-body2, caption, .mdw-caption, button, .mdw-description {
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-}
-
 h1, h2, h3, h4, h5, h6, .mdw-headline1, .mdw-headline2, .mdw-headline3, .mdw-headline4, .mdw-headline5, .mdw-headline6 {
   color: var(--mdw-theme-text--heading);
 }
